Apply a statistic function to all column vectors

Apply a counting summary function like dplyr::n_distinct() or count_na() to every column of a dataframe and return the results along with a percentage of that value.

col_stats(data, fun, print = TRUE)

Arguments

data
A data frame to glimpse.

fun
A function to map to each column.

print
logical; Should all columns be printed as rows?

Value

A tibble with a row for every column and new columns with count and proportion.

Examples

col_stats(dplyr::storms, dplyr::n_distinct)

#> # A tibble: 13 x 4
#>    col         class     n        p
#>    <chr>       <chr> <int>    <dbl>
#>  1 name        <chr>   198 0.0198  
#>  2 year        <dbl>    41 0.00410 
#>  3 month       <dbl>    10 0.000999
#>  4 day         <int>    31 0.00310 
#>  5 hour        <dbl>    24 0.00240 
#>  6 lat         <dbl>   403 0.0403  
#>  7 long        <dbl>   856 0.0855  
#>  8 status      <chr>     3 0.000300
#>  9 category    <ord>     7 0.000699
#> 10 wind        <int>    31 0.00310 
#> 11 pressure    <int>   124 0.0124  
#> 12 ts_diameter <dbl>   109 0.0109  
#> 13 hu_diameter <dbl>    35 0.00350 

col_stats(dplyr::storms, campfin::count_na)

#> # A tibble: 13 x 4
#>    col         class     n     p
#>    <chr>       <chr> <int> <dbl>
#>  1 name        <chr>     0 0    
#>  2 year        <dbl>     0 0    
#>  3 month       <dbl>     0 0    
#>  4 day         <int>     0 0    
#>  5 hour        <dbl>     0 0    
#>  6 lat         <dbl>     0 0    
#>  7 long        <dbl>     0 0    
#>  8 status      <chr>     0 0    
#>  9 category    <ord>     0 0    
#> 10 wind        <int>     0 0    
#> 11 pressure    <int>     0 0    
#> 12 ts_diameter <dbl>  6528 0.652
#> 13 hu_diameter <dbl>  6528 0.652