Count values of a vector

A version of dplyr::count() which uses tibble::enframe() to count the number of values in a single vector.

count_vec(x, sort = TRUE)

Arguments

x
A vector to check.

sort
logical; if TRUE will sort output in descending order of n

Value

A tibble, with counts of each value in n.

See also

Examples

count_vec(x = rivers)

#> # A tibble: 114 x 2
#>    value     n
#>    <dbl> <int>
#>  1   350     4
#>  2   360     4
#>  3   250     3
#>  4   280     3
#>  5   300     3
#>  6   600     3
#>  7   210     2
#>  8   230     2
#>  9   260     2
#> 10   310     2
#> # … with 104 more rows

count_vec(x = sample(x = state.name, size = 1000, replace = TRUE))

#> # A tibble: 50 x 2
#>    value              n
#>    <chr>          <int>
#>  1 North Carolina    30
#>  2 Connecticut       28
#>  3 Hawaii            27
#>  4 Montana           27
#>  5 Idaho             26
#>  6 Minnesota         26
#>  7 Tennessee         26
#>  8 California        24
#>  9 Florida           24
#> 10 West Virginia     24
#> # … with 40 more rows