Create a progress table

Create a tibble with rows for each stage of normalization and columns for the various statistics most useful in assessing the progress of each stage.

progress_table(..., compare)

Arguments

...

Any number of vectors to check.

compare

A vector to compare each of ... against. Useful with valid_zip, valid_state (valid_name), or valid_city.

Value

A table with a row for each vector in ....

Examples

progress_table(state.name, toupper(state.name), compare = valid_name)
#> # A tibble: 2 × 6
#>   stage               prop_in n_distinct prop_na n_out n_diff
#>   <chr>                 <dbl>      <dbl>   <dbl> <dbl>  <dbl>
#> 1 state.name                0         50       0    50     50
#> 2 toupper(state.name)       1         50       0     0      0