Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type, author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contributions\""                                                
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r date()`\""                                                            
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 3"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "if (!interactive()) {"                                                           
#>  [31] "  options(width = 120)"                                                          
#>  [32] "  set.seed(5)"                                                                   
#>  [33] "}"                                                                               
#>  [34] "```"                                                                             
#>  [35] ""                                                                                
#>  [36] "```{r create-docs-dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [37] "doc_dir <- fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"           
#>  [38] "```"                                                                             
#>  [39] ""                                                                                
#>  [40] "## Project"                                                                      
#>  [41] ""                                                                                
#>  [42] "The Accountability Project is an effort to cut across data silos and give"       
#>  [43] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [44] "way to search across huge volumes of public data about people and organizations."
#>  [45] ""                                                                                
#>  [46] "Our goal is to standardize public data on a few key fields by thinking of each"  
#>  [47] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [48] "variables:"                                                                      
#>  [49] ""                                                                                
#>  [50] "1. All **parties** to a transaction."                                            
#>  [51] "2. The **date** of the transaction."                                             
#>  [52] "3. The **amount** of money involved."                                            
#>  [53] ""                                                                                
#>  [54] "## Objectives"                                                                   
#>  [55] ""                                                                                
#>  [56] "This document describes the process used to complete the following objectives:"  
#>  [57] ""                                                                                
#>  [58] "1. How many records are in the database?"                                        
#>  [59] "1. Check for entirely duplicated records."                                       
#>  [60] "1. Check ranges of continuous variables."                                        
#>  [61] "1. Is there anything blank or missing?"                                          
#>  [62] "1. Check for consistency issues."                                                
#>  [63] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [64] "1. Create a `year` field from the transaction date."                             
#>  [65] "1. Make sure there is data on both parties to a transaction."                    
#>  [66] ""                                                                                
#>  [67] "## Packages"                                                                     
#>  [68] ""                                                                                
#>  [69] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [70] "and communicate these results. The `pacman` package will facilitate their"       
#>  [71] "installation and attachment."                                                    
#>  [72] ""                                                                                
#>  [73] "```{r load-packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [74] "if (!require(\"pacman\")) {"                                                     
#>  [75] "  install.packages(\"pacman\")"                                                  
#>  [76] "}"                                                                               
#>  [77] "pacman::p_load("                                                                 
#>  [78] "  tidyverse, # data manipulation"                                                
#>  [79] "  lubridate, # datetime strings"                                                 
#>  [80] "  gluedown, # printing markdown"                                                 
#>  [81] "  janitor, # clean data frames"                                                  
#>  [82] "  campfin, # custom irw tools"                                                   
#>  [83] "  aws.s3, # aws cloud storage"                                                   
#>  [84] "  refinr, # cluster & merge"                                                     
#>  [85] "  scales, # format strings"                                                      
#>  [86] "  knitr, # knit documents"                                                       
#>  [87] "  vroom, # fast reading"                                                         
#>  [88] "  rvest, # scrape html"                                                          
#>  [89] "  glue, # code strings"                                                          
#>  [90] "  here, # project paths"                                                         
#>  [91] "  httr, # http requests"                                                         
#>  [92] "  fs # local storage "                                                           
#>  [93] ")"                                                                               
#>  [94] "```"                                                                             
#>  [95] ""                                                                                
#>  [96] "```{r package-options, echo=FALSE}"                                              
#>  [97] "options(options(knitr.kable.NA = \"\"))"                                         
#>  [98] "```"                                                                             
#>  [99] ""                                                                                
#> [100] "This document should be run as part of the `R_campfin` project, which lives as a"
#> [101] "sub-directory of the more general, language-agnostic"                            
#> [102] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#> [103] ""                                                                                
#> [104] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#> [105] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#> [106] "paths relative to _your_ machine."                                               
#> [107] ""                                                                                
#> [108] "```{r where-here}"                                                               
#> [109] "# where does this document knit?"                                                
#> [110] "here::i_am(\"vt/contribs/docs/vt_contribs_diary.Rmd\")"                          
#> [111] "```"                                                                             
#> [112] ""                                                                                
#> [113] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [114] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [115] ""                                                                                
#> [116] "## Data"                                                                         
#> [117] ""                                                                                
#> [118] ""                                                                                
#> [119] ""                                                                                
#> [120] "## Download"                                                                     
#> [121] ""                                                                                
#> [122] "```{r raw-dir}"                                                                  
#> [123] "raw_url <- \"https://example.com/source_file.csv\""                              
#> [124] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [125] "raw_csv <- path(raw_dir, basename(raw_url))"                                     
#> [126] "```"                                                                             
#> [127] ""                                                                                
#> [128] "```{r raw-download}"                                                             
#> [129] "if (!file_exists(raw_path)) {"                                                   
#> [130] "  download.file(raw_url, raw_path)"                                              
#> [131] "}"                                                                               
#> [132] "```"                                                                             
#> [133] ""                                                                                
#> [134] "## Read"                                                                         
#> [135] ""                                                                                
#> [136] "```{r raw-read}"                                                                 
#> [137] "vtc <- read_delim("                                                              
#> [138] "  file = raw_path,"                                                              
#> [139] "  delim = \",\","                                                                
#> [140] "  escape_backslash = FALSE,"                                                     
#> [141] "  escape_double = FALSE,"                                                        
#> [142] "  col_types = cols("                                                             
#> [143] "    .default = col_character(),"                                                 
#> [144] "    date = col_date_usa(),"                                                      
#> [145] "    amount = col_double()"                                                       
#> [146] "  )"                                                                             
#> [147] ")"                                                                               
#> [148] "```"                                                                             
#> [149] ""                                                                                
#> [150] "```{r clean-names}"                                                              
#> [151] "vtc <- clean_names(vtc, case = \"snake\")"                                       
#> [152] "```"                                                                             
#> [153] ""                                                                                
#> [154] "## Explore"                                                                      
#> [155] ""                                                                                
#> [156] "There are `r comma(nrow(vtc))` rows of `r ncol(vtc)` columns. Each record"       
#> [157] "represents a single Contributions..."                                            
#> [158] ""                                                                                
#> [159] "```{r glimpse}"                                                                  
#> [160] "glimpse(vtc)"                                                                    
#> [161] "tail(vtc)"                                                                       
#> [162] "```"                                                                             
#> [163] ""                                                                                
#> [164] "### Missing"                                                                     
#> [165] ""                                                                                
#> [166] "Columns vary in their degree of missing values."                                 
#> [167] ""                                                                                
#> [168] "```{r na-count}"                                                                 
#> [169] "col_stats(vtc, count_na)"                                                        
#> [170] "```"                                                                             
#> [171] ""                                                                                
#> [172] "We can flag any record missing a key variable needed to identify a transaction." 
#> [173] ""                                                                                
#> [174] "```{r na-flag}"                                                                  
#> [175] "key_vars <- c(\"date\", \"last_name\", \"amount\", \"committee_name\")"          
#> [176] "vtc <- flag_na(vtc, all_of(key_vars))"                                           
#> [177] "sum(vtc$na_flag)"                                                                
#> [178] "```"                                                                             
#> [179] ""                                                                                
#> [180] "```{r na-view}"                                                                  
#> [181] "vtc %>% "                                                                        
#> [182] "  filter(na_flag) %>% "                                                          
#> [183] "  select(all_of(key_vars))"                                                      
#> [184] "```"                                                                             
#> [185] ""                                                                                
#> [186] "### Duplicates"                                                                  
#> [187] ""                                                                                
#> [188] "We can also flag any record completely duplicated across every column."          
#> [189] ""                                                                                
#> [190] "```{r dupe-flag}"                                                                
#> [191] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [192] "sum(vtc$dupe_flag)"                                                              
#> [193] "```"                                                                             
#> [194] ""                                                                                
#> [195] "```{r dupe-view}"                                                                
#> [196] "vtc %>% "                                                                        
#> [197] "  filter(dupe_flag) %>% "                                                        
#> [198] "  select(all_of(key_vars))"                                                      
#> [199] "```"                                                                             
#> [200] ""                                                                                
#> [201] "### Categorical"                                                                 
#> [202] ""                                                                                
#> [203] "```{r distinct-count}"                                                           
#> [204] "col_stats(vtc, n_distinct)"                                                      
#> [205] "```"                                                                             
#> [206] ""                                                                                
#> [207] "```{r distinct-plots, echo=FALSE}"                                               
#> [208] "explore_plot(vtc, type)"                                                         
#> [209] "```"                                                                             
#> [210] ""                                                                                
#> [211] "### Amounts"                                                                     
#> [212] ""                                                                                
#> [213] "```{r amount-round}"                                                             
#> [214] "vtc$amount <- round(vtc$amount, digits = 2)"                                     
#> [215] "```"                                                                             
#> [216] ""                                                                                
#> [217] "```{r amount-summary}"                                                           
#> [218] "summary(vtc$amount)"                                                             
#> [219] "mean(vtc$amount <= 0)"                                                           
#> [220] "```"                                                                             
#> [221] ""                                                                                
#> [222] "These are the records with the minimum and maximum amounts."                     
#> [223] ""                                                                                
#> [224] "```{r amount-minmax}"                                                            
#> [225] "glimpse(vtc[c(which.max(vtc$amount), which.min(vtc$amount)), ])"                 
#> [226] "```"                                                                             
#> [227] ""                                                                                
#> [228] "```{r hist-amount, echo=FALSE}"                                                  
#> [229] "vtc %>%"                                                                         
#> [230] "  ggplot(aes(amount)) +"                                                         
#> [231] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [232] "  scale_y_continuous(labels = comma) +"                                          
#> [233] "  scale_x_continuous("                                                           
#> [234] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [235] "    labels = dollar,"                                                            
#> [236] "    trans = \"log10\""                                                           
#> [237] "  ) +"                                                                           
#> [238] "  labs("                                                                         
#> [239] "    title = \"Vermont Contributions Amount Distribution\","                      
#> [240] "    caption = \"Source: {source}\","                                             
#> [241] "    x = \"Amount\","                                                             
#> [242] "    y = \"Count\""                                                               
#> [243] "  )"                                                                             
#> [244] "```"                                                                             
#> [245] ""                                                                                
#> [246] "### Dates"                                                                       
#> [247] ""                                                                                
#> [248] "We can add the calendar year from `date` with `lubridate::year()`"               
#> [249] ""                                                                                
#> [250] "```{r date-year}"                                                                
#> [251] "vtc <- mutate(vtc, year = year(date))"                                           
#> [252] "```"                                                                             
#> [253] ""                                                                                
#> [254] "```{r date-range}"                                                               
#> [255] "min(vtc$date)"                                                                   
#> [256] "sum(vtc$year < 2000)"                                                            
#> [257] "max(vtc$date)"                                                                   
#> [258] "sum(vtc$date > today())"                                                         
#> [259] "```"                                                                             
#> [260] ""                                                                                
#> [261] "```{r bar-year, echo=FALSE}"                                                     
#> [262] "vtc %>% "                                                                        
#> [263] "  count(year) %>% "                                                              
#> [264] "  mutate(even = is_even(year)) %>% "                                             
#> [265] "  ggplot(aes(x = year, y = n)) +"                                                
#> [266] "  geom_col(aes(fill = even)) + "                                                 
#> [267] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [268] "  scale_y_continuous(labels = comma) +"                                          
#> [269] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [270] "  theme(legend.position = \"bottom\") +"                                         
#> [271] "  labs("                                                                         
#> [272] "    title = \"Vermont Contributions by Year\","                                  
#> [273] "    caption = \"Source: {source}\","                                             
#> [274] "    fill = \"Election Year\","                                                   
#> [275] "    x = \"Year Made\","                                                          
#> [276] "    y = \"Count\""                                                               
#> [277] "  )"                                                                             
#> [278] "```"                                                                             
#> [279] ""                                                                                
#> [280] "## Wrangle"                                                                      
#> [281] ""                                                                                
#> [282] "To improve the searchability of the database, we will perform some consistent,"  
#> [283] "confident string normalization. For geographic variables like city names and"    
#> [284] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [285] "facilitate this process."                                                        
#> [286] ""                                                                                
#> [287] "### Address"                                                                     
#> [288] ""                                                                                
#> [289] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [290] "will force consistence case, remove punctuation, and abbreviate official "       
#> [291] "USPS suffixes."                                                                  
#> [292] ""                                                                                
#> [293] "```{r address-norm}"                                                             
#> [294] "addr_norm <- vtc %>% "                                                           
#> [295] "  distinct(address1, address2) %>% "                                             
#> [296] "  unite("                                                                        
#> [297] "    col = address_full,"                                                         
#> [298] "    starts_with(\"address\"),"                                                   
#> [299] "    sep = \" \","                                                                
#> [300] "    remove = FALSE,"                                                             
#> [301] "    na.rm = TRUE"                                                                
#> [302] "  ) %>% "                                                                        
#> [303] "  mutate("                                                                       
#> [304] "    address_norm = normal_address("                                              
#> [305] "      address = address_full,"                                                   
#> [306] "      abbs = usps_street,"                                                       
#> [307] "      na_rep = TRUE"                                                             
#> [308] "    )"                                                                           
#> [309] "  ) %>% "                                                                        
#> [310] "  select(-address_full)"                                                         
#> [311] "```"                                                                             
#> [312] ""                                                                                
#> [313] "```{r address-view}"                                                             
#> [314] "addr_norm"                                                                       
#> [315] "```"                                                                             
#> [316] ""                                                                                
#> [317] "```{r address-join}"                                                             
#> [318] "vtc <- left_join(vtc, addr_norm, by = c(\"address1\", \"address2\"))"            
#> [319] "```"                                                                             
#> [320] ""                                                                                
#> [321] "### ZIP"                                                                         
#> [322] ""                                                                                
#> [323] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [324] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [325] "zeroes dropped by other programs like Microsoft Excel."                          
#> [326] ""                                                                                
#> [327] "```{r zip-norm}"                                                                 
#> [328] "vtc <- vtc %>% "                                                                 
#> [329] "  mutate("                                                                       
#> [330] "    zip_norm = normal_zip("                                                      
#> [331] "      zip = zip,"                                                                
#> [332] "      na_rep = TRUE"                                                             
#> [333] "    )"                                                                           
#> [334] "  )"                                                                             
#> [335] "```"                                                                             
#> [336] ""                                                                                
#> [337] "```{r zip-progress}"                                                             
#> [338] "progress_table("                                                                 
#> [339] "  vtc$zip,"                                                                      
#> [340] "  vtc$zip_norm,"                                                                 
#> [341] "  compare = valid_zip"                                                           
#> [342] ")"                                                                               
#> [343] "```"                                                                             
#> [344] ""                                                                                
#> [345] "### State"                                                                       
#> [346] ""                                                                                
#> [347] "Valid two digit state abbreviations can be made using the "                      
#> [348] "`campfin::normal_state()` function."                                             
#> [349] ""                                                                                
#> [350] "```{r state-norm}"                                                               
#> [351] "vtc <- vtc %>% "                                                                 
#> [352] "  mutate("                                                                       
#> [353] "    state_norm = normal_state("                                                  
#> [354] "      state = state,"                                                            
#> [355] "      abbreviate = TRUE,"                                                        
#> [356] "      na_rep = TRUE,"                                                            
#> [357] "      valid = valid_state"                                                       
#> [358] "    )"                                                                           
#> [359] "  )"                                                                             
#> [360] "```"                                                                             
#> [361] ""                                                                                
#> [362] "```{r state-view}"                                                               
#> [363] "vtc %>% "                                                                        
#> [364] "  filter(state != state_norm) %>% "                                              
#> [365] "  count(state, state_norm, sort = TRUE)"                                         
#> [366] "```"                                                                             
#> [367] ""                                                                                
#> [368] "```{r state-progress}"                                                           
#> [369] "progress_table("                                                                 
#> [370] "  vtc$state,"                                                                    
#> [371] "  vtc$state_norm,"                                                               
#> [372] "  compare = valid_state"                                                         
#> [373] ")"                                                                               
#> [374] "```"                                                                             
#> [375] ""                                                                                
#> [376] "### City"                                                                        
#> [377] ""                                                                                
#> [378] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [379] "the wide variety of valid cities and formats."                                   
#> [380] ""                                                                                
#> [381] "#### Normal"                                                                     
#> [382] ""                                                                                
#> [383] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [384] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [385] "`invalid_city` values."                                                          
#> [386] ""                                                                                
#> [387] "```{r city-norm}"                                                                
#> [388] "norm_city <- vtc %>% "                                                           
#> [389] "  distinct(city, state_norm, zip_norm) %>% "                                     
#> [390] "  mutate("                                                                       
#> [391] "    city_norm = normal_city("                                                    
#> [392] "      city = city, "                                                             
#> [393] "      abbs = usps_city,"                                                         
#> [394] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [395] "      na = invalid_city,"                                                        
#> [396] "      na_rep = TRUE"                                                             
#> [397] "    )"                                                                           
#> [398] "  )"                                                                             
#> [399] "```"                                                                             
#> [400] ""                                                                                
#> [401] "#### Swap"                                                                       
#> [402] ""                                                                                
#> [403] "We can further improve normalization by comparing our normalized value"          
#> [404] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [405] "If the normalized value is either an abbreviation for or very similar to the"    
#> [406] "expected value, we can confidently swap those two."                              
#> [407] ""                                                                                
#> [408] "```{r city-swap}"                                                                
#> [409] "norm_city <- norm_city %>% "                                                     
#> [410] "  rename(city_raw = city) %>% "                                                  
#> [411] "  left_join("                                                                    
#> [412] "    y = zipcodes,"                                                               
#> [413] "    by = c("                                                                     
#> [414] "      \"state_norm\" = \"state\","                                               
#> [415] "      \"zip_norm\" = \"zip\""                                                    
#> [416] "    )"                                                                           
#> [417] "  ) %>% "                                                                        
#> [418] "  rename(city_match = city) %>% "                                                
#> [419] "  mutate("                                                                       
#> [420] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [421] "    match_dist = str_dist(city_norm, city_match),"                               
#> [422] "    city_swap = if_else("                                                        
#> [423] "      condition = !is.na(match_dist) & (match_abb | match_dist == 1),"           
#> [424] "      true = city_match,"                                                        
#> [425] "      false = city_norm"                                                         
#> [426] "    )"                                                                           
#> [427] "  ) %>% "                                                                        
#> [428] "  select("                                                                       
#> [429] "    -city_match,"                                                                
#> [430] "    -match_dist,"                                                                
#> [431] "    -match_abb"                                                                  
#> [432] "  )"                                                                             
#> [433] "```"                                                                             
#> [434] ""                                                                                
#> [435] "```{r city-rejoin}"                                                              
#> [436] "vtc <- left_join("                                                               
#> [437] "  x = vtc,"                                                                      
#> [438] "  y = norm_city,"                                                                
#> [439] "  by = c("                                                                       
#> [440] "    \"city\" = \"city_raw\", "                                                   
#> [441] "    \"state_norm\", "                                                            
#> [442] "    \"zip_norm\""                                                                
#> [443] "  )"                                                                             
#> [444] ")"                                                                               
#> [445] "```"                                                                             
#> [446] ""                                                                                
#> [447] "#### Refine"                                                                     
#> [448] ""                                                                                
#> [449] "The [OpenRefine][or] algorithms can be used to group similar strings and replace"
#> [450] "the less common versions with their most common counterpart. This can greatly"   
#> [451] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [452] "strings that have a valid city/state/zip combination."                           
#> [453] ""                                                                                
#> [454] "[or]: https://openrefine.org/"                                                   
#> [455] ""                                                                                
#> [456] "```{r city-refine}"                                                              
#> [457] "good_refine <- vtc %>% "                                                         
#> [458] "  mutate("                                                                       
#> [459] "    city_refine = city_swap %>% "                                                
#> [460] "      key_collision_merge() %>% "                                                
#> [461] "      n_gram_merge(numgram = 1)"                                                 
#> [462] "  ) %>% "                                                                        
#> [463] "  filter(city_refine != city_swap) %>% "                                         
#> [464] "  inner_join("                                                                   
#> [465] "    y = zipcodes,"                                                               
#> [466] "    by = c("                                                                     
#> [467] "      \"city_refine\" = \"city\","                                               
#> [468] "      \"state_norm\" = \"state\","                                               
#> [469] "      \"zip_norm\" = \"zip\""                                                    
#> [470] "    )"                                                                           
#> [471] "  )"                                                                             
#> [472] "```"                                                                             
#> [473] ""                                                                                
#> [474] "```{r city-count, echo=FALSE}"                                                   
#> [475] "good_refine %>%"                                                                 
#> [476] "  count("                                                                        
#> [477] "    state_norm, "                                                                
#> [478] "    zip_norm, "                                                                  
#> [479] "    city_swap, "                                                                 
#> [480] "    city_refine,"                                                                
#> [481] "    sort = TRUE"                                                                 
#> [482] "  )"                                                                             
#> [483] "```"                                                                             
#> [484] ""                                                                                
#> [485] "Then we can join the refined values back to the database."                       
#> [486] ""                                                                                
#> [487] "```{r city-join}"                                                                
#> [488] "vtc <- vtc %>% "                                                                 
#> [489] "  left_join(good_refine, by = names(.)) %>% "                                    
#> [490] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [491] "```"                                                                             
#> [492] ""                                                                                
#> [493] "#### Progress"                                                                   
#> [494] ""                                                                                
#> [495] "Our goal for normalization was to increase the proportion of city values known"  
#> [496] "to be valid and reduce the total distinct values by correcting misspellings."    
#> [497] ""                                                                                
#> [498] "```{r city-progress, echo=FALSE}"                                                
#> [499] "many_city <- c(valid_city, extra_city)"                                          
#> [500] "progress <- progress_table("                                                     
#> [501] "  str_to_upper(vtc$city),"                                                       
#> [502] "  vtc$city_norm,"                                                                
#> [503] "  vtc$city_swap,"                                                                
#> [504] "  vtc$city_refine,"                                                              
#> [505] "  compare = many_city"                                                           
#> [506] ") %>% mutate(stage = as_factor(stage))"                                          
#> [507] "progress %>% "                                                                   
#> [508] "    mutate(across(stage, md_code)) %>% "                                         
#> [509] "    kable(digits = 3)"                                                           
#> [510] "```"                                                                             
#> [511] ""                                                                                
#> [512] "You can see how the percentage of valid values increased with each stage."       
#> [513] ""                                                                                
#> [514] "```{r bar-progress, echo=FALSE}"                                                 
#> [515] "raw_in <- percent(prop_in(vtc$city, valid_city))"                                
#> [516] "progress %>% "                                                                   
#> [517] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [518] "  geom_hline(yintercept = 0.99) +"                                               
#> [519] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [520] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [521] "  scale_y_continuous(labels = percent) +"                                        
#> [522] "  labs("                                                                         
#> [523] "    title = \"Vermont City Normalization Progress\","                            
#> [524] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [525] "    x = \"Stage\","                                                              
#> [526] "    y = \"Percent Valid\""                                                       
#> [527] "  )"                                                                             
#> [528] "```"                                                                             
#> [529] ""                                                                                
#> [530] "More importantly, the number of distinct values decreased each stage. We were"   
#> [531] "able to confidently change many distinct invalid values to their valid"          
#> [532] "equivalent."                                                                     
#> [533] ""                                                                                
#> [534] "```{r bar-distinct, echo=FALSE}"                                                 
#> [535] "progress %>% "                                                                   
#> [536] "  select("                                                                       
#> [537] "    stage, "                                                                     
#> [538] "    all = n_distinct,"                                                           
#> [539] "    bad = n_diff"                                                                
#> [540] "  ) %>% "                                                                        
#> [541] "  mutate(good = all - bad) %>% "                                                 
#> [542] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [543] "  mutate(name = name == \"good\") %>% "                                          
#> [544] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [545] "  geom_col(aes(fill = name)) +"                                                  
#> [546] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [547] "  scale_y_continuous(labels = comma) +"                                          
#> [548] "  theme(legend.position = \"bottom\") +"                                         
#> [549] "  labs("                                                                         
#> [550] "    title = \"Vermont City Normalization Progress\","                            
#> [551] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [552] "    x = \"Stage\","                                                              
#> [553] "    y = \"Distinct Values\","                                                    
#> [554] "    fill = \"Valid\""                                                            
#> [555] "  )"                                                                             
#> [556] "```"                                                                             
#> [557] ""                                                                                
#> [558] "Before exporting, we can remove the intermediary normalization columns and"      
#> [559] "rename all added variables with the `_clean` suffix."                            
#> [560] ""                                                                                
#> [561] "```{r clean-select}"                                                             
#> [562] "vtc <- vtc %>% "                                                                 
#> [563] "  select("                                                                       
#> [564] "    -city_norm,"                                                                 
#> [565] "    -city_swap,"                                                                 
#> [566] "    city_clean = city_refine"                                                    
#> [567] "  ) %>% "                                                                        
#> [568] "  rename_all(~str_replace(., \"_norm\", \"_clean\")) %>% "                       
#> [569] "  rename_all(~str_remove(., \"_raw\")) %>% "                                     
#> [570] "  relocate(address_clean, city_clean, state_clean, .before = zip_clean)"         
#> [571] "```"                                                                             
#> [572] ""                                                                                
#> [573] "## Conclude"                                                                     
#> [574] ""                                                                                
#> [575] "```{r clean-glimpse}"                                                            
#> [576] "glimpse(sample_n(vtc, 50))"                                                      
#> [577] "```"                                                                             
#> [578] ""                                                                                
#> [579] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [580] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [581] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [582] "1. There are `r comma(sum(vtc$na_flag))` records missing key variables."         
#> [583] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [584] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [585] ""                                                                                
#> [586] "## Export"                                                                       
#> [587] ""                                                                                
#> [588] "Now the file can be saved on disk for upload to the Accountability server."      
#> [589] ""                                                                                
#> [590] "```{r clean-dir}"                                                                
#> [591] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [592] "clean_csv <- path(clean_dir, \"vt_contribs_clean.csv\")"                         
#> [593] "write_csv(vtc, clean_csv, na = \"\")"                                            
#> [594] "(clean_size <- file_size(clean_path))"                                           
#> [595] "```"                                                                             
#> [596] ""                                                                                
#> [597] "## Upload"                                                                       
#> [598] ""                                                                                
#> [599] "We can use the `aws.s3::put_object()` to upload the text file to the IRW server."
#> [600] ""                                                                                
#> [601] "```{r aws-upload, eval=FALSE}"                                                   
#> [602] "aws_csv <- path(\"csv\", basename(clean_csv))"                                   
#> [603] "if (!object_exists(aws_csv, \"publicaccountability\")) {"                        
#> [604] "  put_object("                                                                   
#> [605] "    file = clean_csv,"                                                           
#> [606] "    object = aws_csv, "                                                          
#> [607] "    bucket = \"publicaccountability\","                                          
#> [608] "    acl = \"public-read\","                                                      
#> [609] "    show_progress = TRUE,"                                                       
#> [610] "    multipart = TRUE"                                                            
#> [611] "  )"                                                                             
#> [612] "}"                                                                               
#> [613] "aws_head <- head_object(aws_csv, \"publicaccountability\")"                      
#> [614] "(aws_size <- as_fs_bytes(attr(aws_head, \"content-length\")))"                   
#> [615] "unname(aws_size == clean_size)"                                                  
#> [616] "```"                                                                             
#> [617] ""                                                                                
#> [618] "## Dictionary"                                                                   
#> [619] ""                                                                                
#> [620] "The following table describes the variables in our final exported file:"         
#> [621] ""                                                                                
#> [622] "```{r dict-make, echo=FALSE}"                                                    
#> [623] "dict_raw <- tibble("                                                             
#> [624] "  var = md_code(names(vtc)),"                                                    
#> [625] "  type = md_code(map_chr(vtc, typeof)),"                                         
#> [626] "  def = c("                                                                      
#> [627] "    \"\""                                                                        
#> [628] "  )"                                                                             
#> [629] ")"                                                                               
#> [630] "```"                                                                             
#> [631] ""                                                                                
#> [632] "```{r dict-md, echo=FALSE}"                                                      
#> [633] "(dict_md <- kable("                                                              
#> [634] "  x = dict_raw,"                                                                 
#> [635] "  format = \"markdown\","                                                        
#> [636] "  col.names = c(\"Column\", \"Type\", \"Definition\")"                           
#> [637] "))"                                                                              
#> [638] "```"                                                                             

use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpixQ4gQ/vt_contribs_diary.Rmd was created