Count values of a vector

A version of dplyr::count() which uses tibble::enframe() to count the number of values in a single vector.

count_vec(x, sort = TRUE)

Arguments

x
A vector to check.

sort
logical; if TRUE will sort output in descending order of n

Value

A tibble, with counts of each value in n.

See also

Examples

count_vec(x = rivers)

#> # A tibble: 114 x 2
#>    value     n
#>    <dbl> <int>
#>  1   350     4
#>  2   360     4
#>  3   250     3
#>  4   280     3
#>  5   300     3
#>  6   600     3
#>  7   210     2
#>  8   230     2
#>  9   260     2
#> 10   310     2
#> # … with 104 more rows

count_vec(x = sample(x = state.name, size = 1000, replace = TRUE))

#> # A tibble: 50 x 2
#>    value             n
#>    <chr>         <int>
#>  1 Oregon           30
#>  2 Kansas           28
#>  3 Tennessee        28
#>  4 New Hampshire    27
#>  5 Hawaii           26
#>  6 Iowa             25
#>  7 Minnesota        25
#>  8 Pennsylvania     25
#>  9 Utah             25
#> 10 Kentucky         24
#> # … with 40 more rows