Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type = c("contribs", "expends", "lobby"), author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contributions\""                                                
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r Sys.time()`\""                                                        
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 2"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "options(width = 99)"                                                             
#>  [31] "set.seed(5)"                                                                     
#>  [32] "```"                                                                             
#>  [33] ""                                                                                
#>  [34] "```{r create_docs_dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [35] "fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"                      
#>  [36] "```"                                                                             
#>  [37] ""                                                                                
#>  [38] "## Project"                                                                      
#>  [39] ""                                                                                
#>  [40] "The Accountability Project is an effort to cut across data silos and give"       
#>  [41] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [42] "way to search across huge volumes of public data about people and organizations."
#>  [43] ""                                                                                
#>  [44] "Our goal is to standardizing public data on a few key fields by thinking of each"
#>  [45] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [46] "variables:"                                                                      
#>  [47] ""                                                                                
#>  [48] "1. All **parties** to a transaction."                                            
#>  [49] "2. The **date** of the transaction."                                             
#>  [50] "3. The **amount** of money involved."                                            
#>  [51] ""                                                                                
#>  [52] "## Objectives"                                                                   
#>  [53] ""                                                                                
#>  [54] "This document describes the process used to complete the following objectives:"  
#>  [55] ""                                                                                
#>  [56] "1. How many records are in the database?"                                        
#>  [57] "1. Check for entirely duplicated records."                                       
#>  [58] "1. Check ranges of continuous variables."                                        
#>  [59] "1. Is there anything blank or missing?"                                          
#>  [60] "1. Check for consistency issues."                                                
#>  [61] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [62] "1. Create a `year` field from the transaction date."                             
#>  [63] "1. Make sure there is data on both parties to a transaction."                    
#>  [64] ""                                                                                
#>  [65] "## Packages"                                                                     
#>  [66] ""                                                                                
#>  [67] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [68] "and communicate these results. The `pacman` package will facilitate their"       
#>  [69] "installation and attachment."                                                    
#>  [70] ""                                                                                
#>  [71] "The IRW's `campfin` package will also have to be installed from GitHub. This"    
#>  [72] "package contains functions custom made to help facilitate the processing of"     
#>  [73] "campaign finance data."                                                          
#>  [74] ""                                                                                
#>  [75] "```{r load_packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [76] "if (!require(\"pacman\")) install.packages(\"pacman\")"                          
#>  [77] "pacman::p_load_gh(\"irworkshop/campfin\")"                                       
#>  [78] "pacman::p_load("                                                                 
#>  [79] "  tidyverse, # data manipulation"                                                
#>  [80] "  lubridate, # datetime strings"                                                 
#>  [81] "  magrittr, # pipe operators"                                                    
#>  [82] "  janitor, # dataframe clean"                                                    
#>  [83] "  refinr, # cluster and merge"                                                   
#>  [84] "  scales, # format strings"                                                      
#>  [85] "  knitr, # knit documents"                                                       
#>  [86] "  vroom, # read files fast"                                                      
#>  [87] "  glue, # combine strings"                                                       
#>  [88] "  here, # relative storage"                                                      
#>  [89] "  fs # search storage "                                                          
#>  [90] ")"                                                                               
#>  [91] "```"                                                                             
#>  [92] ""                                                                                
#>  [93] "This document should be run as part of the `R_campfin` project, which lives as a"
#>  [94] "sub-directory of the more general, language-agnostic"                            
#>  [95] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#>  [96] ""                                                                                
#>  [97] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#>  [98] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#>  [99] "paths relative to _your_ machine."                                               
#> [100] ""                                                                                
#> [101] "```{r where_here}"                                                               
#> [102] "# where does this document knit?"                                                
#> [103] "here::here()"                                                                    
#> [104] "```"                                                                             
#> [105] ""                                                                                
#> [106] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [107] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [108] ""                                                                                
#> [109] "## Data"                                                                         
#> [110] ""                                                                                
#> [111] "## Import"                                                                       
#> [112] ""                                                                                
#> [113] "### Download"                                                                    
#> [114] ""                                                                                
#> [115] "```{r raw_dir}"                                                                  
#> [116] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [117] "```"                                                                             
#> [118] ""                                                                                
#> [119] "```{r raw_download}"                                                             
#> [120] "download.file()"                                                                 
#> [121] "```"                                                                             
#> [122] ""                                                                                
#> [123] "### Read"                                                                        
#> [124] ""                                                                                
#> [125] "```{r raw_read}"                                                                 
#> [126] "vtc <- read_delim("                                                              
#> [127] "  file = dir_ls(raw_dir)"                                                        
#> [128] ")"                                                                               
#> [129] "```"                                                                             
#> [130] ""                                                                                
#> [131] "## Explore"                                                                      
#> [132] ""                                                                                
#> [133] "```{r glimpse}"                                                                  
#> [134] "head(vtc)"                                                                       
#> [135] "tail(vtc)"                                                                       
#> [136] "glimpse(sample_n(vtc, 20))"                                                      
#> [137] "```"                                                                             
#> [138] ""                                                                                
#> [139] "### Missing"                                                                     
#> [140] ""                                                                                
#> [141] "```{r na_count}"                                                                 
#> [142] "col_stats(vtc, count_na)"                                                        
#> [143] "```"                                                                             
#> [144] ""                                                                                
#> [145] "```{r na_flag}"                                                                  
#> [146] "vtc <- vtc %>% flag_na(date, last, amount, committee)"                           
#> [147] "sum(vtc$na_flag)"                                                                
#> [148] "```"                                                                             
#> [149] ""                                                                                
#> [150] "### Duplicates"                                                                  
#> [151] ""                                                                                
#> [152] "```{r dupe_flag}"                                                                
#> [153] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [154] "sum(vtc$dupe_flag)"                                                              
#> [155] "```"                                                                             
#> [156] ""                                                                                
#> [157] "```{r dupe_view}"                                                                
#> [158] "vtc %>% "                                                                        
#> [159] "  filter(dupe_flag) %>% "                                                        
#> [160] "  select(date, last, amount, committee)"                                         
#> [161] "```"                                                                             
#> [162] ""                                                                                
#> [163] "### Categorical"                                                                 
#> [164] ""                                                                                
#> [165] "```{r n_distinct}"                                                               
#> [166] "col_stats(vtc, n_distinct)"                                                      
#> [167] "```"                                                                             
#> [168] ""                                                                                
#> [169] "### Continuous"                                                                  
#> [170] ""                                                                                
#> [171] "#### Amounts"                                                                    
#> [172] ""                                                                                
#> [173] "```{r ammount_summary}"                                                          
#> [174] "summary(vtc$amount)"                                                             
#> [175] "mean(vtc$amount <= 0)"                                                           
#> [176] "```"                                                                             
#> [177] ""                                                                                
#> [178] "```{r hist_amount, echo=FALSE}"                                                  
#> [179] "vtc %>%"                                                                         
#> [180] "  ggplot(aes(amount)) +"                                                         
#> [181] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [182] "  scale_y_continuous(labels = comma) +"                                          
#> [183] "  scale_x_continuous("                                                           
#> [184] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [185] "    labels = dollar,"                                                            
#> [186] "    trans = \"log10\""                                                           
#> [187] "  ) +"                                                                           
#> [188] "  labs("                                                                         
#> [189] "    title = \"Vermont Contributions Amount Distribution\","                      
#> [190] "    subtitle = \"from 2000 to 2019\","                                           
#> [191] "    caption = \"Source: {source}\","                                             
#> [192] "    x = \"Amount\","                                                             
#> [193] "    y = \"Count\""                                                               
#> [194] "  )"                                                                             
#> [195] "```"                                                                             
#> [196] ""                                                                                
#> [197] "#### Dates"                                                                      
#> [198] ""                                                                                
#> [199] "```{r date_year}"                                                                
#> [200] "vtc <- mutate(vtc, year = year(date))"                                           
#> [201] "```"                                                                             
#> [202] ""                                                                                
#> [203] "```{r date_range}"                                                               
#> [204] "min(vtc$date)"                                                                   
#> [205] "sum(vtc$year < 2000)"                                                            
#> [206] "max(vtc$date)"                                                                   
#> [207] "sum(vtc$date > today())"                                                         
#> [208] "```"                                                                             
#> [209] ""                                                                                
#> [210] "```{r bar_year, echo=FALSE}"                                                     
#> [211] "vtc %>% "                                                                        
#> [212] "  count(year) %>% "                                                              
#> [213] "  mutate(even = is_even(year)) %>% "                                             
#> [214] "  ggplot(aes(x = year, y = n)) +"                                                
#> [215] "  geom_col(aes(fill = even)) + "                                                 
#> [216] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [217] "  scale_y_continuous(labels = comma) +"                                          
#> [218] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [219] "  theme(legend.position = \"bottom\") +"                                         
#> [220] "  labs("                                                                         
#> [221] "    title = \"Vermont Contributions by Year\","                                  
#> [222] "    caption = \"Source: {source}\","                                             
#> [223] "    fill = \"Election Year\","                                                   
#> [224] "    x = \"Year Made\","                                                          
#> [225] "    y = \"Count\""                                                               
#> [226] "  )"                                                                             
#> [227] "```"                                                                             
#> [228] ""                                                                                
#> [229] "## Wrangle"                                                                      
#> [230] ""                                                                                
#> [231] "To improve the searchability of the database, we will perform some consistent,"  
#> [232] "confident string normalization. For geographic variables like city names and"    
#> [233] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [234] "facilitate this process."                                                        
#> [235] ""                                                                                
#> [236] "### Address"                                                                     
#> [237] ""                                                                                
#> [238] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [239] "will force consistence case, remove punctuation, and abbreviate official "       
#> [240] "USPS suffixes."                                                                  
#> [241] ""                                                                                
#> [242] "```{r address_norm}"                                                             
#> [243] "vtc <- vtc %>% "                                                                 
#> [244] "  unite("                                                                        
#> [245] "    col = address_full,"                                                         
#> [246] "    starts_with(\"address\"),"                                                   
#> [247] "    sep = \" \","                                                                
#> [248] "    remove = FALSE,"                                                             
#> [249] "    na.rm = TRUE"                                                                
#> [250] "  ) %>% "                                                                        
#> [251] "  mutate("                                                                       
#> [252] "    address_norm = normal_address("                                              
#> [253] "      address = address_full,"                                                   
#> [254] "      abbs = usps_street,"                                                       
#> [255] "      na_rep = TRUE"                                                             
#> [256] "    )"                                                                           
#> [257] "  ) %>% "                                                                        
#> [258] "  select(-address_full)"                                                         
#> [259] "```"                                                                             
#> [260] ""                                                                                
#> [261] "```{r address_view}"                                                             
#> [262] "vtc %>% "                                                                        
#> [263] "  select(contains(\"address\")) %>% "                                            
#> [264] "  distinct() %>% "                                                               
#> [265] "  sample_n(10)"                                                                  
#> [266] "```"                                                                             
#> [267] ""                                                                                
#> [268] "### ZIP"                                                                         
#> [269] ""                                                                                
#> [270] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [271] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [272] "zeroes dropped by other programs like Microsoft Excel."                          
#> [273] ""                                                                                
#> [274] "```{r zip_norm}"                                                                 
#> [275] "vtc <- vtc %>% "                                                                 
#> [276] "  mutate("                                                                       
#> [277] "    zip_norm = normal_zip("                                                      
#> [278] "      zip = zip,"                                                                
#> [279] "      na_rep = TRUE"                                                             
#> [280] "    )"                                                                           
#> [281] "  )"                                                                             
#> [282] "```"                                                                             
#> [283] ""                                                                                
#> [284] "```{r zip_progress}"                                                             
#> [285] "progress_table("                                                                 
#> [286] "  vtc$zip,"                                                                      
#> [287] "  vtc$zip_norm,"                                                                 
#> [288] "  compare = valid_zip"                                                           
#> [289] ")"                                                                               
#> [290] "```"                                                                             
#> [291] ""                                                                                
#> [292] "### State"                                                                       
#> [293] ""                                                                                
#> [294] "Valid two digit state abbreviations can be made using the "                      
#> [295] "`campfin::normal_state()` function."                                             
#> [296] ""                                                                                
#> [297] "```{r state_norm}"                                                               
#> [298] "vtc <- vtc %>% "                                                                 
#> [299] "  mutate("                                                                       
#> [300] "    state_norm = normal_state("                                                  
#> [301] "      state = state,"                                                            
#> [302] "      abbreviate = TRUE,"                                                        
#> [303] "      na_rep = TRUE,"                                                            
#> [304] "      valid = valid_state"                                                       
#> [305] "    )"                                                                           
#> [306] "  )"                                                                             
#> [307] "```"                                                                             
#> [308] ""                                                                                
#> [309] "```{r state_view}"                                                               
#> [310] "vtc %>% "                                                                        
#> [311] "  filter(state != state_norm) %>% "                                              
#> [312] "  count(state, sort = TRUE)"                                                     
#> [313] "```"                                                                             
#> [314] ""                                                                                
#> [315] "```{r state_progress}"                                                           
#> [316] "progress_table("                                                                 
#> [317] "  vtc$state,"                                                                    
#> [318] "  vtc$state_norm,"                                                               
#> [319] "  compare = valid_state"                                                         
#> [320] ")"                                                                               
#> [321] "```"                                                                             
#> [322] ""                                                                                
#> [323] "### City"                                                                        
#> [324] ""                                                                                
#> [325] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [326] "the wide variety of valid cities and formats."                                   
#> [327] ""                                                                                
#> [328] "#### Normal"                                                                     
#> [329] ""                                                                                
#> [330] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [331] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [332] "`invalid_city` values."                                                          
#> [333] ""                                                                                
#> [334] "```{r city_norm}"                                                                
#> [335] "vtc <- vtc %>% "                                                                 
#> [336] "  mutate("                                                                       
#> [337] "    city_norm = normal_city("                                                    
#> [338] "      city = city, "                                                             
#> [339] "      abbs = usps_city,"                                                         
#> [340] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [341] "      na = invalid_city,"                                                        
#> [342] "      na_rep = TRUE"                                                             
#> [343] "    )"                                                                           
#> [344] "  )"                                                                             
#> [345] "```"                                                                             
#> [346] ""                                                                                
#> [347] "#### Swap"                                                                       
#> [348] ""                                                                                
#> [349] "We can further improve normalization by comparing our normalized value"          
#> [350] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [351] "If the normalized value is either an abbreviation for or very similar to the"    
#> [352] "expected value, we can confidently swap those two."                              
#> [353] ""                                                                                
#> [354] "```{r city_swap}"                                                                
#> [355] "vtc <- vtc %>% "                                                                 
#> [356] "  rename(city_raw = city) %>% "                                                  
#> [357] "  left_join("                                                                    
#> [358] "    y = zipcodes,"                                                               
#> [359] "    by = c("                                                                     
#> [360] "      \"state_norm\" = \"state\","                                               
#> [361] "      \"zip_norm\" = \"zip\""                                                    
#> [362] "    )"                                                                           
#> [363] "  ) %>% "                                                                        
#> [364] "  rename(city_match = city) %>% "                                                
#> [365] "  mutate("                                                                       
#> [366] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [367] "    match_dist = str_dist(city_norm, city_match),"                               
#> [368] "    city_swap = if_else("                                                        
#> [369] "      condition = !is.na(match_dist) & match_abb | match_dist == 1,"             
#> [370] "      true = city_match,"                                                        
#> [371] "      false = city_norm"                                                         
#> [372] "    )"                                                                           
#> [373] "  ) %>% "                                                                        
#> [374] "  select("                                                                       
#> [375] "    -city_match,"                                                                
#> [376] "    -match_dist,"                                                                
#> [377] "    -match_abb"                                                                  
#> [378] "  )"                                                                             
#> [379] "```"                                                                             
#> [380] ""                                                                                
#> [381] "#### Refine"                                                                     
#> [382] ""                                                                                
#> [383] "The [OpenRefine] algorithms can be used to group similar strings and replace the"
#> [384] "less common versions with their most common counterpart. This can greatly "      
#> [385] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [386] "strings that have a valid city/state/zip combination."                           
#> [387] ""                                                                                
#> [388] "[or]: https://openrefine.org/"                                                   
#> [389] ""                                                                                
#> [390] "```{r city_refine}"                                                              
#> [391] "good_refine <- vtc %>% "                                                         
#> [392] "  mutate("                                                                       
#> [393] "    city_refine = city_swap %>% "                                                
#> [394] "      key_collision_merge() %>% "                                                
#> [395] "      n_gram_merge(numgram = 1)"                                                 
#> [396] "  ) %>% "                                                                        
#> [397] "  filter(city_refine != city_swap) %>% "                                         
#> [398] "  inner_join("                                                                   
#> [399] "    y = zipcodes,"                                                               
#> [400] "    by = c("                                                                     
#> [401] "      \"city_refine\" = \"city\","                                               
#> [402] "      \"state_norm\" = \"state\","                                               
#> [403] "      \"zip_norm\" = \"zip\""                                                    
#> [404] "    )"                                                                           
#> [405] "  )"                                                                             
#> [406] "```"                                                                             
#> [407] ""                                                                                
#> [408] "```{r city_count, echo=FALSE}"                                                   
#> [409] "good_refine %>%"                                                                 
#> [410] "  count("                                                                        
#> [411] "    state_norm, "                                                                
#> [412] "    zip_norm, "                                                                  
#> [413] "    city_swap, "                                                                 
#> [414] "    city_refine,"                                                                
#> [415] "    sort = TRUE"                                                                 
#> [416] "  )"                                                                             
#> [417] "```"                                                                             
#> [418] ""                                                                                
#> [419] "Then we can join the refined values back to the database."                       
#> [420] ""                                                                                
#> [421] "```{r city_join}"                                                                
#> [422] "vtc <- vtc %>% "                                                                 
#> [423] "  left_join(good_refine) %>% "                                                   
#> [424] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [425] "```"                                                                             
#> [426] ""                                                                                
#> [427] "#### Progress"                                                                   
#> [428] ""                                                                                
#> [429] "```{r city_progress, echo=FALSE}"                                                
#> [430] "many_city <- c(valid_city, extra_city)"                                          
#> [431] "progress <- progress_table("                                                     
#> [432] "  str_to_upper(vtc$city_raw),"                                                   
#> [433] "  vtc$city_norm,"                                                                
#> [434] "  vtc$city_swap,"                                                                
#> [435] "  compare = many_city"                                                           
#> [436] ") %>% mutate(stage = as_factor(stage))"                                          
#> [437] "kable(progress, digits = 3)"                                                     
#> [438] "```"                                                                             
#> [439] ""                                                                                
#> [440] "You can see how the percentage of valid values increased with each stage."       
#> [441] ""                                                                                
#> [442] "```{r bar_progress, echo=FALSE}"                                                 
#> [443] "raw_in <- percent(prop_in(vtc$city_raw, valid_city))"                            
#> [444] "progress %>% "                                                                   
#> [445] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [446] "  geom_hline(yintercept = 0.99) +"                                               
#> [447] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [448] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [449] "  scale_y_continuous(labels = percent) +"                                        
#> [450] "  labs("                                                                         
#> [451] "    title = \"Vermont City Normalization Progress\","                            
#> [452] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [453] "    x = \"Stage\","                                                              
#> [454] "    y = \"Percent Valid\""                                                       
#> [455] "  )"                                                                             
#> [456] "```"                                                                             
#> [457] ""                                                                                
#> [458] "More importantly, the number of distinct values decreased each stage. We were"   
#> [459] "able to confidently change many distinct invalid values to their valid"          
#> [460] "equivalent."                                                                     
#> [461] ""                                                                                
#> [462] "```{r bar_distinct, echo=FALSE}"                                                 
#> [463] "progress %>% "                                                                   
#> [464] "  select("                                                                       
#> [465] "    stage, "                                                                     
#> [466] "    all = n_distinct,"                                                           
#> [467] "    bad = n_diff"                                                                
#> [468] "  ) %>% "                                                                        
#> [469] "  mutate(good = all - bad) %>% "                                                 
#> [470] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [471] "  mutate(name = name == \"good\") %>% "                                          
#> [472] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [473] "  geom_col(aes(fill = name)) +"                                                  
#> [474] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [475] "  scale_y_continuous(labels = comma) +"                                          
#> [476] "  theme(legend.position = \"bottom\") +"                                         
#> [477] "  labs("                                                                         
#> [478] "    title = \"Vermont City Normalization Progress\","                            
#> [479] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [480] "    x = \"Stage\","                                                              
#> [481] "    y = \"Distinct Values\","                                                    
#> [482] "    fill = \"Valid\""                                                            
#> [483] "  )"                                                                             
#> [484] "```"                                                                             
#> [485] ""                                                                                
#> [486] "## Conclude"                                                                     
#> [487] ""                                                                                
#> [488] "```{r clean_select}"                                                             
#> [489] "vtc <- vtc %>% "                                                                 
#> [490] "  select("                                                                       
#> [491] "    -city_norm,"                                                                 
#> [492] "    -city_swap,"                                                                 
#> [493] "    city_clean = city_refine"                                                    
#> [494] "  ) %>% "                                                                        
#> [495] "  rename_all(~str_replace(., \"_norm\", \"_clean\"))"                            
#> [496] "```"                                                                             
#> [497] ""                                                                                
#> [498] "```{r clean_glimpse}"                                                            
#> [499] "glimpse(sample_n(vtc, 20))"                                                      
#> [500] "```"                                                                             
#> [501] ""                                                                                
#> [502] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [503] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [504] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [505] "1. There are `r comma(sum(vtc$na_flag))` records missing ...."                   
#> [506] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [507] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [508] ""                                                                                
#> [509] "## Export"                                                                       
#> [510] ""                                                                                
#> [511] "```{r clean_dir}"                                                                
#> [512] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [513] "```"                                                                             
#> [514] ""                                                                                
#> [515] "```{r clean_write}"                                                              
#> [516] "write_csv("                                                                      
#> [517] "  x = vtc,"                                                                      
#> [518] "  path = path(clean_dir, \"vt_contribs_clean.csv\"),"                            
#> [519] "  na = \"\""                                                                     
#> [520] ")"                                                                               
#> [521] "```"                                                                             
#> [522] ""                                                                                

# use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())