Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type, author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contracts\""                                                    
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r Sys.time()`\""                                                        
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 3"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "if (!interactive()) {"                                                           
#>  [31] "  options(width = 99)"                                                           
#>  [32] "  set.seed(5)"                                                                   
#>  [33] "}"                                                                               
#>  [34] "```"                                                                             
#>  [35] ""                                                                                
#>  [36] "```{r create_docs_dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [37] "doc_dir <- fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"           
#>  [38] "```"                                                                             
#>  [39] ""                                                                                
#>  [40] "## Project"                                                                      
#>  [41] ""                                                                                
#>  [42] "The Accountability Project is an effort to cut across data silos and give"       
#>  [43] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [44] "way to search across huge volumes of public data about people and organizations."
#>  [45] ""                                                                                
#>  [46] "Our goal is to standardizing public data on a few key fields by thinking of each"
#>  [47] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [48] "variables:"                                                                      
#>  [49] ""                                                                                
#>  [50] "1. All **parties** to a transaction."                                            
#>  [51] "2. The **date** of the transaction."                                             
#>  [52] "3. The **amount** of money involved."                                            
#>  [53] ""                                                                                
#>  [54] "## Objectives"                                                                   
#>  [55] ""                                                                                
#>  [56] "This document describes the process used to complete the following objectives:"  
#>  [57] ""                                                                                
#>  [58] "1. How many records are in the database?"                                        
#>  [59] "1. Check for entirely duplicated records."                                       
#>  [60] "1. Check ranges of continuous variables."                                        
#>  [61] "1. Is there anything blank or missing?"                                          
#>  [62] "1. Check for consistency issues."                                                
#>  [63] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [64] "1. Create a `year` field from the transaction date."                             
#>  [65] "1. Make sure there is data on both parties to a transaction."                    
#>  [66] ""                                                                                
#>  [67] "## Packages"                                                                     
#>  [68] ""                                                                                
#>  [69] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [70] "and communicate these results. The `pacman` package will facilitate their"       
#>  [71] "installation and attachment."                                                    
#>  [72] ""                                                                                
#>  [73] "The IRW's `campfin` package will also have to be installed from GitHub. This"    
#>  [74] "package contains functions custom made to help facilitate the processing of"     
#>  [75] "campaign finance data."                                                          
#>  [76] ""                                                                                
#>  [77] "```{r load_packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [78] "if (!require(\"pacman\")) install.packages(\"pacman\")"                          
#>  [79] "pacman::p_load_gh(\"irworkshop/campfin\")"                                       
#>  [80] "pacman::p_load("                                                                 
#>  [81] "  tidyverse, # data manipulation"                                                
#>  [82] "  lubridate, # datetime strings"                                                 
#>  [83] "  gluedown, # printing markdown"                                                 
#>  [84] "  magrittr, # pipe operators"                                                    
#>  [85] "  janitor, # clean data frames"                                                  
#>  [86] "  aws.s3, # upload to aws s3"                                                    
#>  [87] "  refinr, # cluster and merge"                                                   
#>  [88] "  scales, # format strings"                                                      
#>  [89] "  knitr, # knit documents"                                                       
#>  [90] "  vroom, # read files fast"                                                      
#>  [91] "  rvest, # html scraping"                                                        
#>  [92] "  glue, # combine strings"                                                       
#>  [93] "  here, # relative paths"                                                        
#>  [94] "  httr, # http requests"                                                         
#>  [95] "  fs # local storage "                                                           
#>  [96] ")"                                                                               
#>  [97] "```"                                                                             
#>  [98] ""                                                                                
#>  [99] "This document should be run as part of the `R_campfin` project, which lives as a"
#> [100] "sub-directory of the more general, language-agnostic"                            
#> [101] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#> [102] ""                                                                                
#> [103] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#> [104] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#> [105] "paths relative to _your_ machine."                                               
#> [106] ""                                                                                
#> [107] "```{r where_here}"                                                               
#> [108] "# where does this document knit?"                                                
#> [109] "here::here()"                                                                    
#> [110] "```"                                                                             
#> [111] ""                                                                                
#> [112] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [113] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [114] ""                                                                                
#> [115] "## Data"                                                                         
#> [116] ""                                                                                
#> [117] ""                                                                                
#> [118] ""                                                                                
#> [119] "## Download"                                                                     
#> [120] ""                                                                                
#> [121] "```{r raw_dir}"                                                                  
#> [122] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [123] "```"                                                                             
#> [124] ""                                                                                
#> [125] "```{r raw_download}"                                                             
#> [126] "download.file()"                                                                 
#> [127] "```"                                                                             
#> [128] ""                                                                                
#> [129] "## Read"                                                                         
#> [130] ""                                                                                
#> [131] "```{r raw_read}"                                                                 
#> [132] "vtc <- read_delim("                                                              
#> [133] "  file = dir_ls(raw_dir)"                                                        
#> [134] ")"                                                                               
#> [135] "```"                                                                             
#> [136] ""                                                                                
#> [137] "## Explore"                                                                      
#> [138] ""                                                                                
#> [139] "There are `r scales::comma(nrow(vtc))` rows of `r ncol(vtc)` columns."           
#> [140] ""                                                                                
#> [141] "```{r glimpse}"                                                                  
#> [142] "glimpse(vtc)"                                                                    
#> [143] "tail(vtc)"                                                                       
#> [144] "```"                                                                             
#> [145] ""                                                                                
#> [146] "### Missing"                                                                     
#> [147] ""                                                                                
#> [148] "Columns vary in their degree of missing values."                                 
#> [149] ""                                                                                
#> [150] "```{r na_count}"                                                                 
#> [151] "col_stats(vtc, count_na)"                                                        
#> [152] "```"                                                                             
#> [153] ""                                                                                
#> [154] "We can flag any record missing a key variable needed to identify a transaction." 
#> [155] ""                                                                                
#> [156] "```{r na_flag}"                                                                  
#> [157] "vtc <- vtc %>% flag_na(date, last, amount, committee)"                           
#> [158] "sum(vtc$na_flag)"                                                                
#> [159] "```"                                                                             
#> [160] ""                                                                                
#> [161] "```{r na_view}"                                                                  
#> [162] "vtc %>% "                                                                        
#> [163] "  filter(na_flag) %>% "                                                          
#> [164] "  select(date, last, amount, committee)"                                         
#> [165] "```"                                                                             
#> [166] ""                                                                                
#> [167] "### Duplicates"                                                                  
#> [168] ""                                                                                
#> [169] "We can also flag any record completely duplicated across every column."          
#> [170] ""                                                                                
#> [171] "```{r dupe_flag}"                                                                
#> [172] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [173] "sum(vtc$dupe_flag)"                                                              
#> [174] "```"                                                                             
#> [175] ""                                                                                
#> [176] "```{r dupe_view}"                                                                
#> [177] "vtc %>% "                                                                        
#> [178] "  filter(dupe_flag) %>% "                                                        
#> [179] "  select(date, last, amount, committee)"                                         
#> [180] "```"                                                                             
#> [181] ""                                                                                
#> [182] "### Categorical"                                                                 
#> [183] ""                                                                                
#> [184] "```{r distinct_count}"                                                           
#> [185] "col_stats(vtc, n_distinct)"                                                      
#> [186] "```"                                                                             
#> [187] ""                                                                                
#> [188] "```{r distinct_plots, echo=FALSE}"                                               
#> [189] "explore_plot(vtc, type)"                                                         
#> [190] "```"                                                                             
#> [191] ""                                                                                
#> [192] "### Amounts"                                                                     
#> [193] ""                                                                                
#> [194] "```{r amount_summary}"                                                           
#> [195] "summary(vtc$amount)"                                                             
#> [196] "mean(vtc$amount <= 0)"                                                           
#> [197] "```"                                                                             
#> [198] ""                                                                                
#> [199] "These are the records with the minimum and maximum amounts."                     
#> [200] ""                                                                                
#> [201] "```{r amount_minmax}"                                                            
#> [202] "glimpse(vtc[c(which.max(vtc$amount), which.min(vtc$amount)), ])"                 
#> [203] "```"                                                                             
#> [204] ""                                                                                
#> [205] "```{r hist_amount, echo=FALSE}"                                                  
#> [206] "vtc %>%"                                                                         
#> [207] "  ggplot(aes(amount)) +"                                                         
#> [208] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [209] "  scale_y_continuous(labels = comma) +"                                          
#> [210] "  scale_x_continuous("                                                           
#> [211] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [212] "    labels = dollar,"                                                            
#> [213] "    trans = \"log10\""                                                           
#> [214] "  ) +"                                                                           
#> [215] "  labs("                                                                         
#> [216] "    title = \"Vermont Contracts Amount Distribution\","                          
#> [217] "    caption = \"Source: {source}\","                                             
#> [218] "    x = \"Amount\","                                                             
#> [219] "    y = \"Count\""                                                               
#> [220] "  )"                                                                             
#> [221] "```"                                                                             
#> [222] ""                                                                                
#> [223] "### Dates"                                                                       
#> [224] ""                                                                                
#> [225] "We can add the calendar year from `date` with `lubridate::year()`"               
#> [226] ""                                                                                
#> [227] "```{r date_year}"                                                                
#> [228] "vtc <- mutate(vtc, year = year(date))"                                           
#> [229] "```"                                                                             
#> [230] ""                                                                                
#> [231] "```{r date_range}"                                                               
#> [232] "min(vtc$date)"                                                                   
#> [233] "sum(vtc$year < 2000)"                                                            
#> [234] "max(vtc$date)"                                                                   
#> [235] "sum(vtc$date > today())"                                                         
#> [236] "```"                                                                             
#> [237] ""                                                                                
#> [238] "```{r bar_year, echo=FALSE}"                                                     
#> [239] "vtc %>% "                                                                        
#> [240] "  count(year) %>% "                                                              
#> [241] "  mutate(even = is_even(year)) %>% "                                             
#> [242] "  ggplot(aes(x = year, y = n)) +"                                                
#> [243] "  geom_col(aes(fill = even)) + "                                                 
#> [244] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [245] "  scale_y_continuous(labels = comma) +"                                          
#> [246] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [247] "  theme(legend.position = \"bottom\") +"                                         
#> [248] "  labs("                                                                         
#> [249] "    title = \"Vermont Contracts by Year\","                                      
#> [250] "    caption = \"Source: {source}\","                                             
#> [251] "    fill = \"Election Year\","                                                   
#> [252] "    x = \"Year Made\","                                                          
#> [253] "    y = \"Count\""                                                               
#> [254] "  )"                                                                             
#> [255] "```"                                                                             
#> [256] ""                                                                                
#> [257] "## Wrangle"                                                                      
#> [258] ""                                                                                
#> [259] "To improve the searchability of the database, we will perform some consistent,"  
#> [260] "confident string normalization. For geographic variables like city names and"    
#> [261] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [262] "facilitate this process."                                                        
#> [263] ""                                                                                
#> [264] "### Address"                                                                     
#> [265] ""                                                                                
#> [266] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [267] "will force consistence case, remove punctuation, and abbreviate official "       
#> [268] "USPS suffixes."                                                                  
#> [269] ""                                                                                
#> [270] "```{r address_norm}"                                                             
#> [271] "vtc <- vtc %>% "                                                                 
#> [272] "  unite("                                                                        
#> [273] "    col = address_full,"                                                         
#> [274] "    starts_with(\"address\"),"                                                   
#> [275] "    sep = \" \","                                                                
#> [276] "    remove = FALSE,"                                                             
#> [277] "    na.rm = TRUE"                                                                
#> [278] "  ) %>% "                                                                        
#> [279] "  mutate("                                                                       
#> [280] "    address_norm = normal_address("                                              
#> [281] "      address = address_full,"                                                   
#> [282] "      abbs = usps_street,"                                                       
#> [283] "      na_rep = TRUE"                                                             
#> [284] "    )"                                                                           
#> [285] "  ) %>% "                                                                        
#> [286] "  select(-address_full)"                                                         
#> [287] "```"                                                                             
#> [288] ""                                                                                
#> [289] "```{r address_view}"                                                             
#> [290] "vtc %>% "                                                                        
#> [291] "  select(contains(\"address\")) %>% "                                            
#> [292] "  distinct() %>% "                                                               
#> [293] "  sample_n(10)"                                                                  
#> [294] "```"                                                                             
#> [295] ""                                                                                
#> [296] "### ZIP"                                                                         
#> [297] ""                                                                                
#> [298] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [299] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [300] "zeroes dropped by other programs like Microsoft Excel."                          
#> [301] ""                                                                                
#> [302] "```{r zip_norm}"                                                                 
#> [303] "vtc <- vtc %>% "                                                                 
#> [304] "  mutate("                                                                       
#> [305] "    zip_norm = normal_zip("                                                      
#> [306] "      zip = zip,"                                                                
#> [307] "      na_rep = TRUE"                                                             
#> [308] "    )"                                                                           
#> [309] "  )"                                                                             
#> [310] "```"                                                                             
#> [311] ""                                                                                
#> [312] "```{r zip_progress}"                                                             
#> [313] "progress_table("                                                                 
#> [314] "  vtc$zip,"                                                                      
#> [315] "  vtc$zip_norm,"                                                                 
#> [316] "  compare = valid_zip"                                                           
#> [317] ")"                                                                               
#> [318] "```"                                                                             
#> [319] ""                                                                                
#> [320] "### State"                                                                       
#> [321] ""                                                                                
#> [322] "Valid two digit state abbreviations can be made using the "                      
#> [323] "`campfin::normal_state()` function."                                             
#> [324] ""                                                                                
#> [325] "```{r state_norm}"                                                               
#> [326] "vtc <- vtc %>% "                                                                 
#> [327] "  mutate("                                                                       
#> [328] "    state_norm = normal_state("                                                  
#> [329] "      state = state,"                                                            
#> [330] "      abbreviate = TRUE,"                                                        
#> [331] "      na_rep = TRUE,"                                                            
#> [332] "      valid = valid_state"                                                       
#> [333] "    )"                                                                           
#> [334] "  )"                                                                             
#> [335] "```"                                                                             
#> [336] ""                                                                                
#> [337] "```{r state_view}"                                                               
#> [338] "vtc %>% "                                                                        
#> [339] "  filter(state != state_norm) %>% "                                              
#> [340] "  count(state, state_norm, sort = TRUE)"                                         
#> [341] "```"                                                                             
#> [342] ""                                                                                
#> [343] "```{r state_progress}"                                                           
#> [344] "progress_table("                                                                 
#> [345] "  vtc$state,"                                                                    
#> [346] "  vtc$state_norm,"                                                               
#> [347] "  compare = valid_state"                                                         
#> [348] ")"                                                                               
#> [349] "```"                                                                             
#> [350] ""                                                                                
#> [351] "### City"                                                                        
#> [352] ""                                                                                
#> [353] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [354] "the wide variety of valid cities and formats."                                   
#> [355] ""                                                                                
#> [356] "#### Normal"                                                                     
#> [357] ""                                                                                
#> [358] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [359] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [360] "`invalid_city` values."                                                          
#> [361] ""                                                                                
#> [362] "```{r city_norm}"                                                                
#> [363] "vtc <- vtc %>% "                                                                 
#> [364] "  mutate("                                                                       
#> [365] "    city_norm = normal_city("                                                    
#> [366] "      city = city, "                                                             
#> [367] "      abbs = usps_city,"                                                         
#> [368] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [369] "      na = invalid_city,"                                                        
#> [370] "      na_rep = TRUE"                                                             
#> [371] "    )"                                                                           
#> [372] "  )"                                                                             
#> [373] "```"                                                                             
#> [374] ""                                                                                
#> [375] "#### Swap"                                                                       
#> [376] ""                                                                                
#> [377] "We can further improve normalization by comparing our normalized value"          
#> [378] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [379] "If the normalized value is either an abbreviation for or very similar to the"    
#> [380] "expected value, we can confidently swap those two."                              
#> [381] ""                                                                                
#> [382] "```{r city_swap}"                                                                
#> [383] "vtc <- vtc %>% "                                                                 
#> [384] "  rename(city_raw = city) %>% "                                                  
#> [385] "  left_join("                                                                    
#> [386] "    y = zipcodes,"                                                               
#> [387] "    by = c("                                                                     
#> [388] "      \"state_norm\" = \"state\","                                               
#> [389] "      \"zip_norm\" = \"zip\""                                                    
#> [390] "    )"                                                                           
#> [391] "  ) %>% "                                                                        
#> [392] "  rename(city_match = city) %>% "                                                
#> [393] "  mutate("                                                                       
#> [394] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [395] "    match_dist = str_dist(city_norm, city_match),"                               
#> [396] "    city_swap = if_else("                                                        
#> [397] "      condition = !is.na(match_dist) & (match_abb | match_dist == 1),"           
#> [398] "      true = city_match,"                                                        
#> [399] "      false = city_norm"                                                         
#> [400] "    )"                                                                           
#> [401] "  ) %>% "                                                                        
#> [402] "  select("                                                                       
#> [403] "    -city_match,"                                                                
#> [404] "    -match_dist,"                                                                
#> [405] "    -match_abb"                                                                  
#> [406] "  )"                                                                             
#> [407] "```"                                                                             
#> [408] ""                                                                                
#> [409] "#### Refine"                                                                     
#> [410] ""                                                                                
#> [411] "The [OpenRefine][or] algorithms can be used to group similar strings and replace"
#> [412] "the less common versions with their most common counterpart. This can greatly"   
#> [413] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [414] "strings that have a valid city/state/zip combination."                           
#> [415] ""                                                                                
#> [416] "[or]: https://openrefine.org/"                                                   
#> [417] ""                                                                                
#> [418] "```{r city_refine}"                                                              
#> [419] "good_refine <- vtc %>% "                                                         
#> [420] "  mutate("                                                                       
#> [421] "    city_refine = city_swap %>% "                                                
#> [422] "      key_collision_merge() %>% "                                                
#> [423] "      n_gram_merge(numgram = 1)"                                                 
#> [424] "  ) %>% "                                                                        
#> [425] "  filter(city_refine != city_swap) %>% "                                         
#> [426] "  inner_join("                                                                   
#> [427] "    y = zipcodes,"                                                               
#> [428] "    by = c("                                                                     
#> [429] "      \"city_refine\" = \"city\","                                               
#> [430] "      \"state_norm\" = \"state\","                                               
#> [431] "      \"zip_norm\" = \"zip\""                                                    
#> [432] "    )"                                                                           
#> [433] "  )"                                                                             
#> [434] "```"                                                                             
#> [435] ""                                                                                
#> [436] "```{r city_count, echo=FALSE}"                                                   
#> [437] "good_refine %>%"                                                                 
#> [438] "  count("                                                                        
#> [439] "    state_norm, "                                                                
#> [440] "    zip_norm, "                                                                  
#> [441] "    city_swap, "                                                                 
#> [442] "    city_refine,"                                                                
#> [443] "    sort = TRUE"                                                                 
#> [444] "  )"                                                                             
#> [445] "```"                                                                             
#> [446] ""                                                                                
#> [447] "Then we can join the refined values back to the database."                       
#> [448] ""                                                                                
#> [449] "```{r city_join}"                                                                
#> [450] "vtc <- vtc %>% "                                                                 
#> [451] "  left_join(good_refine) %>% "                                                   
#> [452] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [453] "```"                                                                             
#> [454] ""                                                                                
#> [455] "#### Progress"                                                                   
#> [456] ""                                                                                
#> [457] "Our goal for normalization was to increase the proportion of city values known"  
#> [458] "to be valid and reduce the total distinct values by correcting misspellings."    
#> [459] ""                                                                                
#> [460] "```{r city_progress, echo=FALSE}"                                                
#> [461] "many_city <- c(valid_city, extra_city)"                                          
#> [462] "progress <- progress_table("                                                     
#> [463] "  str_to_upper(vtc$city_raw),"                                                   
#> [464] "  vtc$city_norm,"                                                                
#> [465] "  vtc$city_swap,"                                                                
#> [466] "  vtc$city_refine,"                                                              
#> [467] "  compare = many_city"                                                           
#> [468] ") %>% mutate(stage = as_factor(stage))"                                          
#> [469] "kable(progress, digits = 3)"                                                     
#> [470] "```"                                                                             
#> [471] ""                                                                                
#> [472] "You can see how the percentage of valid values increased with each stage."       
#> [473] ""                                                                                
#> [474] "```{r bar_progress, echo=FALSE}"                                                 
#> [475] "raw_in <- percent(prop_in(vtc$city_raw, valid_city))"                            
#> [476] "progress %>% "                                                                   
#> [477] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [478] "  geom_hline(yintercept = 0.99) +"                                               
#> [479] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [480] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [481] "  scale_y_continuous(labels = percent) +"                                        
#> [482] "  labs("                                                                         
#> [483] "    title = \"Vermont City Normalization Progress\","                            
#> [484] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [485] "    x = \"Stage\","                                                              
#> [486] "    y = \"Percent Valid\""                                                       
#> [487] "  )"                                                                             
#> [488] "```"                                                                             
#> [489] ""                                                                                
#> [490] "More importantly, the number of distinct values decreased each stage. We were"   
#> [491] "able to confidently change many distinct invalid values to their valid"          
#> [492] "equivalent."                                                                     
#> [493] ""                                                                                
#> [494] "```{r bar_distinct, echo=FALSE}"                                                 
#> [495] "progress %>% "                                                                   
#> [496] "  select("                                                                       
#> [497] "    stage, "                                                                     
#> [498] "    all = n_distinct,"                                                           
#> [499] "    bad = n_diff"                                                                
#> [500] "  ) %>% "                                                                        
#> [501] "  mutate(good = all - bad) %>% "                                                 
#> [502] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [503] "  mutate(name = name == \"good\") %>% "                                          
#> [504] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [505] "  geom_col(aes(fill = name)) +"                                                  
#> [506] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [507] "  scale_y_continuous(labels = comma) +"                                          
#> [508] "  theme(legend.position = \"bottom\") +"                                         
#> [509] "  labs("                                                                         
#> [510] "    title = \"Vermont City Normalization Progress\","                            
#> [511] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [512] "    x = \"Stage\","                                                              
#> [513] "    y = \"Distinct Values\","                                                    
#> [514] "    fill = \"Valid\""                                                            
#> [515] "  )"                                                                             
#> [516] "```"                                                                             
#> [517] ""                                                                                
#> [518] "## Conclude"                                                                     
#> [519] ""                                                                                
#> [520] "Before exporting, we can remove the intermediary normalization columns and"      
#> [521] "rename all added variables with the `_clean` suffix."                            
#> [522] ""                                                                                
#> [523] "```{r clean_select}"                                                             
#> [524] "vtc <- vtc %>% "                                                                 
#> [525] "  select("                                                                       
#> [526] "    -city_norm,"                                                                 
#> [527] "    -city_swap,"                                                                 
#> [528] "    city_clean = city_refine"                                                    
#> [529] "  ) %>% "                                                                        
#> [530] "  rename_all(~str_replace(., \"_norm\", \"_clean\")) %>% "                       
#> [531] "  rename_all(~str_remove(., \"_raw\"))"                                          
#> [532] "```"                                                                             
#> [533] ""                                                                                
#> [534] "```{r clean_glimpse}"                                                            
#> [535] "glimpse(sample_n(vtc, 50))"                                                      
#> [536] "```"                                                                             
#> [537] ""                                                                                
#> [538] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [539] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [540] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [541] "1. There are `r comma(sum(vtc$na_flag))` records missing key variables."         
#> [542] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [543] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [544] ""                                                                                
#> [545] "## Export"                                                                       
#> [546] ""                                                                                
#> [547] "Now the file can be saved on disk for upload to the Accountability server."      
#> [548] ""                                                                                
#> [549] "```{r clean_dir}"                                                                
#> [550] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [551] "clean_path <- path(clean_dir, \"vt_contribs_clean.csv\")"                        
#> [552] "write_csv(vtc, clean_path, na = \"\")"                                           
#> [553] "(clean_size <- file_size(clean_path))"                                           
#> [554] "file_encoding(clean_path) %>% "                                                  
#> [555] "  mutate(across(path, path.abbrev))"                                             
#> [556] "```"                                                                             
#> [557] ""                                                                                
#> [558] "## Upload"                                                                       
#> [559] ""                                                                                
#> [560] "We can use the `aws.s3::put_object()` to upload the text file to the IRW server."
#> [561] ""                                                                                
#> [562] "```{r aws_upload, eval=FALSE}"                                                   
#> [563] "aws_path <- path(\"csv\", basename(clean_path))"                                 
#> [564] "if (!object_exists(aws_path, \"publicaccountability\")) {"                       
#> [565] "  put_object("                                                                   
#> [566] "    file = clean_path,"                                                          
#> [567] "    object = aws_path, "                                                         
#> [568] "    bucket = \"publicaccountability\","                                          
#> [569] "    acl = \"public-read\","                                                      
#> [570] "    show_progress = TRUE,"                                                       
#> [571] "    multipart = TRUE"                                                            
#> [572] "  )"                                                                             
#> [573] "}"                                                                               
#> [574] "aws_head <- head_object(aws_path, \"publicaccountability\")"                     
#> [575] "(aws_size <- as_fs_bytes(attr(aws_head, \"content-length\")))"                   
#> [576] "unname(aws_size == clean_size)"                                                  
#> [577] "```"                                                                             
#> [578] ""                                                                                
#> [579] "## Dictionary"                                                                   
#> [580] ""                                                                                
#> [581] "The following table describes the variables in our final exported file:"         
#> [582] ""                                                                                
#> [583] "```{r dict_make, echo=FALSE}"                                                    
#> [584] "dict_raw <- tibble("                                                             
#> [585] "  var = md_code(names(vtc)),"                                                    
#> [586] "  type = md_code(map_chr(vtc, typeof)),"                                         
#> [587] "  def = c("                                                                      
#> [588] "    \"\""                                                                        
#> [589] "  )"                                                                             
#> [590] ")"                                                                               
#> [591] "```"                                                                             
#> [592] ""                                                                                
#> [593] "```{r dict_md, echo=FALSE}"                                                      
#> [594] "(dict_md <- kable("                                                              
#> [595] "  x = dict_raw,"                                                                 
#> [596] "  format = \"markdown\","                                                        
#> [597] "  col.names = c(\"Column\", \"Type\", \"Definition\")"                           
#> [598] "))"                                                                              
#> [599] "```"                                                                             

use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())

#> Warning: The `path` argument of `write_lines()` is deprecated as of readr 1.4.0.
#> Please use the `file` argument instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp5MHeCf/vt_contribs_diary.Rmd was created