Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type, author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contracts\""                                                    
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r Sys.time()`\""                                                        
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 2"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "if (!interactive()) {"                                                           
#>  [31] "  options(width = 99)"                                                           
#>  [32] "  set.seed(5)"                                                                   
#>  [33] "}"                                                                               
#>  [34] "```"                                                                             
#>  [35] ""                                                                                
#>  [36] "```{r create_docs_dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [37] "doc_dir <- fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"           
#>  [38] "```"                                                                             
#>  [39] ""                                                                                
#>  [40] "## Project"                                                                      
#>  [41] ""                                                                                
#>  [42] "The Accountability Project is an effort to cut across data silos and give"       
#>  [43] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [44] "way to search across huge volumes of public data about people and organizations."
#>  [45] ""                                                                                
#>  [46] "Our goal is to standardizing public data on a few key fields by thinking of each"
#>  [47] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [48] "variables:"                                                                      
#>  [49] ""                                                                                
#>  [50] "1. All **parties** to a transaction."                                            
#>  [51] "2. The **date** of the transaction."                                             
#>  [52] "3. The **amount** of money involved."                                            
#>  [53] ""                                                                                
#>  [54] "## Objectives"                                                                   
#>  [55] ""                                                                                
#>  [56] "This document describes the process used to complete the following objectives:"  
#>  [57] ""                                                                                
#>  [58] "1. How many records are in the database?"                                        
#>  [59] "1. Check for entirely duplicated records."                                       
#>  [60] "1. Check ranges of continuous variables."                                        
#>  [61] "1. Is there anything blank or missing?"                                          
#>  [62] "1. Check for consistency issues."                                                
#>  [63] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [64] "1. Create a `year` field from the transaction date."                             
#>  [65] "1. Make sure there is data on both parties to a transaction."                    
#>  [66] ""                                                                                
#>  [67] "## Packages"                                                                     
#>  [68] ""                                                                                
#>  [69] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [70] "and communicate these results. The `pacman` package will facilitate their"       
#>  [71] "installation and attachment."                                                    
#>  [72] ""                                                                                
#>  [73] "The IRW's `campfin` package will also have to be installed from GitHub. This"    
#>  [74] "package contains functions custom made to help facilitate the processing of"     
#>  [75] "campaign finance data."                                                          
#>  [76] ""                                                                                
#>  [77] "```{r load_packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [78] "if (!require(\"pacman\")) install.packages(\"pacman\")"                          
#>  [79] "pacman::p_load_gh(\"irworkshop/campfin\")"                                       
#>  [80] "pacman::p_load("                                                                 
#>  [81] "  tidyverse, # data manipulation"                                                
#>  [82] "  lubridate, # datetime strings"                                                 
#>  [83] "  gluedown, # printing markdown"                                                 
#>  [84] "  magrittr, # pipe operators"                                                    
#>  [85] "  janitor, # clean data frames"                                                  
#>  [86] "  refinr, # cluster and merge"                                                   
#>  [87] "  scales, # format strings"                                                      
#>  [88] "  knitr, # knit documents"                                                       
#>  [89] "  vroom, # read files fast"                                                      
#>  [90] "  rvest, # html scraping"                                                        
#>  [91] "  glue, # combine strings"                                                       
#>  [92] "  here, # relative paths"                                                        
#>  [93] "  httr, # http requests"                                                         
#>  [94] "  fs # local storage "                                                           
#>  [95] ")"                                                                               
#>  [96] "```"                                                                             
#>  [97] ""                                                                                
#>  [98] "This document should be run as part of the `R_campfin` project, which lives as a"
#>  [99] "sub-directory of the more general, language-agnostic"                            
#> [100] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#> [101] ""                                                                                
#> [102] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#> [103] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#> [104] "paths relative to _your_ machine."                                               
#> [105] ""                                                                                
#> [106] "```{r where_here}"                                                               
#> [107] "# where does this document knit?"                                                
#> [108] "here::here()"                                                                    
#> [109] "```"                                                                             
#> [110] ""                                                                                
#> [111] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [112] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [113] ""                                                                                
#> [114] "## Data"                                                                         
#> [115] ""                                                                                
#> [116] "## Download"                                                                     
#> [117] ""                                                                                
#> [118] "```{r raw_dir}"                                                                  
#> [119] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [120] "```"                                                                             
#> [121] ""                                                                                
#> [122] "```{r raw_download}"                                                             
#> [123] "download.file()"                                                                 
#> [124] "```"                                                                             
#> [125] ""                                                                                
#> [126] "## Read"                                                                         
#> [127] ""                                                                                
#> [128] "```{r raw_read}"                                                                 
#> [129] "vtc <- read_delim("                                                              
#> [130] "  file = dir_ls(raw_dir)"                                                        
#> [131] ")"                                                                               
#> [132] "```"                                                                             
#> [133] ""                                                                                
#> [134] "## Explore"                                                                      
#> [135] ""                                                                                
#> [136] "```{r glimpse}"                                                                  
#> [137] "glimpse(vtc)"                                                                    
#> [138] "tail(vtc)"                                                                       
#> [139] "```"                                                                             
#> [140] ""                                                                                
#> [141] "### Missing"                                                                     
#> [142] ""                                                                                
#> [143] "```{r na_count}"                                                                 
#> [144] "col_stats(vtc, count_na)"                                                        
#> [145] "```"                                                                             
#> [146] ""                                                                                
#> [147] "```{r na_flag}"                                                                  
#> [148] "vtc <- vtc %>% flag_na(date, last, amount, committee)"                           
#> [149] "sum(vtc$na_flag)"                                                                
#> [150] "```"                                                                             
#> [151] ""                                                                                
#> [152] "```{r na_view}"                                                                  
#> [153] "vtc %>% "                                                                        
#> [154] "  filter(na_flag) %>% "                                                          
#> [155] "  select(date, last, amount, committee)"                                         
#> [156] "```"                                                                             
#> [157] ""                                                                                
#> [158] "### Duplicates"                                                                  
#> [159] ""                                                                                
#> [160] "```{r dupe_flag}"                                                                
#> [161] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [162] "sum(vtc$dupe_flag)"                                                              
#> [163] "```"                                                                             
#> [164] ""                                                                                
#> [165] "```{r dupe_view}"                                                                
#> [166] "vtc %>% "                                                                        
#> [167] "  filter(dupe_flag) %>% "                                                        
#> [168] "  select(date, last, amount, committee)"                                         
#> [169] "```"                                                                             
#> [170] ""                                                                                
#> [171] "### Categorical"                                                                 
#> [172] ""                                                                                
#> [173] "```{r distinct_count}"                                                           
#> [174] "col_stats(vtc, n_distinct)"                                                      
#> [175] "```"                                                                             
#> [176] ""                                                                                
#> [177] "```{r distinct_plots}"                                                           
#> [178] "explore_plot(vtc, type)"                                                         
#> [179] "```"                                                                             
#> [180] ""                                                                                
#> [181] "### Amounts"                                                                     
#> [182] ""                                                                                
#> [183] "```{r ammount_summary}"                                                          
#> [184] "summary(vtc$amount)"                                                             
#> [185] "mean(vtc$amount <= 0)"                                                           
#> [186] "```"                                                                             
#> [187] ""                                                                                
#> [188] "```{r hist_amount, echo=FALSE}"                                                  
#> [189] "vtc %>%"                                                                         
#> [190] "  ggplot(aes(amount)) +"                                                         
#> [191] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [192] "  scale_y_continuous(labels = comma) +"                                          
#> [193] "  scale_x_continuous("                                                           
#> [194] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [195] "    labels = dollar,"                                                            
#> [196] "    trans = \"log10\""                                                           
#> [197] "  ) +"                                                                           
#> [198] "  labs("                                                                         
#> [199] "    title = \"Vermont Contracts Amount Distribution\","                          
#> [200] "    subtitle = \"from 2000 to 2019\","                                           
#> [201] "    caption = \"Source: {source}\","                                             
#> [202] "    x = \"Amount\","                                                             
#> [203] "    y = \"Count\""                                                               
#> [204] "  )"                                                                             
#> [205] "```"                                                                             
#> [206] ""                                                                                
#> [207] "### Dates"                                                                       
#> [208] ""                                                                                
#> [209] "We can add the calendar year from `date` with `lubridate::year()`"               
#> [210] ""                                                                                
#> [211] "```{r date_year}"                                                                
#> [212] "vtc <- mutate(vtc, year = year(date))"                                           
#> [213] "```"                                                                             
#> [214] ""                                                                                
#> [215] "```{r date_range}"                                                               
#> [216] "min(vtc$date)"                                                                   
#> [217] "sum(vtc$year < 2000)"                                                            
#> [218] "max(vtc$date)"                                                                   
#> [219] "sum(vtc$date > today())"                                                         
#> [220] "```"                                                                             
#> [221] ""                                                                                
#> [222] "```{r bar_year, echo=FALSE}"                                                     
#> [223] "vtc %>% "                                                                        
#> [224] "  count(year) %>% "                                                              
#> [225] "  mutate(even = is_even(year)) %>% "                                             
#> [226] "  ggplot(aes(x = year, y = n)) +"                                                
#> [227] "  geom_col(aes(fill = even)) + "                                                 
#> [228] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [229] "  scale_y_continuous(labels = comma) +"                                          
#> [230] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [231] "  theme(legend.position = \"bottom\") +"                                         
#> [232] "  labs("                                                                         
#> [233] "    title = \"Vermont Contracts by Year\","                                      
#> [234] "    caption = \"Source: {source}\","                                             
#> [235] "    fill = \"Election Year\","                                                   
#> [236] "    x = \"Year Made\","                                                          
#> [237] "    y = \"Count\""                                                               
#> [238] "  )"                                                                             
#> [239] "```"                                                                             
#> [240] ""                                                                                
#> [241] "## Wrangle"                                                                      
#> [242] ""                                                                                
#> [243] "To improve the searchability of the database, we will perform some consistent,"  
#> [244] "confident string normalization. For geographic variables like city names and"    
#> [245] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [246] "facilitate this process."                                                        
#> [247] ""                                                                                
#> [248] "### Address"                                                                     
#> [249] ""                                                                                
#> [250] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [251] "will force consistence case, remove punctuation, and abbreviate official "       
#> [252] "USPS suffixes."                                                                  
#> [253] ""                                                                                
#> [254] "```{r address_norm}"                                                             
#> [255] "vtc <- vtc %>% "                                                                 
#> [256] "  unite("                                                                        
#> [257] "    col = address_full,"                                                         
#> [258] "    starts_with(\"address\"),"                                                   
#> [259] "    sep = \" \","                                                                
#> [260] "    remove = FALSE,"                                                             
#> [261] "    na.rm = TRUE"                                                                
#> [262] "  ) %>% "                                                                        
#> [263] "  mutate("                                                                       
#> [264] "    address_norm = normal_address("                                              
#> [265] "      address = address_full,"                                                   
#> [266] "      abbs = usps_street,"                                                       
#> [267] "      na_rep = TRUE"                                                             
#> [268] "    )"                                                                           
#> [269] "  ) %>% "                                                                        
#> [270] "  select(-address_full)"                                                         
#> [271] "```"                                                                             
#> [272] ""                                                                                
#> [273] "```{r address_view}"                                                             
#> [274] "vtc %>% "                                                                        
#> [275] "  select(contains(\"address\")) %>% "                                            
#> [276] "  distinct() %>% "                                                               
#> [277] "  sample_n(10)"                                                                  
#> [278] "```"                                                                             
#> [279] ""                                                                                
#> [280] "### ZIP"                                                                         
#> [281] ""                                                                                
#> [282] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [283] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [284] "zeroes dropped by other programs like Microsoft Excel."                          
#> [285] ""                                                                                
#> [286] "```{r zip_norm}"                                                                 
#> [287] "vtc <- vtc %>% "                                                                 
#> [288] "  mutate("                                                                       
#> [289] "    zip_norm = normal_zip("                                                      
#> [290] "      zip = zip,"                                                                
#> [291] "      na_rep = TRUE"                                                             
#> [292] "    )"                                                                           
#> [293] "  )"                                                                             
#> [294] "```"                                                                             
#> [295] ""                                                                                
#> [296] "```{r zip_progress}"                                                             
#> [297] "progress_table("                                                                 
#> [298] "  vtc$zip,"                                                                      
#> [299] "  vtc$zip_norm,"                                                                 
#> [300] "  compare = valid_zip"                                                           
#> [301] ")"                                                                               
#> [302] "```"                                                                             
#> [303] ""                                                                                
#> [304] "### State"                                                                       
#> [305] ""                                                                                
#> [306] "Valid two digit state abbreviations can be made using the "                      
#> [307] "`campfin::normal_state()` function."                                             
#> [308] ""                                                                                
#> [309] "```{r state_norm}"                                                               
#> [310] "vtc <- vtc %>% "                                                                 
#> [311] "  mutate("                                                                       
#> [312] "    state_norm = normal_state("                                                  
#> [313] "      state = state,"                                                            
#> [314] "      abbreviate = TRUE,"                                                        
#> [315] "      na_rep = TRUE,"                                                            
#> [316] "      valid = valid_state"                                                       
#> [317] "    )"                                                                           
#> [318] "  )"                                                                             
#> [319] "```"                                                                             
#> [320] ""                                                                                
#> [321] "```{r state_view}"                                                               
#> [322] "vtc %>% "                                                                        
#> [323] "  filter(state != state_norm) %>% "                                              
#> [324] "  count(state, sort = TRUE)"                                                     
#> [325] "```"                                                                             
#> [326] ""                                                                                
#> [327] "```{r state_progress}"                                                           
#> [328] "progress_table("                                                                 
#> [329] "  vtc$state,"                                                                    
#> [330] "  vtc$state_norm,"                                                               
#> [331] "  compare = valid_state"                                                         
#> [332] ")"                                                                               
#> [333] "```"                                                                             
#> [334] ""                                                                                
#> [335] "### City"                                                                        
#> [336] ""                                                                                
#> [337] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [338] "the wide variety of valid cities and formats."                                   
#> [339] ""                                                                                
#> [340] "#### Normal"                                                                     
#> [341] ""                                                                                
#> [342] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [343] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [344] "`invalid_city` values."                                                          
#> [345] ""                                                                                
#> [346] "```{r city_norm}"                                                                
#> [347] "vtc <- vtc %>% "                                                                 
#> [348] "  mutate("                                                                       
#> [349] "    city_norm = normal_city("                                                    
#> [350] "      city = city, "                                                             
#> [351] "      abbs = usps_city,"                                                         
#> [352] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [353] "      na = invalid_city,"                                                        
#> [354] "      na_rep = TRUE"                                                             
#> [355] "    )"                                                                           
#> [356] "  )"                                                                             
#> [357] "```"                                                                             
#> [358] ""                                                                                
#> [359] "#### Swap"                                                                       
#> [360] ""                                                                                
#> [361] "We can further improve normalization by comparing our normalized value"          
#> [362] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [363] "If the normalized value is either an abbreviation for or very similar to the"    
#> [364] "expected value, we can confidently swap those two."                              
#> [365] ""                                                                                
#> [366] "```{r city_swap}"                                                                
#> [367] "vtc <- vtc %>% "                                                                 
#> [368] "  rename(city_raw = city) %>% "                                                  
#> [369] "  left_join("                                                                    
#> [370] "    y = zipcodes,"                                                               
#> [371] "    by = c("                                                                     
#> [372] "      \"state_norm\" = \"state\","                                               
#> [373] "      \"zip_norm\" = \"zip\""                                                    
#> [374] "    )"                                                                           
#> [375] "  ) %>% "                                                                        
#> [376] "  rename(city_match = city) %>% "                                                
#> [377] "  mutate("                                                                       
#> [378] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [379] "    match_dist = str_dist(city_norm, city_match),"                               
#> [380] "    city_swap = if_else("                                                        
#> [381] "      condition = !is.na(match_dist) & (match_abb | match_dist == 1),"           
#> [382] "      true = city_match,"                                                        
#> [383] "      false = city_norm"                                                         
#> [384] "    )"                                                                           
#> [385] "  ) %>% "                                                                        
#> [386] "  select("                                                                       
#> [387] "    -city_match,"                                                                
#> [388] "    -match_dist,"                                                                
#> [389] "    -match_abb"                                                                  
#> [390] "  )"                                                                             
#> [391] "```"                                                                             
#> [392] ""                                                                                
#> [393] "#### Refine"                                                                     
#> [394] ""                                                                                
#> [395] "The [OpenRefine] algorithms can be used to group similar strings and replace the"
#> [396] "less common versions with their most common counterpart. This can greatly "      
#> [397] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [398] "strings that have a valid city/state/zip combination."                           
#> [399] ""                                                                                
#> [400] "[or]: https://openrefine.org/"                                                   
#> [401] ""                                                                                
#> [402] "```{r city_refine}"                                                              
#> [403] "good_refine <- vtc %>% "                                                         
#> [404] "  mutate("                                                                       
#> [405] "    city_refine = city_swap %>% "                                                
#> [406] "      key_collision_merge() %>% "                                                
#> [407] "      n_gram_merge(numgram = 1)"                                                 
#> [408] "  ) %>% "                                                                        
#> [409] "  filter(city_refine != city_swap) %>% "                                         
#> [410] "  inner_join("                                                                   
#> [411] "    y = zipcodes,"                                                               
#> [412] "    by = c("                                                                     
#> [413] "      \"city_refine\" = \"city\","                                               
#> [414] "      \"state_norm\" = \"state\","                                               
#> [415] "      \"zip_norm\" = \"zip\""                                                    
#> [416] "    )"                                                                           
#> [417] "  )"                                                                             
#> [418] "```"                                                                             
#> [419] ""                                                                                
#> [420] "```{r city_count, echo=FALSE}"                                                   
#> [421] "good_refine %>%"                                                                 
#> [422] "  count("                                                                        
#> [423] "    state_norm, "                                                                
#> [424] "    zip_norm, "                                                                  
#> [425] "    city_swap, "                                                                 
#> [426] "    city_refine,"                                                                
#> [427] "    sort = TRUE"                                                                 
#> [428] "  )"                                                                             
#> [429] "```"                                                                             
#> [430] ""                                                                                
#> [431] "Then we can join the refined values back to the database."                       
#> [432] ""                                                                                
#> [433] "```{r city_join}"                                                                
#> [434] "vtc <- vtc %>% "                                                                 
#> [435] "  left_join(good_refine) %>% "                                                   
#> [436] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [437] "```"                                                                             
#> [438] ""                                                                                
#> [439] "#### Progress"                                                                   
#> [440] ""                                                                                
#> [441] "```{r city_progress, echo=FALSE}"                                                
#> [442] "many_city <- c(valid_city, extra_city)"                                          
#> [443] "progress <- progress_table("                                                     
#> [444] "  str_to_upper(vtc$city_raw),"                                                   
#> [445] "  vtc$city_norm,"                                                                
#> [446] "  vtc$city_swap,"                                                                
#> [447] "  vtc$city_refine,"                                                              
#> [448] "  compare = many_city"                                                           
#> [449] ") %>% mutate(stage = as_factor(stage))"                                          
#> [450] "kable(progress, digits = 3)"                                                     
#> [451] "```"                                                                             
#> [452] ""                                                                                
#> [453] "You can see how the percentage of valid values increased with each stage."       
#> [454] ""                                                                                
#> [455] "```{r bar_progress, echo=FALSE}"                                                 
#> [456] "raw_in <- percent(prop_in(vtc$city_raw, valid_city))"                            
#> [457] "progress %>% "                                                                   
#> [458] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [459] "  geom_hline(yintercept = 0.99) +"                                               
#> [460] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [461] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [462] "  scale_y_continuous(labels = percent) +"                                        
#> [463] "  labs("                                                                         
#> [464] "    title = \"Vermont City Normalization Progress\","                            
#> [465] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [466] "    x = \"Stage\","                                                              
#> [467] "    y = \"Percent Valid\""                                                       
#> [468] "  )"                                                                             
#> [469] "```"                                                                             
#> [470] ""                                                                                
#> [471] "More importantly, the number of distinct values decreased each stage. We were"   
#> [472] "able to confidently change many distinct invalid values to their valid"          
#> [473] "equivalent."                                                                     
#> [474] ""                                                                                
#> [475] "```{r bar_distinct, echo=FALSE}"                                                 
#> [476] "progress %>% "                                                                   
#> [477] "  select("                                                                       
#> [478] "    stage, "                                                                     
#> [479] "    all = n_distinct,"                                                           
#> [480] "    bad = n_diff"                                                                
#> [481] "  ) %>% "                                                                        
#> [482] "  mutate(good = all - bad) %>% "                                                 
#> [483] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [484] "  mutate(name = name == \"good\") %>% "                                          
#> [485] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [486] "  geom_col(aes(fill = name)) +"                                                  
#> [487] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [488] "  scale_y_continuous(labels = comma) +"                                          
#> [489] "  theme(legend.position = \"bottom\") +"                                         
#> [490] "  labs("                                                                         
#> [491] "    title = \"Vermont City Normalization Progress\","                            
#> [492] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [493] "    x = \"Stage\","                                                              
#> [494] "    y = \"Distinct Values\","                                                    
#> [495] "    fill = \"Valid\""                                                            
#> [496] "  )"                                                                             
#> [497] "```"                                                                             
#> [498] ""                                                                                
#> [499] "## Conclude"                                                                     
#> [500] ""                                                                                
#> [501] "Before exporting, we can remove the intermediary normalization columns and"      
#> [502] "rename all added variables with the `_clean` suffix."                            
#> [503] ""                                                                                
#> [504] "```{r clean_select}"                                                             
#> [505] "vtc <- vtc %>% "                                                                 
#> [506] "  select("                                                                       
#> [507] "    -city_norm,"                                                                 
#> [508] "    -city_swap,"                                                                 
#> [509] "    city_clean = city_refine"                                                    
#> [510] "  ) %>% "                                                                        
#> [511] "  rename_all(~str_replace(., \"_norm\", \"_clean\")) %>% "                       
#> [512] "  rename_all(~str_remove(., \"_raw\"))"                                          
#> [513] "```"                                                                             
#> [514] ""                                                                                
#> [515] "```{r clean_glimpse}"                                                            
#> [516] "glimpse(sample_n(vtc, 20))"                                                      
#> [517] "```"                                                                             
#> [518] ""                                                                                
#> [519] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [520] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [521] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [522] "1. There are `r comma(sum(vtc$na_flag))` records missing key variables."         
#> [523] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [524] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [525] ""                                                                                
#> [526] "## Export"                                                                       
#> [527] ""                                                                                
#> [528] "Now the file can be saved on disk for upload to the Accountability server."      
#> [529] ""                                                                                
#> [530] "```{r clean_dir}"                                                                
#> [531] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [532] "clean_path <- path(clean_dir, \"vt_contribs_clean.csv\")"                        
#> [533] "write_csv(vtc, clean_path, na = \"\")"                                           
#> [534] "file_size(clean_path)"                                                           
#> [535] "file_encoding(clean_path)"                                                       
#> [536] "```"                                                                             
#> [537] ""                                                                                
#> [538] "## Dictionary"                                                                   
#> [539] ""                                                                                
#> [540] "The following table describes the variables in our final exported file:"         
#> [541] ""                                                                                
#> [542] "```{r dict_make, echo=FALSE}"                                                    
#> [543] "dict_raw <- tibble("                                                             
#> [544] "  var = md_code(names(vtc)),"                                                    
#> [545] "  type = md_code(map_chr(vtc, typeof)),"                                         
#> [546] "  def = c("                                                                      
#> [547] "    \"Date contribution was made\","                                             
#> [548] "    \"Contribution amount or correction\","                                      
#> [549] "    \"Contributor name\","                                                       
#> [550] "    \"Contributor unique ID\","                                                  
#> [551] "    \"Contributor street address\","                                             
#> [552] "    \"Contributor secondary address\","                                          
#> [553] "    \"Contributor city name\","                                                  
#> [554] "    \"Contributor state abbreviation\","                                         
#> [555] "    \"Contributor ZIP+4 code\","                                                 
#> [556] "    \"Contributor occupation\","                                                 
#> [557] "    \"Contributor employer name\","                                              
#> [558] "    \"Contributor type\","                                                       
#> [559] "    \"Recipient committee name\","                                               
#> [560] "    \"Recipient unique ID\","                                                    
#> [561] "    \"Contribution method\","                                                    
#> [562] "    \"Recipient election or office sought\","                                    
#> [563] "    \"Flag for missing date, amount, or name\","                                 
#> [564] "    \"Flag for completely duplicated record\","                                  
#> [565] "    \"Calendar year of contribution date\","                                     
#> [566] "    \"Normalized combined street address\","                                     
#> [567] "    \"Normalized 5-digit ZIP code\","                                            
#> [568] "    \"Normalized 2-digit state abbreviation\","                                  
#> [569] "    \"Normalized city name\""                                                    
#> [570] "  )"                                                                             
#> [571] ")"                                                                               
#> [572] "```"                                                                             
#> [573] ""                                                                                
#> [574] "```{r dict_md, echo=FALSE}"                                                      
#> [575] "(dict_md <- kable("                                                              
#> [576] "  x = dict_raw,"                                                                 
#> [577] "  format = \"markdown\","                                                        
#> [578] "  col.names = c(\"Column\", \"Type\", \"Definition\")"                           
#> [579] "))"                                                                              
#> [580] "```"                                                                             
#> [581] ""                                                                                
#> [582] "```{r dict_write}"                                                               
#> [583] "write_lines("                                                                    
#> [584] "  x = c(\"# Vermont Contracts Data Dictionary\\n\", dict_md),"                   
#> [585] "  path = here(\"vt\", \"contribs\", \"vt_contribs_dict.md\"),"                   
#> [586] ")"                                                                               
#> [587] "```"                                                                             
#> [588] ""                                                                                

use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())

#> /tmp/Rtmp9Wj4wz/vt_contribs_diary.Rmd was created