Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type, author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contributions\""                                                
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r date()`\""                                                            
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 3"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "if (!interactive()) {"                                                           
#>  [31] "  options(width = 120)"                                                          
#>  [32] "  set.seed(5)"                                                                   
#>  [33] "}"                                                                               
#>  [34] "```"                                                                             
#>  [35] ""                                                                                
#>  [36] "```{r create-docs-dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [37] "doc_dir <- fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"           
#>  [38] "```"                                                                             
#>  [39] ""                                                                                
#>  [40] "## Project"                                                                      
#>  [41] ""                                                                                
#>  [42] "The Accountability Project is an effort to cut across data silos and give"       
#>  [43] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [44] "way to search across huge volumes of public data about people and organizations."
#>  [45] ""                                                                                
#>  [46] "Our goal is to standardize public data on a few key fields by thinking of each"  
#>  [47] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [48] "variables:"                                                                      
#>  [49] ""                                                                                
#>  [50] "1. All **parties** to a transaction."                                            
#>  [51] "2. The **date** of the transaction."                                             
#>  [52] "3. The **amount** of money involved."                                            
#>  [53] ""                                                                                
#>  [54] "## Objectives"                                                                   
#>  [55] ""                                                                                
#>  [56] "This document describes the process used to complete the following objectives:"  
#>  [57] ""                                                                                
#>  [58] "1. How many records are in the database?"                                        
#>  [59] "1. Check for entirely duplicated records."                                       
#>  [60] "1. Check ranges of continuous variables."                                        
#>  [61] "1. Is there anything blank or missing?"                                          
#>  [62] "1. Check for consistency issues."                                                
#>  [63] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [64] "1. Create a `year` field from the transaction date."                             
#>  [65] "1. Make sure there is data on both parties to a transaction."                    
#>  [66] ""                                                                                
#>  [67] "## Packages"                                                                     
#>  [68] ""                                                                                
#>  [69] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [70] "and communicate these results. The `pacman` package will facilitate their"       
#>  [71] "installation and attachment."                                                    
#>  [72] ""                                                                                
#>  [73] "```{r load-packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [74] "if (!require(\"pacman\")) {"                                                     
#>  [75] "  install.packages(\"pacman\")"                                                  
#>  [76] "}"                                                                               
#>  [77] "pacman::p_load("                                                                 
#>  [78] "  tidyverse, # data manipulation"                                                
#>  [79] "  lubridate, # datetime strings"                                                 
#>  [80] "  gluedown, # printing markdown"                                                 
#>  [81] "  janitor, # clean data frames"                                                  
#>  [82] "  campfin, # custom irw tools"                                                   
#>  [83] "  aws.s3, # aws cloud storage"                                                   
#>  [84] "  refinr, # cluster & merge"                                                     
#>  [85] "  scales, # format strings"                                                      
#>  [86] "  knitr, # knit documents"                                                       
#>  [87] "  vroom, # fast reading"                                                         
#>  [88] "  rvest, # scrape html"                                                          
#>  [89] "  glue, # code strings"                                                          
#>  [90] "  here, # project paths"                                                         
#>  [91] "  httr, # http requests"                                                         
#>  [92] "  fs # local storage "                                                           
#>  [93] ")"                                                                               
#>  [94] "```"                                                                             
#>  [95] ""                                                                                
#>  [96] "This document should be run as part of the `R_campfin` project, which lives as a"
#>  [97] "sub-directory of the more general, language-agnostic"                            
#>  [98] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#>  [99] ""                                                                                
#> [100] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#> [101] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#> [102] "paths relative to _your_ machine."                                               
#> [103] ""                                                                                
#> [104] "```{r where-here}"                                                               
#> [105] "# where does this document knit?"                                                
#> [106] "here::i_am(\"vt/contribs/docs/vt_contribs_diary.Rmd\")"                          
#> [107] "```"                                                                             
#> [108] ""                                                                                
#> [109] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [110] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [111] ""                                                                                
#> [112] "## Data"                                                                         
#> [113] ""                                                                                
#> [114] ""                                                                                
#> [115] ""                                                                                
#> [116] "## Download"                                                                     
#> [117] ""                                                                                
#> [118] "```{r raw-dir}"                                                                  
#> [119] "raw_url <- \"https://example.com/source_file.csv\""                              
#> [120] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [121] "raw_path <- path(raw_dir, basename(raw_url))"                                    
#> [122] "```"                                                                             
#> [123] ""                                                                                
#> [124] "```{r raw-download}"                                                             
#> [125] "if (!file_exists(raw_path)) {"                                                   
#> [126] "  download.file(raw_url, raw_path)"                                              
#> [127] "}"                                                                               
#> [128] "```"                                                                             
#> [129] ""                                                                                
#> [130] "## Read"                                                                         
#> [131] ""                                                                                
#> [132] "```{r raw-read}"                                                                 
#> [133] "vtc <- read_delim("                                                              
#> [134] "  file = raw_path,"                                                              
#> [135] "  delim = \",\","                                                                
#> [136] "  escape_backslash = FALSE,"                                                     
#> [137] "  escape_double = FALSE,"                                                        
#> [138] "  col_types = cols("                                                             
#> [139] "    .default = col_character(),"                                                 
#> [140] "    date = col_date_usa(),"                                                      
#> [141] "    amount = col_double()"                                                       
#> [142] "  )"                                                                             
#> [143] ")"                                                                               
#> [144] "```"                                                                             
#> [145] ""                                                                                
#> [146] "```{r clean-names}"                                                              
#> [147] "vtc <- clean_names(vtc, case = \"snake\")"                                       
#> [148] "```"                                                                             
#> [149] ""                                                                                
#> [150] "## Explore"                                                                      
#> [151] ""                                                                                
#> [152] "There are `r comma(nrow(vtc))` rows of `r ncol(vtc)` columns. Each record"       
#> [153] "represents a single Contributions..."                                            
#> [154] ""                                                                                
#> [155] "```{r glimpse}"                                                                  
#> [156] "glimpse(vtc)"                                                                    
#> [157] "tail(vtc)"                                                                       
#> [158] "```"                                                                             
#> [159] ""                                                                                
#> [160] "### Missing"                                                                     
#> [161] ""                                                                                
#> [162] "Columns vary in their degree of missing values."                                 
#> [163] ""                                                                                
#> [164] "```{r na-count}"                                                                 
#> [165] "col_stats(vtc, count_na)"                                                        
#> [166] "```"                                                                             
#> [167] ""                                                                                
#> [168] "We can flag any record missing a key variable needed to identify a transaction." 
#> [169] ""                                                                                
#> [170] "```{r na-flag}"                                                                  
#> [171] "key_vars <- c(\"date\", \"last_name\", \"amount\", \"committee_name\")"          
#> [172] "vtc <- flag_na(vtc, all_of(key_vars))"                                           
#> [173] "sum(vtc$na_flag)"                                                                
#> [174] "```"                                                                             
#> [175] ""                                                                                
#> [176] "```{r na-view}"                                                                  
#> [177] "vtc %>% "                                                                        
#> [178] "  filter(na_flag) %>% "                                                          
#> [179] "  select(all_of(key_vars))"                                                      
#> [180] "```"                                                                             
#> [181] ""                                                                                
#> [182] "### Duplicates"                                                                  
#> [183] ""                                                                                
#> [184] "We can also flag any record completely duplicated across every column."          
#> [185] ""                                                                                
#> [186] "```{r dupe-flag}"                                                                
#> [187] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [188] "sum(vtc$dupe_flag)"                                                              
#> [189] "```"                                                                             
#> [190] ""                                                                                
#> [191] "```{r dupe-view}"                                                                
#> [192] "vtc %>% "                                                                        
#> [193] "  filter(dupe_flag) %>% "                                                        
#> [194] "  select(all_of(key_vars))"                                                      
#> [195] "```"                                                                             
#> [196] ""                                                                                
#> [197] "### Categorical"                                                                 
#> [198] ""                                                                                
#> [199] "```{r distinct-count}"                                                           
#> [200] "col_stats(vtc, n_distinct)"                                                      
#> [201] "```"                                                                             
#> [202] ""                                                                                
#> [203] "```{r distinct-plots, echo=FALSE}"                                               
#> [204] "explore_plot(vtc, type)"                                                         
#> [205] "```"                                                                             
#> [206] ""                                                                                
#> [207] "### Amounts"                                                                     
#> [208] ""                                                                                
#> [209] "```{r amount-round}"                                                             
#> [210] "vtc$amount <- round(vtc$amount, digits = 2)"                                     
#> [211] "```"                                                                             
#> [212] ""                                                                                
#> [213] "```{r amount-summary}"                                                           
#> [214] "summary(vtc$amount)"                                                             
#> [215] "mean(vtc$amount <= 0)"                                                           
#> [216] "```"                                                                             
#> [217] ""                                                                                
#> [218] "These are the records with the minimum and maximum amounts."                     
#> [219] ""                                                                                
#> [220] "```{r amount-minmax}"                                                            
#> [221] "glimpse(vtc[c(which.max(vtc$amount), which.min(vtc$amount)), ])"                 
#> [222] "```"                                                                             
#> [223] ""                                                                                
#> [224] "```{r hist-amount, echo=FALSE}"                                                  
#> [225] "vtc %>%"                                                                         
#> [226] "  ggplot(aes(amount)) +"                                                         
#> [227] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [228] "  scale_y_continuous(labels = comma) +"                                          
#> [229] "  scale_x_continuous("                                                           
#> [230] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [231] "    labels = dollar,"                                                            
#> [232] "    trans = \"log10\""                                                           
#> [233] "  ) +"                                                                           
#> [234] "  labs("                                                                         
#> [235] "    title = \"Vermont Contributions Amount Distribution\","                      
#> [236] "    caption = \"Source: {source}\","                                             
#> [237] "    x = \"Amount\","                                                             
#> [238] "    y = \"Count\""                                                               
#> [239] "  )"                                                                             
#> [240] "```"                                                                             
#> [241] ""                                                                                
#> [242] "### Dates"                                                                       
#> [243] ""                                                                                
#> [244] "We can add the calendar year from `date` with `lubridate::year()`"               
#> [245] ""                                                                                
#> [246] "```{r date-year}"                                                                
#> [247] "vtc <- mutate(vtc, year = year(date))"                                           
#> [248] "```"                                                                             
#> [249] ""                                                                                
#> [250] "```{r date-range}"                                                               
#> [251] "min(vtc$date)"                                                                   
#> [252] "sum(vtc$year < 2000)"                                                            
#> [253] "max(vtc$date)"                                                                   
#> [254] "sum(vtc$date > today())"                                                         
#> [255] "```"                                                                             
#> [256] ""                                                                                
#> [257] "```{r bar-year, echo=FALSE}"                                                     
#> [258] "vtc %>% "                                                                        
#> [259] "  count(year) %>% "                                                              
#> [260] "  mutate(even = is_even(year)) %>% "                                             
#> [261] "  ggplot(aes(x = year, y = n)) +"                                                
#> [262] "  geom_col(aes(fill = even)) + "                                                 
#> [263] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [264] "  scale_y_continuous(labels = comma) +"                                          
#> [265] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [266] "  theme(legend.position = \"bottom\") +"                                         
#> [267] "  labs("                                                                         
#> [268] "    title = \"Vermont Contributions by Year\","                                  
#> [269] "    caption = \"Source: {source}\","                                             
#> [270] "    fill = \"Election Year\","                                                   
#> [271] "    x = \"Year Made\","                                                          
#> [272] "    y = \"Count\""                                                               
#> [273] "  )"                                                                             
#> [274] "```"                                                                             
#> [275] ""                                                                                
#> [276] "## Wrangle"                                                                      
#> [277] ""                                                                                
#> [278] "To improve the searchability of the database, we will perform some consistent,"  
#> [279] "confident string normalization. For geographic variables like city names and"    
#> [280] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [281] "facilitate this process."                                                        
#> [282] ""                                                                                
#> [283] "### Address"                                                                     
#> [284] ""                                                                                
#> [285] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [286] "will force consistence case, remove punctuation, and abbreviate official "       
#> [287] "USPS suffixes."                                                                  
#> [288] ""                                                                                
#> [289] "```{r address-norm}"                                                             
#> [290] "addr_norm <- vtc %>% "                                                           
#> [291] "  distinct(address1, address2) %>% "                                             
#> [292] "  unite("                                                                        
#> [293] "    col = address_full,"                                                         
#> [294] "    starts_with(\"address\"),"                                                   
#> [295] "    sep = \" \","                                                                
#> [296] "    remove = FALSE,"                                                             
#> [297] "    na.rm = TRUE"                                                                
#> [298] "  ) %>% "                                                                        
#> [299] "  mutate("                                                                       
#> [300] "    address_norm = normal_address("                                              
#> [301] "      address = address_full,"                                                   
#> [302] "      abbs = usps_street,"                                                       
#> [303] "      na_rep = TRUE"                                                             
#> [304] "    )"                                                                           
#> [305] "  ) %>% "                                                                        
#> [306] "  select(-address_full)"                                                         
#> [307] "```"                                                                             
#> [308] ""                                                                                
#> [309] "```{r address-view}"                                                             
#> [310] "addr_norm"                                                                       
#> [311] "```"                                                                             
#> [312] ""                                                                                
#> [313] "```{r address-join}"                                                             
#> [314] "vtc <- left_join(vtc, addr_norm, by = c(\"address1\", \"address2\"))"            
#> [315] "```"                                                                             
#> [316] ""                                                                                
#> [317] "### ZIP"                                                                         
#> [318] ""                                                                                
#> [319] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [320] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [321] "zeroes dropped by other programs like Microsoft Excel."                          
#> [322] ""                                                                                
#> [323] "```{r zip-norm}"                                                                 
#> [324] "vtc <- vtc %>% "                                                                 
#> [325] "  mutate("                                                                       
#> [326] "    zip_norm = normal_zip("                                                      
#> [327] "      zip = zip,"                                                                
#> [328] "      na_rep = TRUE"                                                             
#> [329] "    )"                                                                           
#> [330] "  )"                                                                             
#> [331] "```"                                                                             
#> [332] ""                                                                                
#> [333] "```{r zip-progress}"                                                             
#> [334] "progress_table("                                                                 
#> [335] "  vtc$zip,"                                                                      
#> [336] "  vtc$zip_norm,"                                                                 
#> [337] "  compare = valid_zip"                                                           
#> [338] ")"                                                                               
#> [339] "```"                                                                             
#> [340] ""                                                                                
#> [341] "### State"                                                                       
#> [342] ""                                                                                
#> [343] "Valid two digit state abbreviations can be made using the "                      
#> [344] "`campfin::normal_state()` function."                                             
#> [345] ""                                                                                
#> [346] "```{r state-norm}"                                                               
#> [347] "vtc <- vtc %>% "                                                                 
#> [348] "  mutate("                                                                       
#> [349] "    state_norm = normal_state("                                                  
#> [350] "      state = state,"                                                            
#> [351] "      abbreviate = TRUE,"                                                        
#> [352] "      na_rep = TRUE,"                                                            
#> [353] "      valid = valid_state"                                                       
#> [354] "    )"                                                                           
#> [355] "  )"                                                                             
#> [356] "```"                                                                             
#> [357] ""                                                                                
#> [358] "```{r state-view}"                                                               
#> [359] "vtc %>% "                                                                        
#> [360] "  filter(state != state_norm) %>% "                                              
#> [361] "  count(state, state_norm, sort = TRUE)"                                         
#> [362] "```"                                                                             
#> [363] ""                                                                                
#> [364] "```{r state-progress}"                                                           
#> [365] "progress_table("                                                                 
#> [366] "  vtc$state,"                                                                    
#> [367] "  vtc$state_norm,"                                                               
#> [368] "  compare = valid_state"                                                         
#> [369] ")"                                                                               
#> [370] "```"                                                                             
#> [371] ""                                                                                
#> [372] "### City"                                                                        
#> [373] ""                                                                                
#> [374] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [375] "the wide variety of valid cities and formats."                                   
#> [376] ""                                                                                
#> [377] "#### Normal"                                                                     
#> [378] ""                                                                                
#> [379] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [380] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [381] "`invalid_city` values."                                                          
#> [382] ""                                                                                
#> [383] "```{r city-norm}"                                                                
#> [384] "norm_city <- vtc %>% "                                                           
#> [385] "  distinct(city, state_norm, zip_norm) %>% "                                     
#> [386] "  mutate("                                                                       
#> [387] "    city_norm = normal_city("                                                    
#> [388] "      city = city, "                                                             
#> [389] "      abbs = usps_city,"                                                         
#> [390] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [391] "      na = invalid_city,"                                                        
#> [392] "      na_rep = TRUE"                                                             
#> [393] "    )"                                                                           
#> [394] "  )"                                                                             
#> [395] "```"                                                                             
#> [396] ""                                                                                
#> [397] "#### Swap"                                                                       
#> [398] ""                                                                                
#> [399] "We can further improve normalization by comparing our normalized value"          
#> [400] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [401] "If the normalized value is either an abbreviation for or very similar to the"    
#> [402] "expected value, we can confidently swap those two."                              
#> [403] ""                                                                                
#> [404] "```{r city-swap}"                                                                
#> [405] "norm_city <- norm_city %>% "                                                     
#> [406] "  rename(city_raw = city) %>% "                                                  
#> [407] "  left_join("                                                                    
#> [408] "    y = zipcodes,"                                                               
#> [409] "    by = c("                                                                     
#> [410] "      \"state_norm\" = \"state\","                                               
#> [411] "      \"zip_norm\" = \"zip\""                                                    
#> [412] "    )"                                                                           
#> [413] "  ) %>% "                                                                        
#> [414] "  rename(city_match = city) %>% "                                                
#> [415] "  mutate("                                                                       
#> [416] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [417] "    match_dist = str_dist(city_norm, city_match),"                               
#> [418] "    city_swap = if_else("                                                        
#> [419] "      condition = !is.na(match_dist) & (match_abb | match_dist == 1),"           
#> [420] "      true = city_match,"                                                        
#> [421] "      false = city_norm"                                                         
#> [422] "    )"                                                                           
#> [423] "  ) %>% "                                                                        
#> [424] "  select("                                                                       
#> [425] "    -city_match,"                                                                
#> [426] "    -match_dist,"                                                                
#> [427] "    -match_abb"                                                                  
#> [428] "  )"                                                                             
#> [429] "```"                                                                             
#> [430] ""                                                                                
#> [431] "```{r city-rejoin}"                                                              
#> [432] "vtc <- left_join("                                                               
#> [433] "  x = vtc,"                                                                      
#> [434] "  y = norm_city,"                                                                
#> [435] "  by = c("                                                                       
#> [436] "    \"city\" = \"city_raw\", "                                                   
#> [437] "    \"state_norm\", "                                                            
#> [438] "    \"zip_norm\""                                                                
#> [439] "  )"                                                                             
#> [440] ")"                                                                               
#> [441] "```"                                                                             
#> [442] ""                                                                                
#> [443] "#### Refine"                                                                     
#> [444] ""                                                                                
#> [445] "The [OpenRefine][or] algorithms can be used to group similar strings and replace"
#> [446] "the less common versions with their most common counterpart. This can greatly"   
#> [447] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [448] "strings that have a valid city/state/zip combination."                           
#> [449] ""                                                                                
#> [450] "[or]: https://openrefine.org/"                                                   
#> [451] ""                                                                                
#> [452] "```{r city-refine}"                                                              
#> [453] "good_refine <- vtc %>% "                                                         
#> [454] "  mutate("                                                                       
#> [455] "    city_refine = city_swap %>% "                                                
#> [456] "      key_collision_merge() %>% "                                                
#> [457] "      n_gram_merge(numgram = 1)"                                                 
#> [458] "  ) %>% "                                                                        
#> [459] "  filter(city_refine != city_swap) %>% "                                         
#> [460] "  inner_join("                                                                   
#> [461] "    y = zipcodes,"                                                               
#> [462] "    by = c("                                                                     
#> [463] "      \"city_refine\" = \"city\","                                               
#> [464] "      \"state_norm\" = \"state\","                                               
#> [465] "      \"zip_norm\" = \"zip\""                                                    
#> [466] "    )"                                                                           
#> [467] "  )"                                                                             
#> [468] "```"                                                                             
#> [469] ""                                                                                
#> [470] "```{r city-count, echo=FALSE}"                                                   
#> [471] "good_refine %>%"                                                                 
#> [472] "  count("                                                                        
#> [473] "    state_norm, "                                                                
#> [474] "    zip_norm, "                                                                  
#> [475] "    city_swap, "                                                                 
#> [476] "    city_refine,"                                                                
#> [477] "    sort = TRUE"                                                                 
#> [478] "  )"                                                                             
#> [479] "```"                                                                             
#> [480] ""                                                                                
#> [481] "Then we can join the refined values back to the database."                       
#> [482] ""                                                                                
#> [483] "```{r city-join}"                                                                
#> [484] "vtc <- vtc %>% "                                                                 
#> [485] "  left_join(good_refine) %>% "                                                   
#> [486] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [487] "```"                                                                             
#> [488] ""                                                                                
#> [489] "#### Progress"                                                                   
#> [490] ""                                                                                
#> [491] "Our goal for normalization was to increase the proportion of city values known"  
#> [492] "to be valid and reduce the total distinct values by correcting misspellings."    
#> [493] ""                                                                                
#> [494] "```{r city-progress, echo=FALSE}"                                                
#> [495] "many_city <- c(valid_city, extra_city)"                                          
#> [496] "progress <- progress_table("                                                     
#> [497] "  str_to_upper(vtc$city),"                                                       
#> [498] "  vtc$city_norm,"                                                                
#> [499] "  vtc$city_swap,"                                                                
#> [500] "  vtc$city_refine,"                                                              
#> [501] "  compare = many_city"                                                           
#> [502] ") %>% mutate(stage = as_factor(stage))"                                          
#> [503] "progress %>% "                                                                   
#> [504] "    mutate(across(stage, md_code)) %>% "                                         
#> [505] "    kable(digits = 3)"                                                           
#> [506] "```"                                                                             
#> [507] ""                                                                                
#> [508] "You can see how the percentage of valid values increased with each stage."       
#> [509] ""                                                                                
#> [510] "```{r bar-progress, echo=FALSE}"                                                 
#> [511] "raw_in <- percent(prop_in(vtc$city, valid_city))"                                
#> [512] "progress %>% "                                                                   
#> [513] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [514] "  geom_hline(yintercept = 0.99) +"                                               
#> [515] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [516] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [517] "  scale_y_continuous(labels = percent) +"                                        
#> [518] "  labs("                                                                         
#> [519] "    title = \"Vermont City Normalization Progress\","                            
#> [520] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [521] "    x = \"Stage\","                                                              
#> [522] "    y = \"Percent Valid\""                                                       
#> [523] "  )"                                                                             
#> [524] "```"                                                                             
#> [525] ""                                                                                
#> [526] "More importantly, the number of distinct values decreased each stage. We were"   
#> [527] "able to confidently change many distinct invalid values to their valid"          
#> [528] "equivalent."                                                                     
#> [529] ""                                                                                
#> [530] "```{r bar-distinct, echo=FALSE}"                                                 
#> [531] "progress %>% "                                                                   
#> [532] "  select("                                                                       
#> [533] "    stage, "                                                                     
#> [534] "    all = n_distinct,"                                                           
#> [535] "    bad = n_diff"                                                                
#> [536] "  ) %>% "                                                                        
#> [537] "  mutate(good = all - bad) %>% "                                                 
#> [538] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [539] "  mutate(name = name == \"good\") %>% "                                          
#> [540] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [541] "  geom_col(aes(fill = name)) +"                                                  
#> [542] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [543] "  scale_y_continuous(labels = comma) +"                                          
#> [544] "  theme(legend.position = \"bottom\") +"                                         
#> [545] "  labs("                                                                         
#> [546] "    title = \"Vermont City Normalization Progress\","                            
#> [547] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [548] "    x = \"Stage\","                                                              
#> [549] "    y = \"Distinct Values\","                                                    
#> [550] "    fill = \"Valid\""                                                            
#> [551] "  )"                                                                             
#> [552] "```"                                                                             
#> [553] ""                                                                                
#> [554] "Before exporting, we can remove the intermediary normalization columns and"      
#> [555] "rename all added variables with the `_clean` suffix."                            
#> [556] ""                                                                                
#> [557] "```{r clean-select}"                                                             
#> [558] "vtc <- vtc %>% "                                                                 
#> [559] "  select("                                                                       
#> [560] "    -city_norm,"                                                                 
#> [561] "    -city_swap,"                                                                 
#> [562] "    city_clean = city_refine"                                                    
#> [563] "  ) %>% "                                                                        
#> [564] "  rename_all(~str_replace(., \"_norm\", \"_clean\")) %>% "                       
#> [565] "  rename_all(~str_remove(., \"_raw\")) %>% "                                     
#> [566] "  relocate(address_clean, city_clean, state_clean, .before = zip_clean)"         
#> [567] "```"                                                                             
#> [568] ""                                                                                
#> [569] "## Conclude"                                                                     
#> [570] ""                                                                                
#> [571] "```{r clean-glimpse}"                                                            
#> [572] "glimpse(sample_n(vtc, 50))"                                                      
#> [573] "```"                                                                             
#> [574] ""                                                                                
#> [575] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [576] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [577] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [578] "1. There are `r comma(sum(vtc$na_flag))` records missing key variables."         
#> [579] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [580] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [581] ""                                                                                
#> [582] "## Export"                                                                       
#> [583] ""                                                                                
#> [584] "Now the file can be saved on disk for upload to the Accountability server."      
#> [585] ""                                                                                
#> [586] "```{r clean-dir}"                                                                
#> [587] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [588] "clean_path <- path(clean_dir, \"vt_contribs_clean.csv\")"                        
#> [589] "write_csv(vtc, clean_path, na = \"\")"                                           
#> [590] "(clean_size <- file_size(clean_path))"                                           
#> [591] "```"                                                                             
#> [592] ""                                                                                
#> [593] "## Upload"                                                                       
#> [594] ""                                                                                
#> [595] "We can use the `aws.s3::put_object()` to upload the text file to the IRW server."
#> [596] ""                                                                                
#> [597] "```{r aws-upload, eval=FALSE}"                                                   
#> [598] "aws_path <- path(\"csv\", basename(clean_path))"                                 
#> [599] "if (!object_exists(aws_path, \"publicaccountability\")) {"                       
#> [600] "  put_object("                                                                   
#> [601] "    file = clean_path,"                                                          
#> [602] "    object = aws_path, "                                                         
#> [603] "    bucket = \"publicaccountability\","                                          
#> [604] "    acl = \"public-read\","                                                      
#> [605] "    show_progress = TRUE,"                                                       
#> [606] "    multipart = TRUE"                                                            
#> [607] "  )"                                                                             
#> [608] "}"                                                                               
#> [609] "aws_head <- head_object(aws_path, \"publicaccountability\")"                     
#> [610] "(aws_size <- as_fs_bytes(attr(aws_head, \"content-length\")))"                   
#> [611] "unname(aws_size == clean_size)"                                                  
#> [612] "```"                                                                             
#> [613] ""                                                                                
#> [614] "## Dictionary"                                                                   
#> [615] ""                                                                                
#> [616] "The following table describes the variables in our final exported file:"         
#> [617] ""                                                                                
#> [618] "```{r dict-make, echo=FALSE}"                                                    
#> [619] "dict_raw <- tibble("                                                             
#> [620] "  var = md_code(names(vtc)),"                                                    
#> [621] "  type = md_code(map_chr(vtc, typeof)),"                                         
#> [622] "  def = c("                                                                      
#> [623] "    \"\""                                                                        
#> [624] "  )"                                                                             
#> [625] ")"                                                                               
#> [626] "```"                                                                             
#> [627] ""                                                                                
#> [628] "```{r dict-md, echo=FALSE}"                                                      
#> [629] "(dict_md <- kable("                                                              
#> [630] "  x = dict_raw,"                                                                 
#> [631] "  format = \"markdown\","                                                        
#> [632] "  col.names = c(\"Column\", \"Type\", \"Definition\")"                           
#> [633] "))"                                                                              
#> [634] "```"                                                                             

use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpFtsfUm/vt_contribs_diary.Rmd was created