Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type, author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contracts\""                                                    
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r Sys.time()`\""                                                        
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 2"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "if (!interactive()) {"                                                           
#>  [31] "  options(width = 99)"                                                           
#>  [32] "  set.seed(5)"                                                                   
#>  [33] "}"                                                                               
#>  [34] "```"                                                                             
#>  [35] ""                                                                                
#>  [36] "```{r create_docs_dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [37] "doc_dir <- fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"           
#>  [38] "```"                                                                             
#>  [39] ""                                                                                
#>  [40] "## Project"                                                                      
#>  [41] ""                                                                                
#>  [42] "The Accountability Project is an effort to cut across data silos and give"       
#>  [43] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [44] "way to search across huge volumes of public data about people and organizations."
#>  [45] ""                                                                                
#>  [46] "Our goal is to standardizing public data on a few key fields by thinking of each"
#>  [47] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [48] "variables:"                                                                      
#>  [49] ""                                                                                
#>  [50] "1. All **parties** to a transaction."                                            
#>  [51] "2. The **date** of the transaction."                                             
#>  [52] "3. The **amount** of money involved."                                            
#>  [53] ""                                                                                
#>  [54] "## Objectives"                                                                   
#>  [55] ""                                                                                
#>  [56] "This document describes the process used to complete the following objectives:"  
#>  [57] ""                                                                                
#>  [58] "1. How many records are in the database?"                                        
#>  [59] "1. Check for entirely duplicated records."                                       
#>  [60] "1. Check ranges of continuous variables."                                        
#>  [61] "1. Is there anything blank or missing?"                                          
#>  [62] "1. Check for consistency issues."                                                
#>  [63] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [64] "1. Create a `year` field from the transaction date."                             
#>  [65] "1. Make sure there is data on both parties to a transaction."                    
#>  [66] ""                                                                                
#>  [67] "## Packages"                                                                     
#>  [68] ""                                                                                
#>  [69] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [70] "and communicate these results. The `pacman` package will facilitate their"       
#>  [71] "installation and attachment."                                                    
#>  [72] ""                                                                                
#>  [73] "The IRW's `campfin` package will also have to be installed from GitHub. This"    
#>  [74] "package contains functions custom made to help facilitate the processing of"     
#>  [75] "campaign finance data."                                                          
#>  [76] ""                                                                                
#>  [77] "```{r load_packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [78] "if (!require(\"pacman\")) install.packages(\"pacman\")"                          
#>  [79] "pacman::p_load_gh(\"irworkshop/campfin\")"                                       
#>  [80] "pacman::p_load("                                                                 
#>  [81] "  tidyverse, # data manipulation"                                                
#>  [82] "  lubridate, # datetime strings"                                                 
#>  [83] "  gluedown, # printing markdown"                                                 
#>  [84] "  magrittr, # pipe operators"                                                    
#>  [85] "  janitor, # clean data frames"                                                  
#>  [86] "  refinr, # cluster and merge"                                                   
#>  [87] "  scales, # format strings"                                                      
#>  [88] "  knitr, # knit documents"                                                       
#>  [89] "  vroom, # read files fast"                                                      
#>  [90] "  rvest, # html scraping"                                                        
#>  [91] "  glue, # combine strings"                                                       
#>  [92] "  here, # relative paths"                                                        
#>  [93] "  httr, # http requests"                                                         
#>  [94] "  fs # local storage "                                                           
#>  [95] ")"                                                                               
#>  [96] "```"                                                                             
#>  [97] ""                                                                                
#>  [98] "This document should be run as part of the `R_campfin` project, which lives as a"
#>  [99] "sub-directory of the more general, language-agnostic"                            
#> [100] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#> [101] ""                                                                                
#> [102] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#> [103] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#> [104] "paths relative to _your_ machine."                                               
#> [105] ""                                                                                
#> [106] "```{r where_here}"                                                               
#> [107] "# where does this document knit?"                                                
#> [108] "here::here()"                                                                    
#> [109] "```"                                                                             
#> [110] ""                                                                                
#> [111] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [112] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [113] ""                                                                                
#> [114] "## Data"                                                                         
#> [115] ""                                                                                
#> [116] "## Download"                                                                     
#> [117] ""                                                                                
#> [118] "```{r raw_dir}"                                                                  
#> [119] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [120] "```"                                                                             
#> [121] ""                                                                                
#> [122] "```{r raw_download}"                                                             
#> [123] "download.file()"                                                                 
#> [124] "```"                                                                             
#> [125] ""                                                                                
#> [126] "## Read"                                                                         
#> [127] ""                                                                                
#> [128] "```{r raw_read}"                                                                 
#> [129] "vtc <- read_delim("                                                              
#> [130] "  file = dir_ls(raw_dir)"                                                        
#> [131] ")"                                                                               
#> [132] "```"                                                                             
#> [133] ""                                                                                
#> [134] "## Explore"                                                                      
#> [135] ""                                                                                
#> [136] "```{r glimpse}"                                                                  
#> [137] "glimpse(vtc)"                                                                    
#> [138] "tail(vtc)"                                                                       
#> [139] "```"                                                                             
#> [140] ""                                                                                
#> [141] "### Missing"                                                                     
#> [142] ""                                                                                
#> [143] "```{r na_count}"                                                                 
#> [144] "col_stats(vtc, count_na)"                                                        
#> [145] "```"                                                                             
#> [146] ""                                                                                
#> [147] "```{r na_flag}"                                                                  
#> [148] "vtc <- vtc %>% flag_na(date, last, amount, committee)"                           
#> [149] "sum(vtc$na_flag)"                                                                
#> [150] "```"                                                                             
#> [151] ""                                                                                
#> [152] "```{r na_view}"                                                                  
#> [153] "vtc %>% "                                                                        
#> [154] "  filter(na_flag) %>% "                                                          
#> [155] "  select(date, last, amount, committee)"                                         
#> [156] "```"                                                                             
#> [157] ""                                                                                
#> [158] "### Duplicates"                                                                  
#> [159] ""                                                                                
#> [160] "```{r dupe_flag}"                                                                
#> [161] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [162] "sum(vtc$dupe_flag)"                                                              
#> [163] "```"                                                                             
#> [164] ""                                                                                
#> [165] "```{r dupe_view}"                                                                
#> [166] "vtc %>% "                                                                        
#> [167] "  filter(dupe_flag) %>% "                                                        
#> [168] "  select(date, last, amount, committee)"                                         
#> [169] "```"                                                                             
#> [170] ""                                                                                
#> [171] "### Categorical"                                                                 
#> [172] ""                                                                                
#> [173] "```{r distinct_count}"                                                           
#> [174] "col_stats(vtc, n_distinct)"                                                      
#> [175] "```"                                                                             
#> [176] ""                                                                                
#> [177] "```{r distinct_plots}"                                                           
#> [178] "explore_plot(vtc, type)"                                                         
#> [179] "```"                                                                             
#> [180] ""                                                                                
#> [181] "### Amounts"                                                                     
#> [182] ""                                                                                
#> [183] "```{r ammount_summary}"                                                          
#> [184] "summary(vtc$amount)"                                                             
#> [185] "mean(vtc$amount <= 0)"                                                           
#> [186] "```"                                                                             
#> [187] ""                                                                                
#> [188] "```{r hist_amount, echo=FALSE}"                                                  
#> [189] "vtc %>%"                                                                         
#> [190] "  ggplot(aes(amount)) +"                                                         
#> [191] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [192] "  scale_y_continuous(labels = comma) +"                                          
#> [193] "  scale_x_continuous("                                                           
#> [194] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [195] "    labels = dollar,"                                                            
#> [196] "    trans = \"log10\""                                                           
#> [197] "  ) +"                                                                           
#> [198] "  labs("                                                                         
#> [199] "    title = \"Vermont Contracts Amount Distribution\","                          
#> [200] "    caption = \"Source: {source}\","                                             
#> [201] "    x = \"Amount\","                                                             
#> [202] "    y = \"Count\""                                                               
#> [203] "  )"                                                                             
#> [204] "```"                                                                             
#> [205] ""                                                                                
#> [206] "### Dates"                                                                       
#> [207] ""                                                                                
#> [208] "We can add the calendar year from `date` with `lubridate::year()`"               
#> [209] ""                                                                                
#> [210] "```{r date_year}"                                                                
#> [211] "vtc <- mutate(vtc, year = year(date))"                                           
#> [212] "```"                                                                             
#> [213] ""                                                                                
#> [214] "```{r date_range}"                                                               
#> [215] "min(vtc$date)"                                                                   
#> [216] "sum(vtc$year < 2000)"                                                            
#> [217] "max(vtc$date)"                                                                   
#> [218] "sum(vtc$date > today())"                                                         
#> [219] "```"                                                                             
#> [220] ""                                                                                
#> [221] "```{r bar_year, echo=FALSE}"                                                     
#> [222] "vtc %>% "                                                                        
#> [223] "  count(year) %>% "                                                              
#> [224] "  mutate(even = is_even(year)) %>% "                                             
#> [225] "  ggplot(aes(x = year, y = n)) +"                                                
#> [226] "  geom_col(aes(fill = even)) + "                                                 
#> [227] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [228] "  scale_y_continuous(labels = comma) +"                                          
#> [229] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [230] "  theme(legend.position = \"bottom\") +"                                         
#> [231] "  labs("                                                                         
#> [232] "    title = \"Vermont Contracts by Year\","                                      
#> [233] "    caption = \"Source: {source}\","                                             
#> [234] "    fill = \"Election Year\","                                                   
#> [235] "    x = \"Year Made\","                                                          
#> [236] "    y = \"Count\""                                                               
#> [237] "  )"                                                                             
#> [238] "```"                                                                             
#> [239] ""                                                                                
#> [240] "## Wrangle"                                                                      
#> [241] ""                                                                                
#> [242] "To improve the searchability of the database, we will perform some consistent,"  
#> [243] "confident string normalization. For geographic variables like city names and"    
#> [244] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [245] "facilitate this process."                                                        
#> [246] ""                                                                                
#> [247] "### Address"                                                                     
#> [248] ""                                                                                
#> [249] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [250] "will force consistence case, remove punctuation, and abbreviate official "       
#> [251] "USPS suffixes."                                                                  
#> [252] ""                                                                                
#> [253] "```{r address_norm}"                                                             
#> [254] "vtc <- vtc %>% "                                                                 
#> [255] "  unite("                                                                        
#> [256] "    col = address_full,"                                                         
#> [257] "    starts_with(\"address\"),"                                                   
#> [258] "    sep = \" \","                                                                
#> [259] "    remove = FALSE,"                                                             
#> [260] "    na.rm = TRUE"                                                                
#> [261] "  ) %>% "                                                                        
#> [262] "  mutate("                                                                       
#> [263] "    address_norm = normal_address("                                              
#> [264] "      address = address_full,"                                                   
#> [265] "      abbs = usps_street,"                                                       
#> [266] "      na_rep = TRUE"                                                             
#> [267] "    )"                                                                           
#> [268] "  ) %>% "                                                                        
#> [269] "  select(-address_full)"                                                         
#> [270] "```"                                                                             
#> [271] ""                                                                                
#> [272] "```{r address_view}"                                                             
#> [273] "vtc %>% "                                                                        
#> [274] "  select(contains(\"address\")) %>% "                                            
#> [275] "  distinct() %>% "                                                               
#> [276] "  sample_n(10)"                                                                  
#> [277] "```"                                                                             
#> [278] ""                                                                                
#> [279] "### ZIP"                                                                         
#> [280] ""                                                                                
#> [281] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [282] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [283] "zeroes dropped by other programs like Microsoft Excel."                          
#> [284] ""                                                                                
#> [285] "```{r zip_norm}"                                                                 
#> [286] "vtc <- vtc %>% "                                                                 
#> [287] "  mutate("                                                                       
#> [288] "    zip_norm = normal_zip("                                                      
#> [289] "      zip = zip,"                                                                
#> [290] "      na_rep = TRUE"                                                             
#> [291] "    )"                                                                           
#> [292] "  )"                                                                             
#> [293] "```"                                                                             
#> [294] ""                                                                                
#> [295] "```{r zip_progress}"                                                             
#> [296] "progress_table("                                                                 
#> [297] "  vtc$zip,"                                                                      
#> [298] "  vtc$zip_norm,"                                                                 
#> [299] "  compare = valid_zip"                                                           
#> [300] ")"                                                                               
#> [301] "```"                                                                             
#> [302] ""                                                                                
#> [303] "### State"                                                                       
#> [304] ""                                                                                
#> [305] "Valid two digit state abbreviations can be made using the "                      
#> [306] "`campfin::normal_state()` function."                                             
#> [307] ""                                                                                
#> [308] "```{r state_norm}"                                                               
#> [309] "vtc <- vtc %>% "                                                                 
#> [310] "  mutate("                                                                       
#> [311] "    state_norm = normal_state("                                                  
#> [312] "      state = state,"                                                            
#> [313] "      abbreviate = TRUE,"                                                        
#> [314] "      na_rep = TRUE,"                                                            
#> [315] "      valid = valid_state"                                                       
#> [316] "    )"                                                                           
#> [317] "  )"                                                                             
#> [318] "```"                                                                             
#> [319] ""                                                                                
#> [320] "```{r state_view}"                                                               
#> [321] "vtc %>% "                                                                        
#> [322] "  filter(state != state_norm) %>% "                                              
#> [323] "  count(state, sort = TRUE)"                                                     
#> [324] "```"                                                                             
#> [325] ""                                                                                
#> [326] "```{r state_progress}"                                                           
#> [327] "progress_table("                                                                 
#> [328] "  vtc$state,"                                                                    
#> [329] "  vtc$state_norm,"                                                               
#> [330] "  compare = valid_state"                                                         
#> [331] ")"                                                                               
#> [332] "```"                                                                             
#> [333] ""                                                                                
#> [334] "### City"                                                                        
#> [335] ""                                                                                
#> [336] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [337] "the wide variety of valid cities and formats."                                   
#> [338] ""                                                                                
#> [339] "#### Normal"                                                                     
#> [340] ""                                                                                
#> [341] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [342] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [343] "`invalid_city` values."                                                          
#> [344] ""                                                                                
#> [345] "```{r city_norm}"                                                                
#> [346] "vtc <- vtc %>% "                                                                 
#> [347] "  mutate("                                                                       
#> [348] "    city_norm = normal_city("                                                    
#> [349] "      city = city, "                                                             
#> [350] "      abbs = usps_city,"                                                         
#> [351] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [352] "      na = invalid_city,"                                                        
#> [353] "      na_rep = TRUE"                                                             
#> [354] "    )"                                                                           
#> [355] "  )"                                                                             
#> [356] "```"                                                                             
#> [357] ""                                                                                
#> [358] "#### Swap"                                                                       
#> [359] ""                                                                                
#> [360] "We can further improve normalization by comparing our normalized value"          
#> [361] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [362] "If the normalized value is either an abbreviation for or very similar to the"    
#> [363] "expected value, we can confidently swap those two."                              
#> [364] ""                                                                                
#> [365] "```{r city_swap}"                                                                
#> [366] "vtc <- vtc %>% "                                                                 
#> [367] "  rename(city_raw = city) %>% "                                                  
#> [368] "  left_join("                                                                    
#> [369] "    y = zipcodes,"                                                               
#> [370] "    by = c("                                                                     
#> [371] "      \"state_norm\" = \"state\","                                               
#> [372] "      \"zip_norm\" = \"zip\""                                                    
#> [373] "    )"                                                                           
#> [374] "  ) %>% "                                                                        
#> [375] "  rename(city_match = city) %>% "                                                
#> [376] "  mutate("                                                                       
#> [377] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [378] "    match_dist = str_dist(city_norm, city_match),"                               
#> [379] "    city_swap = if_else("                                                        
#> [380] "      condition = !is.na(match_dist) & (match_abb | match_dist == 1),"           
#> [381] "      true = city_match,"                                                        
#> [382] "      false = city_norm"                                                         
#> [383] "    )"                                                                           
#> [384] "  ) %>% "                                                                        
#> [385] "  select("                                                                       
#> [386] "    -city_match,"                                                                
#> [387] "    -match_dist,"                                                                
#> [388] "    -match_abb"                                                                  
#> [389] "  )"                                                                             
#> [390] "```"                                                                             
#> [391] ""                                                                                
#> [392] "#### Refine"                                                                     
#> [393] ""                                                                                
#> [394] "The [OpenRefine][or] algorithms can be used to group similar strings and replace"
#> [395] "the less common versions with their most common counterpart. This can greatly"   
#> [396] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [397] "strings that have a valid city/state/zip combination."                           
#> [398] ""                                                                                
#> [399] "[or]: https://openrefine.org/"                                                   
#> [400] ""                                                                                
#> [401] "```{r city_refine}"                                                              
#> [402] "good_refine <- vtc %>% "                                                         
#> [403] "  mutate("                                                                       
#> [404] "    city_refine = city_swap %>% "                                                
#> [405] "      key_collision_merge() %>% "                                                
#> [406] "      n_gram_merge(numgram = 1)"                                                 
#> [407] "  ) %>% "                                                                        
#> [408] "  filter(city_refine != city_swap) %>% "                                         
#> [409] "  inner_join("                                                                   
#> [410] "    y = zipcodes,"                                                               
#> [411] "    by = c("                                                                     
#> [412] "      \"city_refine\" = \"city\","                                               
#> [413] "      \"state_norm\" = \"state\","                                               
#> [414] "      \"zip_norm\" = \"zip\""                                                    
#> [415] "    )"                                                                           
#> [416] "  )"                                                                             
#> [417] "```"                                                                             
#> [418] ""                                                                                
#> [419] "```{r city_count, echo=FALSE}"                                                   
#> [420] "good_refine %>%"                                                                 
#> [421] "  count("                                                                        
#> [422] "    state_norm, "                                                                
#> [423] "    zip_norm, "                                                                  
#> [424] "    city_swap, "                                                                 
#> [425] "    city_refine,"                                                                
#> [426] "    sort = TRUE"                                                                 
#> [427] "  )"                                                                             
#> [428] "```"                                                                             
#> [429] ""                                                                                
#> [430] "Then we can join the refined values back to the database."                       
#> [431] ""                                                                                
#> [432] "```{r city_join}"                                                                
#> [433] "vtc <- vtc %>% "                                                                 
#> [434] "  left_join(good_refine) %>% "                                                   
#> [435] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [436] "```"                                                                             
#> [437] ""                                                                                
#> [438] "#### Progress"                                                                   
#> [439] ""                                                                                
#> [440] "```{r city_progress, echo=FALSE}"                                                
#> [441] "many_city <- c(valid_city, extra_city)"                                          
#> [442] "progress <- progress_table("                                                     
#> [443] "  str_to_upper(vtc$city_raw),"                                                   
#> [444] "  vtc$city_norm,"                                                                
#> [445] "  vtc$city_swap,"                                                                
#> [446] "  vtc$city_refine,"                                                              
#> [447] "  compare = many_city"                                                           
#> [448] ") %>% mutate(stage = as_factor(stage))"                                          
#> [449] "kable(progress, digits = 3)"                                                     
#> [450] "```"                                                                             
#> [451] ""                                                                                
#> [452] "You can see how the percentage of valid values increased with each stage."       
#> [453] ""                                                                                
#> [454] "```{r bar_progress, echo=FALSE}"                                                 
#> [455] "raw_in <- percent(prop_in(vtc$city_raw, valid_city))"                            
#> [456] "progress %>% "                                                                   
#> [457] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [458] "  geom_hline(yintercept = 0.99) +"                                               
#> [459] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [460] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [461] "  scale_y_continuous(labels = percent) +"                                        
#> [462] "  labs("                                                                         
#> [463] "    title = \"Vermont City Normalization Progress\","                            
#> [464] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [465] "    x = \"Stage\","                                                              
#> [466] "    y = \"Percent Valid\""                                                       
#> [467] "  )"                                                                             
#> [468] "```"                                                                             
#> [469] ""                                                                                
#> [470] "More importantly, the number of distinct values decreased each stage. We were"   
#> [471] "able to confidently change many distinct invalid values to their valid"          
#> [472] "equivalent."                                                                     
#> [473] ""                                                                                
#> [474] "```{r bar_distinct, echo=FALSE}"                                                 
#> [475] "progress %>% "                                                                   
#> [476] "  select("                                                                       
#> [477] "    stage, "                                                                     
#> [478] "    all = n_distinct,"                                                           
#> [479] "    bad = n_diff"                                                                
#> [480] "  ) %>% "                                                                        
#> [481] "  mutate(good = all - bad) %>% "                                                 
#> [482] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [483] "  mutate(name = name == \"good\") %>% "                                          
#> [484] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [485] "  geom_col(aes(fill = name)) +"                                                  
#> [486] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [487] "  scale_y_continuous(labels = comma) +"                                          
#> [488] "  theme(legend.position = \"bottom\") +"                                         
#> [489] "  labs("                                                                         
#> [490] "    title = \"Vermont City Normalization Progress\","                            
#> [491] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [492] "    x = \"Stage\","                                                              
#> [493] "    y = \"Distinct Values\","                                                    
#> [494] "    fill = \"Valid\""                                                            
#> [495] "  )"                                                                             
#> [496] "```"                                                                             
#> [497] ""                                                                                
#> [498] "## Conclude"                                                                     
#> [499] ""                                                                                
#> [500] "Before exporting, we can remove the intermediary normalization columns and"      
#> [501] "rename all added variables with the `_clean` suffix."                            
#> [502] ""                                                                                
#> [503] "```{r clean_select}"                                                             
#> [504] "vtc <- vtc %>% "                                                                 
#> [505] "  select("                                                                       
#> [506] "    -city_norm,"                                                                 
#> [507] "    -city_swap,"                                                                 
#> [508] "    city_clean = city_refine"                                                    
#> [509] "  ) %>% "                                                                        
#> [510] "  rename_all(~str_replace(., \"_norm\", \"_clean\")) %>% "                       
#> [511] "  rename_all(~str_remove(., \"_raw\"))"                                          
#> [512] "```"                                                                             
#> [513] ""                                                                                
#> [514] "```{r clean_glimpse}"                                                            
#> [515] "glimpse(sample_n(vtc, 20))"                                                      
#> [516] "```"                                                                             
#> [517] ""                                                                                
#> [518] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [519] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [520] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [521] "1. There are `r comma(sum(vtc$na_flag))` records missing key variables."         
#> [522] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [523] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [524] ""                                                                                
#> [525] "## Export"                                                                       
#> [526] ""                                                                                
#> [527] "Now the file can be saved on disk for upload to the Accountability server."      
#> [528] ""                                                                                
#> [529] "```{r clean_dir}"                                                                
#> [530] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [531] "clean_path <- path(clean_dir, \"vt_contribs_clean.csv\")"                        
#> [532] "write_csv(vtc, clean_path, na = \"\")"                                           
#> [533] "file_size(clean_path)"                                                           
#> [534] "file_encoding(clean_path)"                                                       
#> [535] "```"                                                                             
#> [536] ""                                                                                
#> [537] "## Upload"                                                                       
#> [538] ""                                                                                
#> [539] "Using the [duckr] R package, we can wrap around the [duck] commnand line tool to"
#> [540] "upload the file to the IRW server."                                              
#> [541] ""                                                                                
#> [542] "[duckr]: https://github.com/kiernann/duckr"                                      
#> [543] "[duck]: https://duck.sh/"                                                        
#> [544] ""                                                                                
#> [545] "```{r clean_upload, eval=FALSE}"                                                 
#> [546] "# remotes::install_github(\"kiernann/duckr\")"                                   
#> [547] "s3_dir <- \"s3:/publicaccountability/csv/\""                                     
#> [548] "s3_path <- path(s3_dir, basename(clean_path))"                                   
#> [549] "if (require(duckr)) {"                                                           
#> [550] "  duckr::duck_upload(clean_path, s3_path)"                                       
#> [551] "}"                                                                               
#> [552] "```"                                                                             
#> [553] ""                                                                                
#> [554] "## Dictionary"                                                                   
#> [555] ""                                                                                
#> [556] "The following table describes the variables in our final exported file:"         
#> [557] ""                                                                                
#> [558] "```{r dict_make, echo=FALSE}"                                                    
#> [559] "dict_raw <- tibble("                                                             
#> [560] "  var = md_code(names(vtc)),"                                                    
#> [561] "  type = md_code(map_chr(vtc, typeof)),"                                         
#> [562] "  def = c("                                                                      
#> [563] "    \"Date contribution was made\","                                             
#> [564] "    \"Contribution amount or correction\","                                      
#> [565] "    \"Contributor name\","                                                       
#> [566] "    \"Contributor unique ID\","                                                  
#> [567] "    \"Contributor street address\","                                             
#> [568] "    \"Contributor secondary address\","                                          
#> [569] "    \"Contributor city name\","                                                  
#> [570] "    \"Contributor state abbreviation\","                                         
#> [571] "    \"Contributor ZIP+4 code\","                                                 
#> [572] "    \"Contributor occupation\","                                                 
#> [573] "    \"Contributor employer name\","                                              
#> [574] "    \"Contributor type\","                                                       
#> [575] "    \"Recipient committee name\","                                               
#> [576] "    \"Recipient unique ID\","                                                    
#> [577] "    \"Contribution method\","                                                    
#> [578] "    \"Recipient election or office sought\","                                    
#> [579] "    \"Flag for missing date, amount, or name\","                                 
#> [580] "    \"Flag for completely duplicated record\","                                  
#> [581] "    \"Calendar year of contribution date\","                                     
#> [582] "    \"Normalized combined street address\","                                     
#> [583] "    \"Normalized 5-digit ZIP code\","                                            
#> [584] "    \"Normalized 2-digit state abbreviation\","                                  
#> [585] "    \"Normalized city name\""                                                    
#> [586] "  )"                                                                             
#> [587] ")"                                                                               
#> [588] "```"                                                                             
#> [589] ""                                                                                
#> [590] "```{r dict_md, echo=FALSE}"                                                      
#> [591] "(dict_md <- kable("                                                              
#> [592] "  x = dict_raw,"                                                                 
#> [593] "  format = \"markdown\","                                                        
#> [594] "  col.names = c(\"Column\", \"Type\", \"Definition\")"                           
#> [595] "))"                                                                              
#> [596] "```"                                                                             
#> [597] ""                                                                                
#> [598] "```{r dict_write}"                                                               
#> [599] "write_lines("                                                                    
#> [600] "  x = c(\"# Vermont Contracts Data Dictionary\\n\", dict_md),"                   
#> [601] "  path = here(\"vt\", \"contribs\", \"vt_contribs_dict.md\"),"                   
#> [602] ")"                                                                               
#> [603] "```"                                                                             

use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())

#> /tmp/Rtmpo9IqyK/vt_contribs_diary.Rmd was created