Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type, author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contracts\""                                                    
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r Sys.time()`\""                                                        
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 3"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "if (!interactive()) {"                                                           
#>  [31] "  options(width = 99)"                                                           
#>  [32] "  set.seed(5)"                                                                   
#>  [33] "}"                                                                               
#>  [34] "```"                                                                             
#>  [35] ""                                                                                
#>  [36] "```{r create_docs_dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [37] "doc_dir <- fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"           
#>  [38] "```"                                                                             
#>  [39] ""                                                                                
#>  [40] "## Project"                                                                      
#>  [41] ""                                                                                
#>  [42] "The Accountability Project is an effort to cut across data silos and give"       
#>  [43] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [44] "way to search across huge volumes of public data about people and organizations."
#>  [45] ""                                                                                
#>  [46] "Our goal is to standardizing public data on a few key fields by thinking of each"
#>  [47] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [48] "variables:"                                                                      
#>  [49] ""                                                                                
#>  [50] "1. All **parties** to a transaction."                                            
#>  [51] "2. The **date** of the transaction."                                             
#>  [52] "3. The **amount** of money involved."                                            
#>  [53] ""                                                                                
#>  [54] "## Objectives"                                                                   
#>  [55] ""                                                                                
#>  [56] "This document describes the process used to complete the following objectives:"  
#>  [57] ""                                                                                
#>  [58] "1. How many records are in the database?"                                        
#>  [59] "1. Check for entirely duplicated records."                                       
#>  [60] "1. Check ranges of continuous variables."                                        
#>  [61] "1. Is there anything blank or missing?"                                          
#>  [62] "1. Check for consistency issues."                                                
#>  [63] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [64] "1. Create a `year` field from the transaction date."                             
#>  [65] "1. Make sure there is data on both parties to a transaction."                    
#>  [66] ""                                                                                
#>  [67] "## Packages"                                                                     
#>  [68] ""                                                                                
#>  [69] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [70] "and communicate these results. The `pacman` package will facilitate their"       
#>  [71] "installation and attachment."                                                    
#>  [72] ""                                                                                
#>  [73] "The IRW's `campfin` package will also have to be installed from GitHub. This"    
#>  [74] "package contains functions custom made to help facilitate the processing of"     
#>  [75] "campaign finance data."                                                          
#>  [76] ""                                                                                
#>  [77] "```{r load_packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [78] "if (!require(\"pacman\")) install.packages(\"pacman\")"                          
#>  [79] "pacman::p_load("                                                                 
#>  [80] "  tidyverse, # data manipulation"                                                
#>  [81] "  lubridate, # datetime strings"                                                 
#>  [82] "  gluedown, # printing markdown"                                                 
#>  [83] "  janitor, # clean data frames"                                                  
#>  [84] "  campfin, # custom irw tools"                                                   
#>  [85] "  aws.s3, # aws cloud storage"                                                   
#>  [86] "  refinr, # cluster & merge"                                                     
#>  [87] "  scales, # format strings"                                                      
#>  [88] "  knitr, # knit documents"                                                       
#>  [89] "  vroom, # fast reading"                                                         
#>  [90] "  rvest, # scrape html"                                                          
#>  [91] "  glue, # code strings"                                                          
#>  [92] "  here, # project paths"                                                         
#>  [93] "  httr, # http requests"                                                         
#>  [94] "  fs # local storage "                                                           
#>  [95] ")"                                                                               
#>  [96] "```"                                                                             
#>  [97] ""                                                                                
#>  [98] "This document should be run as part of the `R_campfin` project, which lives as a"
#>  [99] "sub-directory of the more general, language-agnostic"                            
#> [100] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#> [101] ""                                                                                
#> [102] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#> [103] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#> [104] "paths relative to _your_ machine."                                               
#> [105] ""                                                                                
#> [106] "```{r where_here}"                                                               
#> [107] "# where does this document knit?"                                                
#> [108] "here::dr_here(show_reason = FALSE)"                                              
#> [109] "```"                                                                             
#> [110] ""                                                                                
#> [111] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [112] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [113] ""                                                                                
#> [114] "## Data"                                                                         
#> [115] ""                                                                                
#> [116] ""                                                                                
#> [117] ""                                                                                
#> [118] "## Download"                                                                     
#> [119] ""                                                                                
#> [120] "```{r raw_dir}"                                                                  
#> [121] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [122] "```"                                                                             
#> [123] ""                                                                                
#> [124] "```{r raw_download}"                                                             
#> [125] "download.file()"                                                                 
#> [126] "```"                                                                             
#> [127] ""                                                                                
#> [128] "## Read"                                                                         
#> [129] ""                                                                                
#> [130] "```{r raw_read}"                                                                 
#> [131] "vtc <- read_delim("                                                              
#> [132] "  file = dir_ls(raw_dir)"                                                        
#> [133] ")"                                                                               
#> [134] "```"                                                                             
#> [135] ""                                                                                
#> [136] "## Explore"                                                                      
#> [137] ""                                                                                
#> [138] "There are `r scales::comma(nrow(vtc))` rows of `r ncol(vtc)` columns."           
#> [139] ""                                                                                
#> [140] "```{r glimpse}"                                                                  
#> [141] "glimpse(vtc)"                                                                    
#> [142] "tail(vtc)"                                                                       
#> [143] "```"                                                                             
#> [144] ""                                                                                
#> [145] "### Missing"                                                                     
#> [146] ""                                                                                
#> [147] "Columns vary in their degree of missing values."                                 
#> [148] ""                                                                                
#> [149] "```{r na_count}"                                                                 
#> [150] "col_stats(vtc, count_na)"                                                        
#> [151] "```"                                                                             
#> [152] ""                                                                                
#> [153] "We can flag any record missing a key variable needed to identify a transaction." 
#> [154] ""                                                                                
#> [155] "```{r na_flag}"                                                                  
#> [156] "vtc <- vtc %>% flag_na(date, last, amount, committee)"                           
#> [157] "sum(vtc$na_flag)"                                                                
#> [158] "```"                                                                             
#> [159] ""                                                                                
#> [160] "```{r na_view}"                                                                  
#> [161] "vtc %>% "                                                                        
#> [162] "  filter(na_flag) %>% "                                                          
#> [163] "  select(date, last, amount, committee)"                                         
#> [164] "```"                                                                             
#> [165] ""                                                                                
#> [166] "### Duplicates"                                                                  
#> [167] ""                                                                                
#> [168] "We can also flag any record completely duplicated across every column."          
#> [169] ""                                                                                
#> [170] "```{r dupe_flag}"                                                                
#> [171] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [172] "sum(vtc$dupe_flag)"                                                              
#> [173] "```"                                                                             
#> [174] ""                                                                                
#> [175] "```{r dupe_view}"                                                                
#> [176] "vtc %>% "                                                                        
#> [177] "  filter(dupe_flag) %>% "                                                        
#> [178] "  select(date, last, amount, committee)"                                         
#> [179] "```"                                                                             
#> [180] ""                                                                                
#> [181] "### Categorical"                                                                 
#> [182] ""                                                                                
#> [183] "```{r distinct_count}"                                                           
#> [184] "col_stats(vtc, n_distinct)"                                                      
#> [185] "```"                                                                             
#> [186] ""                                                                                
#> [187] "```{r distinct_plots, echo=FALSE}"                                               
#> [188] "explore_plot(vtc, type)"                                                         
#> [189] "```"                                                                             
#> [190] ""                                                                                
#> [191] "### Amounts"                                                                     
#> [192] ""                                                                                
#> [193] "```{r amount_summary}"                                                           
#> [194] "summary(vtc$amount)"                                                             
#> [195] "mean(vtc$amount <= 0)"                                                           
#> [196] "```"                                                                             
#> [197] ""                                                                                
#> [198] "These are the records with the minimum and maximum amounts."                     
#> [199] ""                                                                                
#> [200] "```{r amount_minmax}"                                                            
#> [201] "glimpse(vtc[c(which.max(vtc$amount), which.min(vtc$amount)), ])"                 
#> [202] "```"                                                                             
#> [203] ""                                                                                
#> [204] "```{r hist_amount, echo=FALSE}"                                                  
#> [205] "vtc %>%"                                                                         
#> [206] "  ggplot(aes(amount)) +"                                                         
#> [207] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [208] "  scale_y_continuous(labels = comma) +"                                          
#> [209] "  scale_x_continuous("                                                           
#> [210] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [211] "    labels = dollar,"                                                            
#> [212] "    trans = \"log10\""                                                           
#> [213] "  ) +"                                                                           
#> [214] "  labs("                                                                         
#> [215] "    title = \"Vermont Contracts Amount Distribution\","                          
#> [216] "    caption = \"Source: {source}\","                                             
#> [217] "    x = \"Amount\","                                                             
#> [218] "    y = \"Count\""                                                               
#> [219] "  )"                                                                             
#> [220] "```"                                                                             
#> [221] ""                                                                                
#> [222] "### Dates"                                                                       
#> [223] ""                                                                                
#> [224] "We can add the calendar year from `date` with `lubridate::year()`"               
#> [225] ""                                                                                
#> [226] "```{r date_year}"                                                                
#> [227] "vtc <- mutate(vtc, year = year(date))"                                           
#> [228] "```"                                                                             
#> [229] ""                                                                                
#> [230] "```{r date_range}"                                                               
#> [231] "min(vtc$date)"                                                                   
#> [232] "sum(vtc$year < 2000)"                                                            
#> [233] "max(vtc$date)"                                                                   
#> [234] "sum(vtc$date > today())"                                                         
#> [235] "```"                                                                             
#> [236] ""                                                                                
#> [237] "```{r bar_year, echo=FALSE}"                                                     
#> [238] "vtc %>% "                                                                        
#> [239] "  count(year) %>% "                                                              
#> [240] "  mutate(even = is_even(year)) %>% "                                             
#> [241] "  ggplot(aes(x = year, y = n)) +"                                                
#> [242] "  geom_col(aes(fill = even)) + "                                                 
#> [243] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [244] "  scale_y_continuous(labels = comma) +"                                          
#> [245] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [246] "  theme(legend.position = \"bottom\") +"                                         
#> [247] "  labs("                                                                         
#> [248] "    title = \"Vermont Contracts by Year\","                                      
#> [249] "    caption = \"Source: {source}\","                                             
#> [250] "    fill = \"Election Year\","                                                   
#> [251] "    x = \"Year Made\","                                                          
#> [252] "    y = \"Count\""                                                               
#> [253] "  )"                                                                             
#> [254] "```"                                                                             
#> [255] ""                                                                                
#> [256] "## Wrangle"                                                                      
#> [257] ""                                                                                
#> [258] "To improve the searchability of the database, we will perform some consistent,"  
#> [259] "confident string normalization. For geographic variables like city names and"    
#> [260] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [261] "facilitate this process."                                                        
#> [262] ""                                                                                
#> [263] "### Address"                                                                     
#> [264] ""                                                                                
#> [265] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [266] "will force consistence case, remove punctuation, and abbreviate official "       
#> [267] "USPS suffixes."                                                                  
#> [268] ""                                                                                
#> [269] "```{r address_norm}"                                                             
#> [270] "vtc <- vtc %>% "                                                                 
#> [271] "  unite("                                                                        
#> [272] "    col = address_full,"                                                         
#> [273] "    starts_with(\"address\"),"                                                   
#> [274] "    sep = \" \","                                                                
#> [275] "    remove = FALSE,"                                                             
#> [276] "    na.rm = TRUE"                                                                
#> [277] "  ) %>% "                                                                        
#> [278] "  mutate("                                                                       
#> [279] "    address_norm = normal_address("                                              
#> [280] "      address = address_full,"                                                   
#> [281] "      abbs = usps_street,"                                                       
#> [282] "      na_rep = TRUE"                                                             
#> [283] "    )"                                                                           
#> [284] "  ) %>% "                                                                        
#> [285] "  select(-address_full)"                                                         
#> [286] "```"                                                                             
#> [287] ""                                                                                
#> [288] "```{r address_view}"                                                             
#> [289] "vtc %>% "                                                                        
#> [290] "  select(contains(\"address\")) %>% "                                            
#> [291] "  distinct() %>% "                                                               
#> [292] "  sample_n(10)"                                                                  
#> [293] "```"                                                                             
#> [294] ""                                                                                
#> [295] "### ZIP"                                                                         
#> [296] ""                                                                                
#> [297] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [298] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [299] "zeroes dropped by other programs like Microsoft Excel."                          
#> [300] ""                                                                                
#> [301] "```{r zip_norm}"                                                                 
#> [302] "vtc <- vtc %>% "                                                                 
#> [303] "  mutate("                                                                       
#> [304] "    zip_norm = normal_zip("                                                      
#> [305] "      zip = zip,"                                                                
#> [306] "      na_rep = TRUE"                                                             
#> [307] "    )"                                                                           
#> [308] "  )"                                                                             
#> [309] "```"                                                                             
#> [310] ""                                                                                
#> [311] "```{r zip_progress}"                                                             
#> [312] "progress_table("                                                                 
#> [313] "  vtc$zip,"                                                                      
#> [314] "  vtc$zip_norm,"                                                                 
#> [315] "  compare = valid_zip"                                                           
#> [316] ")"                                                                               
#> [317] "```"                                                                             
#> [318] ""                                                                                
#> [319] "### State"                                                                       
#> [320] ""                                                                                
#> [321] "Valid two digit state abbreviations can be made using the "                      
#> [322] "`campfin::normal_state()` function."                                             
#> [323] ""                                                                                
#> [324] "```{r state_norm}"                                                               
#> [325] "vtc <- vtc %>% "                                                                 
#> [326] "  mutate("                                                                       
#> [327] "    state_norm = normal_state("                                                  
#> [328] "      state = state,"                                                            
#> [329] "      abbreviate = TRUE,"                                                        
#> [330] "      na_rep = TRUE,"                                                            
#> [331] "      valid = valid_state"                                                       
#> [332] "    )"                                                                           
#> [333] "  )"                                                                             
#> [334] "```"                                                                             
#> [335] ""                                                                                
#> [336] "```{r state_view}"                                                               
#> [337] "vtc %>% "                                                                        
#> [338] "  filter(state != state_norm) %>% "                                              
#> [339] "  count(state, state_norm, sort = TRUE)"                                         
#> [340] "```"                                                                             
#> [341] ""                                                                                
#> [342] "```{r state_progress}"                                                           
#> [343] "progress_table("                                                                 
#> [344] "  vtc$state,"                                                                    
#> [345] "  vtc$state_norm,"                                                               
#> [346] "  compare = valid_state"                                                         
#> [347] ")"                                                                               
#> [348] "```"                                                                             
#> [349] ""                                                                                
#> [350] "### City"                                                                        
#> [351] ""                                                                                
#> [352] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [353] "the wide variety of valid cities and formats."                                   
#> [354] ""                                                                                
#> [355] "#### Normal"                                                                     
#> [356] ""                                                                                
#> [357] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [358] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [359] "`invalid_city` values."                                                          
#> [360] ""                                                                                
#> [361] "```{r city_norm}"                                                                
#> [362] "vtc <- vtc %>% "                                                                 
#> [363] "  mutate("                                                                       
#> [364] "    city_norm = normal_city("                                                    
#> [365] "      city = city, "                                                             
#> [366] "      abbs = usps_city,"                                                         
#> [367] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [368] "      na = invalid_city,"                                                        
#> [369] "      na_rep = TRUE"                                                             
#> [370] "    )"                                                                           
#> [371] "  )"                                                                             
#> [372] "```"                                                                             
#> [373] ""                                                                                
#> [374] "#### Swap"                                                                       
#> [375] ""                                                                                
#> [376] "We can further improve normalization by comparing our normalized value"          
#> [377] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [378] "If the normalized value is either an abbreviation for or very similar to the"    
#> [379] "expected value, we can confidently swap those two."                              
#> [380] ""                                                                                
#> [381] "```{r city_swap}"                                                                
#> [382] "vtc <- vtc %>% "                                                                 
#> [383] "  rename(city_raw = city) %>% "                                                  
#> [384] "  left_join("                                                                    
#> [385] "    y = zipcodes,"                                                               
#> [386] "    by = c("                                                                     
#> [387] "      \"state_norm\" = \"state\","                                               
#> [388] "      \"zip_norm\" = \"zip\""                                                    
#> [389] "    )"                                                                           
#> [390] "  ) %>% "                                                                        
#> [391] "  rename(city_match = city) %>% "                                                
#> [392] "  mutate("                                                                       
#> [393] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [394] "    match_dist = str_dist(city_norm, city_match),"                               
#> [395] "    city_swap = if_else("                                                        
#> [396] "      condition = !is.na(match_dist) & (match_abb | match_dist == 1),"           
#> [397] "      true = city_match,"                                                        
#> [398] "      false = city_norm"                                                         
#> [399] "    )"                                                                           
#> [400] "  ) %>% "                                                                        
#> [401] "  select("                                                                       
#> [402] "    -city_match,"                                                                
#> [403] "    -match_dist,"                                                                
#> [404] "    -match_abb"                                                                  
#> [405] "  )"                                                                             
#> [406] "```"                                                                             
#> [407] ""                                                                                
#> [408] "#### Refine"                                                                     
#> [409] ""                                                                                
#> [410] "The [OpenRefine][or] algorithms can be used to group similar strings and replace"
#> [411] "the less common versions with their most common counterpart. This can greatly"   
#> [412] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [413] "strings that have a valid city/state/zip combination."                           
#> [414] ""                                                                                
#> [415] "[or]: https://openrefine.org/"                                                   
#> [416] ""                                                                                
#> [417] "```{r city_refine}"                                                              
#> [418] "good_refine <- vtc %>% "                                                         
#> [419] "  mutate("                                                                       
#> [420] "    city_refine = city_swap %>% "                                                
#> [421] "      key_collision_merge() %>% "                                                
#> [422] "      n_gram_merge(numgram = 1)"                                                 
#> [423] "  ) %>% "                                                                        
#> [424] "  filter(city_refine != city_swap) %>% "                                         
#> [425] "  inner_join("                                                                   
#> [426] "    y = zipcodes,"                                                               
#> [427] "    by = c("                                                                     
#> [428] "      \"city_refine\" = \"city\","                                               
#> [429] "      \"state_norm\" = \"state\","                                               
#> [430] "      \"zip_norm\" = \"zip\""                                                    
#> [431] "    )"                                                                           
#> [432] "  )"                                                                             
#> [433] "```"                                                                             
#> [434] ""                                                                                
#> [435] "```{r city_count, echo=FALSE}"                                                   
#> [436] "good_refine %>%"                                                                 
#> [437] "  count("                                                                        
#> [438] "    state_norm, "                                                                
#> [439] "    zip_norm, "                                                                  
#> [440] "    city_swap, "                                                                 
#> [441] "    city_refine,"                                                                
#> [442] "    sort = TRUE"                                                                 
#> [443] "  )"                                                                             
#> [444] "```"                                                                             
#> [445] ""                                                                                
#> [446] "Then we can join the refined values back to the database."                       
#> [447] ""                                                                                
#> [448] "```{r city_join}"                                                                
#> [449] "vtc <- vtc %>% "                                                                 
#> [450] "  left_join(good_refine) %>% "                                                   
#> [451] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [452] "```"                                                                             
#> [453] ""                                                                                
#> [454] "#### Progress"                                                                   
#> [455] ""                                                                                
#> [456] "Our goal for normalization was to increase the proportion of city values known"  
#> [457] "to be valid and reduce the total distinct values by correcting misspellings."    
#> [458] ""                                                                                
#> [459] "```{r city_progress, echo=FALSE}"                                                
#> [460] "many_city <- c(valid_city, extra_city)"                                          
#> [461] "progress <- progress_table("                                                     
#> [462] "  str_to_upper(vtc$city_raw),"                                                   
#> [463] "  vtc$city_norm,"                                                                
#> [464] "  vtc$city_swap,"                                                                
#> [465] "  vtc$city_refine,"                                                              
#> [466] "  compare = many_city"                                                           
#> [467] ") %>% mutate(stage = as_factor(stage))"                                          
#> [468] "kable(progress, digits = 3)"                                                     
#> [469] "```"                                                                             
#> [470] ""                                                                                
#> [471] "You can see how the percentage of valid values increased with each stage."       
#> [472] ""                                                                                
#> [473] "```{r bar_progress, echo=FALSE}"                                                 
#> [474] "raw_in <- percent(prop_in(vtc$city_raw, valid_city))"                            
#> [475] "progress %>% "                                                                   
#> [476] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [477] "  geom_hline(yintercept = 0.99) +"                                               
#> [478] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [479] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [480] "  scale_y_continuous(labels = percent) +"                                        
#> [481] "  labs("                                                                         
#> [482] "    title = \"Vermont City Normalization Progress\","                            
#> [483] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [484] "    x = \"Stage\","                                                              
#> [485] "    y = \"Percent Valid\""                                                       
#> [486] "  )"                                                                             
#> [487] "```"                                                                             
#> [488] ""                                                                                
#> [489] "More importantly, the number of distinct values decreased each stage. We were"   
#> [490] "able to confidently change many distinct invalid values to their valid"          
#> [491] "equivalent."                                                                     
#> [492] ""                                                                                
#> [493] "```{r bar_distinct, echo=FALSE}"                                                 
#> [494] "progress %>% "                                                                   
#> [495] "  select("                                                                       
#> [496] "    stage, "                                                                     
#> [497] "    all = n_distinct,"                                                           
#> [498] "    bad = n_diff"                                                                
#> [499] "  ) %>% "                                                                        
#> [500] "  mutate(good = all - bad) %>% "                                                 
#> [501] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [502] "  mutate(name = name == \"good\") %>% "                                          
#> [503] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [504] "  geom_col(aes(fill = name)) +"                                                  
#> [505] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [506] "  scale_y_continuous(labels = comma) +"                                          
#> [507] "  theme(legend.position = \"bottom\") +"                                         
#> [508] "  labs("                                                                         
#> [509] "    title = \"Vermont City Normalization Progress\","                            
#> [510] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [511] "    x = \"Stage\","                                                              
#> [512] "    y = \"Distinct Values\","                                                    
#> [513] "    fill = \"Valid\""                                                            
#> [514] "  )"                                                                             
#> [515] "```"                                                                             
#> [516] ""                                                                                
#> [517] "## Conclude"                                                                     
#> [518] ""                                                                                
#> [519] "Before exporting, we can remove the intermediary normalization columns and"      
#> [520] "rename all added variables with the `_clean` suffix."                            
#> [521] ""                                                                                
#> [522] "```{r clean_select}"                                                             
#> [523] "vtc <- vtc %>% "                                                                 
#> [524] "  select("                                                                       
#> [525] "    -city_norm,"                                                                 
#> [526] "    -city_swap,"                                                                 
#> [527] "    city_clean = city_refine"                                                    
#> [528] "  ) %>% "                                                                        
#> [529] "  rename_all(~str_replace(., \"_norm\", \"_clean\")) %>% "                       
#> [530] "  rename_all(~str_remove(., \"_raw\"))"                                          
#> [531] "```"                                                                             
#> [532] ""                                                                                
#> [533] "```{r clean_glimpse}"                                                            
#> [534] "glimpse(sample_n(vtc, 50))"                                                      
#> [535] "```"                                                                             
#> [536] ""                                                                                
#> [537] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [538] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [539] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [540] "1. There are `r comma(sum(vtc$na_flag))` records missing key variables."         
#> [541] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [542] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [543] ""                                                                                
#> [544] "## Export"                                                                       
#> [545] ""                                                                                
#> [546] "Now the file can be saved on disk for upload to the Accountability server."      
#> [547] ""                                                                                
#> [548] "```{r clean_dir}"                                                                
#> [549] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [550] "clean_path <- path(clean_dir, \"vt_contribs_clean.csv\")"                        
#> [551] "write_csv(vtc, clean_path, na = \"\")"                                           
#> [552] "(clean_size <- file_size(clean_path))"                                           
#> [553] "file_encoding(clean_path) %>% "                                                  
#> [554] "  mutate(across(path, path.abbrev))"                                             
#> [555] "```"                                                                             
#> [556] ""                                                                                
#> [557] "## Upload"                                                                       
#> [558] ""                                                                                
#> [559] "We can use the `aws.s3::put_object()` to upload the text file to the IRW server."
#> [560] ""                                                                                
#> [561] "```{r aws_upload, eval=FALSE}"                                                   
#> [562] "aws_path <- path(\"csv\", basename(clean_path))"                                 
#> [563] "if (!object_exists(aws_path, \"publicaccountability\")) {"                       
#> [564] "  put_object("                                                                   
#> [565] "    file = clean_path,"                                                          
#> [566] "    object = aws_path, "                                                         
#> [567] "    bucket = \"publicaccountability\","                                          
#> [568] "    acl = \"public-read\","                                                      
#> [569] "    show_progress = TRUE,"                                                       
#> [570] "    multipart = TRUE"                                                            
#> [571] "  )"                                                                             
#> [572] "}"                                                                               
#> [573] "aws_head <- head_object(aws_path, \"publicaccountability\")"                     
#> [574] "(aws_size <- as_fs_bytes(attr(aws_head, \"content-length\")))"                   
#> [575] "unname(aws_size == clean_size)"                                                  
#> [576] "```"                                                                             
#> [577] ""                                                                                
#> [578] "## Dictionary"                                                                   
#> [579] ""                                                                                
#> [580] "The following table describes the variables in our final exported file:"         
#> [581] ""                                                                                
#> [582] "```{r dict_make, echo=FALSE}"                                                    
#> [583] "dict_raw <- tibble("                                                             
#> [584] "  var = md_code(names(vtc)),"                                                    
#> [585] "  type = md_code(map_chr(vtc, typeof)),"                                         
#> [586] "  def = c("                                                                      
#> [587] "    \"\""                                                                        
#> [588] "  )"                                                                             
#> [589] ")"                                                                               
#> [590] "```"                                                                             
#> [591] ""                                                                                
#> [592] "```{r dict_md, echo=FALSE}"                                                      
#> [593] "(dict_md <- kable("                                                              
#> [594] "  x = dict_raw,"                                                                 
#> [595] "  format = \"markdown\","                                                        
#> [596] "  col.names = c(\"Column\", \"Type\", \"Definition\")"                           
#> [597] "))"                                                                              
#> [598] "```"                                                                             

use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpffLsVu/vt_contribs_diary.Rmd was created