Create a new template data diary

Take the arguments supplied and put them into the appropriate places in a new template diary. Write the new template diary in the supplied directory.

use_diary(st, type, author, auto = FALSE)

Arguments

st
The USPS state abbreviation.

type
The type of data, one of "contribs", "expends", or "lobby".

author
The author name of the new diary.

auto
If TRUE, file is created in the correct working directory. If FALSE, a plain character string is returned. If a directory name, the file is automatically written to that directory.

Value

The file path of new diary, invisibly.

Examples

use_diary("VT", "contribs", "Kiernan Nicholls", FALSE)

#>   [1] "---"                                                                             
#>   [2] "title: \"Vermont Contributions\""                                                
#>   [3] "author: \"Kiernan Nicholls\""                                                    
#>   [4] "date: \"`r date()`\""                                                            
#>   [5] "output:"                                                                         
#>   [6] "  github_document: "                                                             
#>   [7] "    df_print: tibble"                                                            
#>   [8] "    toc: true"                                                                   
#>   [9] "    toc_dept: 3"                                                                 
#>  [10] "editor_options: "                                                                
#>  [11] "  chunk_output_type: console"                                                    
#>  [12] "---"                                                                             
#>  [13] ""                                                                                
#>  [14] "<!-- Place comments regarding knitting here -->"                                 
#>  [15] ""                                                                                
#>  [16] "```{r setup, include=FALSE, purl=FALSE}"                                         
#>  [17] "library(knitr)"                                                                  
#>  [18] "opts_chunk$set("                                                                 
#>  [19] "  eval = TRUE,"                                                                  
#>  [20] "  echo = TRUE,"                                                                  
#>  [21] "  warning = FALSE,"                                                              
#>  [22] "  message = FALSE,"                                                              
#>  [23] "  error = FALSE,"                                                                
#>  [24] "  collapse = TRUE,"                                                              
#>  [25] "  comment = \"#>\","                                                             
#>  [26] "  fig.path = \"../plots/\","                                                     
#>  [27] "  fig.width = 10,"                                                               
#>  [28] "  dpi = 300"                                                                     
#>  [29] ")"                                                                               
#>  [30] "if (!interactive()) {"                                                           
#>  [31] "  options(width = 120)"                                                          
#>  [32] "  set.seed(5)"                                                                   
#>  [33] "}"                                                                               
#>  [34] "```"                                                                             
#>  [35] ""                                                                                
#>  [36] "```{r create_docs_dir, eval=FALSE, echo=FALSE, include=FALSE}"                   
#>  [37] "doc_dir <- fs::dir_create(here::here(\"vt\", \"contribs\", \"docs\"))"           
#>  [38] "```"                                                                             
#>  [39] ""                                                                                
#>  [40] "## Project"                                                                      
#>  [41] ""                                                                                
#>  [42] "The Accountability Project is an effort to cut across data silos and give"       
#>  [43] "journalists, policy professionals, activists, and the public at large a simple"  
#>  [44] "way to search across huge volumes of public data about people and organizations."
#>  [45] ""                                                                                
#>  [46] "Our goal is to standardize public data on a few key fields by thinking of each"  
#>  [47] "dataset row as a transaction. For each transaction there should be (at least) 3" 
#>  [48] "variables:"                                                                      
#>  [49] ""                                                                                
#>  [50] "1. All **parties** to a transaction."                                            
#>  [51] "2. The **date** of the transaction."                                             
#>  [52] "3. The **amount** of money involved."                                            
#>  [53] ""                                                                                
#>  [54] "## Objectives"                                                                   
#>  [55] ""                                                                                
#>  [56] "This document describes the process used to complete the following objectives:"  
#>  [57] ""                                                                                
#>  [58] "1. How many records are in the database?"                                        
#>  [59] "1. Check for entirely duplicated records."                                       
#>  [60] "1. Check ranges of continuous variables."                                        
#>  [61] "1. Is there anything blank or missing?"                                          
#>  [62] "1. Check for consistency issues."                                                
#>  [63] "1. Create a five-digit ZIP Code called `zip`."                                   
#>  [64] "1. Create a `year` field from the transaction date."                             
#>  [65] "1. Make sure there is data on both parties to a transaction."                    
#>  [66] ""                                                                                
#>  [67] "## Packages"                                                                     
#>  [68] ""                                                                                
#>  [69] "The following packages are needed to collect, manipulate, visualize, analyze,"   
#>  [70] "and communicate these results. The `pacman` package will facilitate their"       
#>  [71] "installation and attachment."                                                    
#>  [72] ""                                                                                
#>  [73] "```{r load_packages, message=FALSE, warning=FALSE, error=FALSE}"                 
#>  [74] "if (!require(\"pacman\")) {"                                                     
#>  [75] "  install.packages(\"pacman\")"                                                  
#>  [76] "}"                                                                               
#>  [77] "pacman::p_load("                                                                 
#>  [78] "  tidyverse, # data manipulation"                                                
#>  [79] "  lubridate, # datetime strings"                                                 
#>  [80] "  gluedown, # printing markdown"                                                 
#>  [81] "  janitor, # clean data frames"                                                  
#>  [82] "  campfin, # custom irw tools"                                                   
#>  [83] "  aws.s3, # aws cloud storage"                                                   
#>  [84] "  refinr, # cluster & merge"                                                     
#>  [85] "  scales, # format strings"                                                      
#>  [86] "  knitr, # knit documents"                                                       
#>  [87] "  vroom, # fast reading"                                                         
#>  [88] "  rvest, # scrape html"                                                          
#>  [89] "  glue, # code strings"                                                          
#>  [90] "  here, # project paths"                                                         
#>  [91] "  httr, # http requests"                                                         
#>  [92] "  fs # local storage "                                                           
#>  [93] ")"                                                                               
#>  [94] "```"                                                                             
#>  [95] ""                                                                                
#>  [96] "This document should be run as part of the `R_campfin` project, which lives as a"
#>  [97] "sub-directory of the more general, language-agnostic"                            
#>  [98] "[`irworkshop/accountability_datacleaning`][tap] GitHub repository."              
#>  [99] ""                                                                                
#> [100] "The `R_campfin` project uses the [RStudio projects][rproj] feature and should be"
#> [101] "run as such. The project also uses the dynamic `here::here()` tool for file"     
#> [102] "paths relative to _your_ machine."                                               
#> [103] ""                                                                                
#> [104] "```{r where_here}"                                                               
#> [105] "# where does this document knit?"                                                
#> [106] "here::i_am(\"vt/contribs/docs/vt_contribs_diary.Rmd\")"                          
#> [107] "```"                                                                             
#> [108] ""                                                                                
#> [109] "[tap]: https://github.com/irworkshop/accountability_datacleaning"                
#> [110] "[rproj]: https://support.rstudio.com/hc/en-us/articles/200526207-Using-Projects" 
#> [111] ""                                                                                
#> [112] "## Data"                                                                         
#> [113] ""                                                                                
#> [114] ""                                                                                
#> [115] ""                                                                                
#> [116] "## Download"                                                                     
#> [117] ""                                                                                
#> [118] "```{r raw_dir}"                                                                  
#> [119] "raw_url <- \"https://example.com/source_file.csv\""                              
#> [120] "raw_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"raw\"))"            
#> [121] "raw_path <- path(raw_dir, basename(raw_url))"                                    
#> [122] "```"                                                                             
#> [123] ""                                                                                
#> [124] "```{r raw_download}"                                                             
#> [125] "if (!file_exists(raw_path)) {"                                                   
#> [126] "  download.file(raw_url, raw_path)"                                              
#> [127] "}"                                                                               
#> [128] "```"                                                                             
#> [129] ""                                                                                
#> [130] "## Read"                                                                         
#> [131] ""                                                                                
#> [132] "```{r raw_read}"                                                                 
#> [133] "vtc <- read_delim("                                                              
#> [134] "  file = raw_path,"                                                              
#> [135] "  delim = \",\","                                                                
#> [136] "  escape_backslash = FALSE,"                                                     
#> [137] "  escape_double = FALSE,"                                                        
#> [138] "  col_types = cols("                                                             
#> [139] "    .default = col_character(),"                                                 
#> [140] "    date = col_date_usa(),"                                                      
#> [141] "    amount = col_double()"                                                       
#> [142] "  )"                                                                             
#> [143] ")"                                                                               
#> [144] "```"                                                                             
#> [145] ""                                                                                
#> [146] "## Explore"                                                                      
#> [147] ""                                                                                
#> [148] "There are `r comma(nrow(vtc))` rows of `r ncol(vtc)` columns. Each record"       
#> [149] "represents a single Contributions..."                                            
#> [150] ""                                                                                
#> [151] "```{r glimpse}"                                                                  
#> [152] "glimpse(vtc)"                                                                    
#> [153] "tail(vtc)"                                                                       
#> [154] "```"                                                                             
#> [155] ""                                                                                
#> [156] "### Missing"                                                                     
#> [157] ""                                                                                
#> [158] "Columns vary in their degree of missing values."                                 
#> [159] ""                                                                                
#> [160] "```{r na_count}"                                                                 
#> [161] "col_stats(vtc, count_na)"                                                        
#> [162] "```"                                                                             
#> [163] ""                                                                                
#> [164] "We can flag any record missing a key variable needed to identify a transaction." 
#> [165] ""                                                                                
#> [166] "```{r na_flag}"                                                                  
#> [167] "key_vars <- c(\"date\", \"last\", \"amount\", \"committee\")"                    
#> [168] "vtc <- flag_na(vtc, all_of(key_vars))"                                           
#> [169] "sum(vtc$na_flag)"                                                                
#> [170] "```"                                                                             
#> [171] ""                                                                                
#> [172] "```{r na_view}"                                                                  
#> [173] "vtc %>% "                                                                        
#> [174] "  filter(na_flag) %>% "                                                          
#> [175] "  select(all_of(key_vars))"                                                      
#> [176] "```"                                                                             
#> [177] ""                                                                                
#> [178] "### Duplicates"                                                                  
#> [179] ""                                                                                
#> [180] "We can also flag any record completely duplicated across every column."          
#> [181] ""                                                                                
#> [182] "```{r dupe_flag}"                                                                
#> [183] "vtc <- flag_dupes(vtc, -id)"                                                     
#> [184] "sum(vtc$dupe_flag)"                                                              
#> [185] "```"                                                                             
#> [186] ""                                                                                
#> [187] "```{r dupe_view}"                                                                
#> [188] "vtc %>% "                                                                        
#> [189] "  filter(dupe_flag) %>% "                                                        
#> [190] "  select(all_of(key_vars))"                                                      
#> [191] "```"                                                                             
#> [192] ""                                                                                
#> [193] "### Categorical"                                                                 
#> [194] ""                                                                                
#> [195] "```{r distinct_count}"                                                           
#> [196] "col_stats(vtc, n_distinct)"                                                      
#> [197] "```"                                                                             
#> [198] ""                                                                                
#> [199] "```{r distinct_plots, echo=FALSE}"                                               
#> [200] "explore_plot(vtc, type)"                                                         
#> [201] "```"                                                                             
#> [202] ""                                                                                
#> [203] "### Amounts"                                                                     
#> [204] ""                                                                                
#> [205] "```{r amount_summary}"                                                           
#> [206] "summary(vtc$amount)"                                                             
#> [207] "mean(vtc$amount <= 0)"                                                           
#> [208] "```"                                                                             
#> [209] ""                                                                                
#> [210] "These are the records with the minimum and maximum amounts."                     
#> [211] ""                                                                                
#> [212] "```{r amount_minmax}"                                                            
#> [213] "glimpse(vtc[c(which.max(vtc$amount), which.min(vtc$amount)), ])"                 
#> [214] "```"                                                                             
#> [215] ""                                                                                
#> [216] "```{r hist_amount, echo=FALSE}"                                                  
#> [217] "vtc %>%"                                                                         
#> [218] "  ggplot(aes(amount)) +"                                                         
#> [219] "  geom_histogram(fill = dark2[\"purple\"]) +"                                    
#> [220] "  scale_y_continuous(labels = comma) +"                                          
#> [221] "  scale_x_continuous("                                                           
#> [222] "    breaks = c(1 %o% 10^(0:6)),"                                                 
#> [223] "    labels = dollar,"                                                            
#> [224] "    trans = \"log10\""                                                           
#> [225] "  ) +"                                                                           
#> [226] "  labs("                                                                         
#> [227] "    title = \"Vermont Contributions Amount Distribution\","                      
#> [228] "    caption = \"Source: {source}\","                                             
#> [229] "    x = \"Amount\","                                                             
#> [230] "    y = \"Count\""                                                               
#> [231] "  )"                                                                             
#> [232] "```"                                                                             
#> [233] ""                                                                                
#> [234] "### Dates"                                                                       
#> [235] ""                                                                                
#> [236] "We can add the calendar year from `date` with `lubridate::year()`"               
#> [237] ""                                                                                
#> [238] "```{r date_year}"                                                                
#> [239] "vtc <- mutate(vtc, year = year(date))"                                           
#> [240] "```"                                                                             
#> [241] ""                                                                                
#> [242] "```{r date_range}"                                                               
#> [243] "min(vtc$date)"                                                                   
#> [244] "sum(vtc$year < 2000)"                                                            
#> [245] "max(vtc$date)"                                                                   
#> [246] "sum(vtc$date > today())"                                                         
#> [247] "```"                                                                             
#> [248] ""                                                                                
#> [249] "```{r bar_year, echo=FALSE}"                                                     
#> [250] "vtc %>% "                                                                        
#> [251] "  count(year) %>% "                                                              
#> [252] "  mutate(even = is_even(year)) %>% "                                             
#> [253] "  ggplot(aes(x = year, y = n)) +"                                                
#> [254] "  geom_col(aes(fill = even)) + "                                                 
#> [255] "  scale_fill_brewer(palette = \"Dark2\") +"                                      
#> [256] "  scale_y_continuous(labels = comma) +"                                          
#> [257] "  scale_x_continuous(breaks = seq(2000, 2020, by = 2)) +"                        
#> [258] "  theme(legend.position = \"bottom\") +"                                         
#> [259] "  labs("                                                                         
#> [260] "    title = \"Vermont Contributions by Year\","                                  
#> [261] "    caption = \"Source: {source}\","                                             
#> [262] "    fill = \"Election Year\","                                                   
#> [263] "    x = \"Year Made\","                                                          
#> [264] "    y = \"Count\""                                                               
#> [265] "  )"                                                                             
#> [266] "```"                                                                             
#> [267] ""                                                                                
#> [268] "## Wrangle"                                                                      
#> [269] ""                                                                                
#> [270] "To improve the searchability of the database, we will perform some consistent,"  
#> [271] "confident string normalization. For geographic variables like city names and"    
#> [272] "ZIP codes, the corresponding `campfin::normal_*()` functions are tailor made to "
#> [273] "facilitate this process."                                                        
#> [274] ""                                                                                
#> [275] "### Address"                                                                     
#> [276] ""                                                                                
#> [277] "For the street `addresss` variable, the `campfin::normal_address()` function"    
#> [278] "will force consistence case, remove punctuation, and abbreviate official "       
#> [279] "USPS suffixes."                                                                  
#> [280] ""                                                                                
#> [281] "```{r address_norm}"                                                             
#> [282] "vtc <- vtc %>% "                                                                 
#> [283] "  unite("                                                                        
#> [284] "    col = address_full,"                                                         
#> [285] "    starts_with(\"address\"),"                                                   
#> [286] "    sep = \" \","                                                                
#> [287] "    remove = FALSE,"                                                             
#> [288] "    na.rm = TRUE"                                                                
#> [289] "  ) %>% "                                                                        
#> [290] "  mutate("                                                                       
#> [291] "    address_norm = normal_address("                                              
#> [292] "      address = address_full,"                                                   
#> [293] "      abbs = usps_street,"                                                       
#> [294] "      na_rep = TRUE"                                                             
#> [295] "    )"                                                                           
#> [296] "  ) %>% "                                                                        
#> [297] "  select(-address_full)"                                                         
#> [298] "```"                                                                             
#> [299] ""                                                                                
#> [300] "```{r address_view}"                                                             
#> [301] "vtc %>% "                                                                        
#> [302] "  select(contains(\"address\")) %>% "                                            
#> [303] "  distinct() %>% "                                                               
#> [304] "  sample_n(10)"                                                                  
#> [305] "```"                                                                             
#> [306] ""                                                                                
#> [307] "### ZIP"                                                                         
#> [308] ""                                                                                
#> [309] "For ZIP codes, the `campfin::normal_zip()` function will attempt to create"      
#> [310] "valid _five_ digit codes by removing the ZIP+4 suffix and returning leading"     
#> [311] "zeroes dropped by other programs like Microsoft Excel."                          
#> [312] ""                                                                                
#> [313] "```{r zip_norm}"                                                                 
#> [314] "vtc <- vtc %>% "                                                                 
#> [315] "  mutate("                                                                       
#> [316] "    zip_norm = normal_zip("                                                      
#> [317] "      zip = zip,"                                                                
#> [318] "      na_rep = TRUE"                                                             
#> [319] "    )"                                                                           
#> [320] "  )"                                                                             
#> [321] "```"                                                                             
#> [322] ""                                                                                
#> [323] "```{r zip_progress}"                                                             
#> [324] "progress_table("                                                                 
#> [325] "  vtc$zip,"                                                                      
#> [326] "  vtc$zip_norm,"                                                                 
#> [327] "  compare = valid_zip"                                                           
#> [328] ")"                                                                               
#> [329] "```"                                                                             
#> [330] ""                                                                                
#> [331] "### State"                                                                       
#> [332] ""                                                                                
#> [333] "Valid two digit state abbreviations can be made using the "                      
#> [334] "`campfin::normal_state()` function."                                             
#> [335] ""                                                                                
#> [336] "```{r state_norm}"                                                               
#> [337] "vtc <- vtc %>% "                                                                 
#> [338] "  mutate("                                                                       
#> [339] "    state_norm = normal_state("                                                  
#> [340] "      state = state,"                                                            
#> [341] "      abbreviate = TRUE,"                                                        
#> [342] "      na_rep = TRUE,"                                                            
#> [343] "      valid = valid_state"                                                       
#> [344] "    )"                                                                           
#> [345] "  )"                                                                             
#> [346] "```"                                                                             
#> [347] ""                                                                                
#> [348] "```{r state_view}"                                                               
#> [349] "vtc %>% "                                                                        
#> [350] "  filter(state != state_norm) %>% "                                              
#> [351] "  count(state, state_norm, sort = TRUE)"                                         
#> [352] "```"                                                                             
#> [353] ""                                                                                
#> [354] "```{r state_progress}"                                                           
#> [355] "progress_table("                                                                 
#> [356] "  vtc$state,"                                                                    
#> [357] "  vtc$state_norm,"                                                               
#> [358] "  compare = valid_state"                                                         
#> [359] ")"                                                                               
#> [360] "```"                                                                             
#> [361] ""                                                                                
#> [362] "### City"                                                                        
#> [363] ""                                                                                
#> [364] "Cities are the most difficult geographic variable to normalize, simply due to"   
#> [365] "the wide variety of valid cities and formats."                                   
#> [366] ""                                                                                
#> [367] "#### Normal"                                                                     
#> [368] ""                                                                                
#> [369] "The `campfin::normal_city()` function is a good start, again converting case,"   
#> [370] "removing punctuation, but _expanding_ USPS abbreviations. We can also remove"    
#> [371] "`invalid_city` values."                                                          
#> [372] ""                                                                                
#> [373] "```{r city_norm}"                                                                
#> [374] "norm_city <- vtc %>% "                                                           
#> [375] "  distinct(city, state_norm, zip_norm) %>% "                                     
#> [376] "  mutate("                                                                       
#> [377] "    city_norm = normal_city("                                                    
#> [378] "      city = city, "                                                             
#> [379] "      abbs = usps_city,"                                                         
#> [380] "      states = c(\"VT\", \"DC\", \"VERMONT\"),"                                  
#> [381] "      na = invalid_city,"                                                        
#> [382] "      na_rep = TRUE"                                                             
#> [383] "    )"                                                                           
#> [384] "  )"                                                                             
#> [385] "```"                                                                             
#> [386] ""                                                                                
#> [387] "#### Swap"                                                                       
#> [388] ""                                                                                
#> [389] "We can further improve normalization by comparing our normalized value"          
#> [390] "against the _expected_ value for that record's state abbreviation and ZIP code." 
#> [391] "If the normalized value is either an abbreviation for or very similar to the"    
#> [392] "expected value, we can confidently swap those two."                              
#> [393] ""                                                                                
#> [394] "```{r city_swap}"                                                                
#> [395] "norm_city <- norm_city %>% "                                                     
#> [396] "  rename(city_raw = city) %>% "                                                  
#> [397] "  left_join("                                                                    
#> [398] "    y = zipcodes,"                                                               
#> [399] "    by = c("                                                                     
#> [400] "      \"state_norm\" = \"state\","                                               
#> [401] "      \"zip_norm\" = \"zip\""                                                    
#> [402] "    )"                                                                           
#> [403] "  ) %>% "                                                                        
#> [404] "  rename(city_match = city) %>% "                                                
#> [405] "  mutate("                                                                       
#> [406] "    match_abb = is_abbrev(city_norm, city_match),"                               
#> [407] "    match_dist = str_dist(city_norm, city_match),"                               
#> [408] "    city_swap = if_else("                                                        
#> [409] "      condition = !is.na(match_dist) & (match_abb | match_dist == 1),"           
#> [410] "      true = city_match,"                                                        
#> [411] "      false = city_norm"                                                         
#> [412] "    )"                                                                           
#> [413] "  ) %>% "                                                                        
#> [414] "  select("                                                                       
#> [415] "    -city_match,"                                                                
#> [416] "    -match_dist,"                                                                
#> [417] "    -match_abb"                                                                  
#> [418] "  )"                                                                             
#> [419] "```"                                                                             
#> [420] ""                                                                                
#> [421] "```{r city_rejoin}"                                                              
#> [422] "vtc <- left_join("                                                               
#> [423] "  x = vtc,"                                                                      
#> [424] "  y = norm_city,"                                                                
#> [425] "  by = c("                                                                       
#> [426] "    \"city_raw\", "                                                              
#> [427] "    \"state_norm\", "                                                            
#> [428] "    \"zip_norm\""                                                                
#> [429] "  )"                                                                             
#> [430] ")"                                                                               
#> [431] "```"                                                                             
#> [432] ""                                                                                
#> [433] "#### Refine"                                                                     
#> [434] ""                                                                                
#> [435] "The [OpenRefine][or] algorithms can be used to group similar strings and replace"
#> [436] "the less common versions with their most common counterpart. This can greatly"   
#> [437] "reduce inconsistency, but with low confidence; we will only keep any refined"    
#> [438] "strings that have a valid city/state/zip combination."                           
#> [439] ""                                                                                
#> [440] "[or]: https://openrefine.org/"                                                   
#> [441] ""                                                                                
#> [442] "```{r city_refine}"                                                              
#> [443] "good_refine <- vtc %>% "                                                         
#> [444] "  mutate("                                                                       
#> [445] "    city_refine = city_swap %>% "                                                
#> [446] "      key_collision_merge() %>% "                                                
#> [447] "      n_gram_merge(numgram = 1)"                                                 
#> [448] "  ) %>% "                                                                        
#> [449] "  filter(city_refine != city_swap) %>% "                                         
#> [450] "  inner_join("                                                                   
#> [451] "    y = zipcodes,"                                                               
#> [452] "    by = c("                                                                     
#> [453] "      \"city_refine\" = \"city\","                                               
#> [454] "      \"state_norm\" = \"state\","                                               
#> [455] "      \"zip_norm\" = \"zip\""                                                    
#> [456] "    )"                                                                           
#> [457] "  )"                                                                             
#> [458] "```"                                                                             
#> [459] ""                                                                                
#> [460] "```{r city_count, echo=FALSE}"                                                   
#> [461] "good_refine %>%"                                                                 
#> [462] "  count("                                                                        
#> [463] "    state_norm, "                                                                
#> [464] "    zip_norm, "                                                                  
#> [465] "    city_swap, "                                                                 
#> [466] "    city_refine,"                                                                
#> [467] "    sort = TRUE"                                                                 
#> [468] "  )"                                                                             
#> [469] "```"                                                                             
#> [470] ""                                                                                
#> [471] "Then we can join the refined values back to the database."                       
#> [472] ""                                                                                
#> [473] "```{r city_join}"                                                                
#> [474] "vtc <- vtc %>% "                                                                 
#> [475] "  left_join(good_refine) %>% "                                                   
#> [476] "  mutate(city_refine = coalesce(city_refine, city_swap))"                        
#> [477] "```"                                                                             
#> [478] ""                                                                                
#> [479] "#### Progress"                                                                   
#> [480] ""                                                                                
#> [481] "Our goal for normalization was to increase the proportion of city values known"  
#> [482] "to be valid and reduce the total distinct values by correcting misspellings."    
#> [483] ""                                                                                
#> [484] "```{r city_progress, echo=FALSE}"                                                
#> [485] "many_city <- c(valid_city, extra_city)"                                          
#> [486] "progress <- progress_table("                                                     
#> [487] "  str_to_upper(vtc$city_raw),"                                                   
#> [488] "  vtc$city_norm,"                                                                
#> [489] "  vtc$city_swap,"                                                                
#> [490] "  vtc$city_refine,"                                                              
#> [491] "  compare = many_city"                                                           
#> [492] ") %>% mutate(stage = as_factor(stage))"                                          
#> [493] "kable(progress, digits = 3)"                                                     
#> [494] "```"                                                                             
#> [495] ""                                                                                
#> [496] "You can see how the percentage of valid values increased with each stage."       
#> [497] ""                                                                                
#> [498] "```{r bar_progress, echo=FALSE}"                                                 
#> [499] "raw_in <- percent(prop_in(vtc$city_raw, valid_city))"                            
#> [500] "progress %>% "                                                                   
#> [501] "  ggplot(aes(x = stage, y = prop_in)) +"                                         
#> [502] "  geom_hline(yintercept = 0.99) +"                                               
#> [503] "  geom_col(fill = dark2[\"purple\"]) +"                                          
#> [504] "  coord_cartesian(ylim = c(0.75, 1)) +"                                          
#> [505] "  scale_y_continuous(labels = percent) +"                                        
#> [506] "  labs("                                                                         
#> [507] "    title = \"Vermont City Normalization Progress\","                            
#> [508] "    subtitle = glue(\"Raw at {raw_in} before conversion to uppercase\"),"        
#> [509] "    x = \"Stage\","                                                              
#> [510] "    y = \"Percent Valid\""                                                       
#> [511] "  )"                                                                             
#> [512] "```"                                                                             
#> [513] ""                                                                                
#> [514] "More importantly, the number of distinct values decreased each stage. We were"   
#> [515] "able to confidently change many distinct invalid values to their valid"          
#> [516] "equivalent."                                                                     
#> [517] ""                                                                                
#> [518] "```{r bar_distinct, echo=FALSE}"                                                 
#> [519] "progress %>% "                                                                   
#> [520] "  select("                                                                       
#> [521] "    stage, "                                                                     
#> [522] "    all = n_distinct,"                                                           
#> [523] "    bad = n_diff"                                                                
#> [524] "  ) %>% "                                                                        
#> [525] "  mutate(good = all - bad) %>% "                                                 
#> [526] "  pivot_longer(c(\"good\", \"bad\")) %>% "                                       
#> [527] "  mutate(name = name == \"good\") %>% "                                          
#> [528] "  ggplot(aes(x = stage, y = value)) +"                                           
#> [529] "  geom_col(aes(fill = name)) +"                                                  
#> [530] "  scale_fill_brewer(palette = \"Dark2\", direction = -1) +"                      
#> [531] "  scale_y_continuous(labels = comma) +"                                          
#> [532] "  theme(legend.position = \"bottom\") +"                                         
#> [533] "  labs("                                                                         
#> [534] "    title = \"Vermont City Normalization Progress\","                            
#> [535] "    subtitle = \"Distinct values, valid and invalid\","                          
#> [536] "    x = \"Stage\","                                                              
#> [537] "    y = \"Distinct Values\","                                                    
#> [538] "    fill = \"Valid\""                                                            
#> [539] "  )"                                                                             
#> [540] "```"                                                                             
#> [541] ""                                                                                
#> [542] "## Conclude"                                                                     
#> [543] ""                                                                                
#> [544] "Before exporting, we can remove the intermediary normalization columns and"      
#> [545] "rename all added variables with the `_clean` suffix."                            
#> [546] ""                                                                                
#> [547] "```{r clean_select}"                                                             
#> [548] "vtc <- vtc %>% "                                                                 
#> [549] "  select("                                                                       
#> [550] "    -city_norm,"                                                                 
#> [551] "    -city_swap,"                                                                 
#> [552] "    city_clean = city_refine"                                                    
#> [553] "  ) %>% "                                                                        
#> [554] "  rename_all(~str_replace(., \"_norm\", \"_clean\")) %>% "                       
#> [555] "  rename_all(~str_remove(., \"_raw\"))"                                          
#> [556] "```"                                                                             
#> [557] ""                                                                                
#> [558] "```{r clean_glimpse}"                                                            
#> [559] "glimpse(sample_n(vtc, 50))"                                                      
#> [560] "```"                                                                             
#> [561] ""                                                                                
#> [562] "1. There are `r comma(nrow(vtc))` records in the database."                      
#> [563] "1. There are `r comma(sum(vtc$dupe_flag))` duplicate records in the database."   
#> [564] "1. The range and distribution of `amount` and `date` seem reasonable."           
#> [565] "1. There are `r comma(sum(vtc$na_flag))` records missing key variables."         
#> [566] "1. Consistency in geographic data has been improved with `campfin::normal_*()`." 
#> [567] "1. The 4-digit `year` variable has been created with `lubridate::year()`."       
#> [568] ""                                                                                
#> [569] "## Export"                                                                       
#> [570] ""                                                                                
#> [571] "Now the file can be saved on disk for upload to the Accountability server."      
#> [572] ""                                                                                
#> [573] "```{r clean_dir}"                                                                
#> [574] "clean_dir <- dir_create(here(\"vt\", \"contribs\", \"data\", \"clean\"))"        
#> [575] "clean_path <- path(clean_dir, \"vt_contribs_clean.csv\")"                        
#> [576] "write_csv(vtc, clean_path, na = \"\")"                                           
#> [577] "(clean_size <- file_size(clean_path))"                                           
#> [578] "file_encoding(clean_path) %>% "                                                  
#> [579] "  mutate(across(path, path.abbrev))"                                             
#> [580] "```"                                                                             
#> [581] ""                                                                                
#> [582] "## Upload"                                                                       
#> [583] ""                                                                                
#> [584] "We can use the `aws.s3::put_object()` to upload the text file to the IRW server."
#> [585] ""                                                                                
#> [586] "```{r aws_upload, eval=FALSE}"                                                   
#> [587] "aws_path <- path(\"csv\", basename(clean_path))"                                 
#> [588] "if (!object_exists(aws_path, \"publicaccountability\")) {"                       
#> [589] "  put_object("                                                                   
#> [590] "    file = clean_path,"                                                          
#> [591] "    object = aws_path, "                                                         
#> [592] "    bucket = \"publicaccountability\","                                          
#> [593] "    acl = \"public-read\","                                                      
#> [594] "    show_progress = TRUE,"                                                       
#> [595] "    multipart = TRUE"                                                            
#> [596] "  )"                                                                             
#> [597] "}"                                                                               
#> [598] "aws_head <- head_object(aws_path, \"publicaccountability\")"                     
#> [599] "(aws_size <- as_fs_bytes(attr(aws_head, \"content-length\")))"                   
#> [600] "unname(aws_size == clean_size)"                                                  
#> [601] "```"                                                                             
#> [602] ""                                                                                
#> [603] "## Dictionary"                                                                   
#> [604] ""                                                                                
#> [605] "The following table describes the variables in our final exported file:"         
#> [606] ""                                                                                
#> [607] "```{r dict_make, echo=FALSE}"                                                    
#> [608] "dict_raw <- tibble("                                                             
#> [609] "  var = md_code(names(vtc)),"                                                    
#> [610] "  type = md_code(map_chr(vtc, typeof)),"                                         
#> [611] "  def = c("                                                                      
#> [612] "    \"\""                                                                        
#> [613] "  )"                                                                             
#> [614] ")"                                                                               
#> [615] "```"                                                                             
#> [616] ""                                                                                
#> [617] "```{r dict_md, echo=FALSE}"                                                      
#> [618] "(dict_md <- kable("                                                              
#> [619] "  x = dict_raw,"                                                                 
#> [620] "  format = \"markdown\","                                                        
#> [621] "  col.names = c(\"Column\", \"Type\", \"Definition\")"                           
#> [622] "))"                                                                              
#> [623] "```"                                                                             

use_diary("VT", "contribs", "Kiernan Nicholls", tempdir())

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpypC30L/vt_contribs_diary.Rmd was created